### 2023-05-22 #5487 @editorialbot
[PRE REVIEW]: mmh3: A Python extension for MurmurHash3
**Submitting author:** @hajimes (<a href="http://orcid.org/0000-0001-8542-1768">Hajime Senuma</a>)
**Repository:** https://github.com/hajimes/mmh3
**Branch with paper.md** (empty if default branch): paper
**Version:** v4.0.0
