### 2023-05-23 db0208e
Fix doi info in bibtext; add paper dir
