+++ paper/paper.bib
@@ -1,0 +1,178 @@
+@article{adja_blockchain-based_2021,
+	title        = {A blockchain-based certificate revocation management and status verification system},
+	author       = {Adja, Yves Christian Elloh and Hammi, Badis and Serhrouchni, Ahmed and Zeadally, Sherali},
+	year         = 2021,
+	journal      = {Computers \& Security},
+	volume       = 104,
+	pages        = 102209,
+	doi          = {10.1016/j.cose.2021.102209},
+	issn         = {0167-4048},
+	url          = {https://www.sciencedirect.com/science/article/pii/S016740482100033X},
+	keywords     = {Authentication, Blockchain, Bloom filter, Certificate, Decentralization, PKI, Revocation, Security, X509}
+}
+@misc{appleby_murmurhash3_2011,
+	title        = {{MurmurHash3} and {SMHasher}},
+	author       = {Appleby, Austin},
+	year         = 2011,
+	url          = {https://github.com/aappleby/smhasher}
+}
+@article{Bloom1970,
+	title        = {Space/Time Trade-Offs in Hash Coding with Allowable Errors},
+	author       = {Bloom, Burton H.},
+	year         = 1970,
+	month        = {jul},
+	journal      = {Commun. ACM},
+	publisher    = {Association for Computing Machinery},
+	address      = {New York, NY, USA},
+	volume       = 13,
+	number       = 7,
+	pages        = {422–426},
+	doi          = {10.1145/362686.362692},
+	issn         = {0001-0782},
+	url          = {https://doi.org/10.1145/362686.362692},
+	issue_date   = {July 1970},
+	numpages     = 5,
+	keywords     = {retrieval efficiency, storage efficiency, hash addressing, scatter storage, searching, storage layout, retrieval trade-offs, hash coding}
+}
+@inproceedings{Broder1997a,
+	title        = {On the resemblance and containment of documents},
+	author       = {Broder, Andrei Z.},
+	year         = 1997,
+	booktitle    = {Proceedings. {Compression} and {Complexity} of {SEQUENCES} 1997},
+	publisher    = {IEEE Comput. Soc},
+	pages        = {21--29},
+	doi          = {10.1109/SEQUEN.1997.666900},
+	isbn         = {0-8186-8132-2},
+	url          = {http://ieeexplore.ieee.org/lpdocs/epic03/wrapper.htm?arnumber=666900},
+	note         = {ISSN: 0818681322},
+	keywords     = {MinHash}
+}
+@misc{collet_xxhash_2012,
+	title        = {{xxHash}},
+	author       = {Collet, Yan},
+	year         = 2012,
+	url          = {https://github.com/Cyan4973/xxHash}
+}
+@misc{du_xxhash_2014,
+	title        = {{xxhash}},
+	author       = {Du, Yue},
+	year         = 2014,
+	url          = {https://github.com/ifduyue/python-xxhash}
+}
+@misc{faraday_security_understanding_2022,
+	title        = {Understanding {Spring4Shell}},
+	author       = {{Faraday Security}},
+	year         = 2022,
+	month        = jul,
+	url          = {https://faradaysec.com/understanding-spring4shell/}
+}
+@book{gorelick_high_2020,
+	title        = {High {Performance} {Python}: {Practical} {Performant} {Programming} for {Humans}},
+	author       = {Gorelick, Micha and Ozsvald, Ian},
+	year         = 2020,
+	month        = jun,
+	publisher    = {O'Reilly Media},
+	isbn         = {978-1-4920-5502-0},
+	edition      = {2nd edition}
+}
+@software{hugo_van_kemenade_2023_7882688,
[…]
