### 2023-05-23 a04f729
Format paper.md so that it passes markdownlint
