+++ paper/paper.md
@@ -20,40 +20,122 @@ bibliography: paper.bib
 
 ---
 
+<!-- markdownlint-disable MD025 -->
 # Summary
-This decade has witnessed the rapid evolution of artificial intelligence (AI), notably in the field of natural language processsing (NLP), as represented by the popularity of OpenAI’s ChatGPT. Another important advancement in computer science and engineering is found in the field of the Internet of Things (IoT), a crucial component of ubiquitous computing, as represented by the development of Shodan, the world's first IoT search engine.
 
-Underlying these advancements are high-performance algorithms and data structures that use non-cryptographic hash functions. This type of hash functions is generally characterized by four properties; they are fast, their resulting bits are statistically well-distributed, they have an avalanche effect, meaning a one-bit difference in a key changes at least half of the resulting bits, and they are collision resistant. Because cryptographic strength is not required in these use cases, they can leverage the efficiency of non-cryptographic hash functions.
+This decade has witnessed the rapid evolution of artificial intelligence (AI),
+notably in the field of natural language processsing (NLP), as represented by
+the popularity of OpenAI’s ChatGPT. Another important advancement in computer
+science and engineering is found in the field of the Internet of Things (IoT),
+a crucial component of ubiquitous computing, as represented by the development
+of Shodan, the world's first IoT search engine.
 
-MurmurHash3 and its test suite, SMHasher, was developed by @appleby_murmurhash3_2011 and is one of the earliest and most continuously popular hash functions specifically designed to implement the characteristics mentioned above.
+Underlying these advancements are high-performance algorithms and data
+structures that use non-cryptographic hash functions. This type of hash
+functions is generally characterized by four properties; they are fast, their
+resulting bits are statistically well-distributed, they have an avalanche
+effect, meaning a one-bit difference in a key changes at least half of the
+resulting bits, and they are collision resistant. Because cryptographic
+strength is not required in these use cases, they can leverage the efficiency
+of non-cryptographic hash functions.
 
-`mmh3` was launched in 2011 as a Python wrapper for MurmurHash3 and has been maintained ever since. Its API is very simple to use for Python programmers, as it offers both simple one-shot hash functions and hasher classes that allow incremental updating, whose methods are compliant to `hashlib`, a part of the Python Standard Library. The library provides Python wheels (i.e., pre-built binary packages) for immediate use on various platforms, including Linux (x86\_64, aarch64, i686, ppc64le, and s390x), Windows (win32, win\_amd64, and win\_arm64), and macOS (Intel Mac and Apple Silicon). From version 4.0.0, `mmh3` has been published under the MIT License, an OSI-approved permissive open-source license.
+MurmurHash3 and its test suite, SMHasher, was developed
+by @appleby_murmurhash3_2011 and is one of the earliest and most continuously
+popular hash functions specifically designed to implement the characteristics
+mentioned above.
 
-As of May 1, 2023, `mmh3` was being downloaded more than 2 million times per month, and it ranks as the 1,027th most downloaded PyPI package (of around 450,000 projects), showing that only 0.23% of the remaining packages in the PyPI ecosystem are more popular [@hugo_van_kemenade_2023_7882688]. According to PePy, as of May 22, 2023, the total downloads of this library exceeded 78 millions.
+`mmh3` was launched in 2011 as a Python wrapper for MurmurHash3 and has been
+maintained ever since. Its API is very simple to use for Python programmers,
+as it offers both simple one-shot hash functions and hasher classes that allow
+incremental updating, whose methods are compliant to `hashlib`, a part of the
[…]
