### 2023-05-24 3ea5ddb
Remove invalid doi
