+++ paper/paper.bib
@@ -149,8 +149,7 @@
 	year         = 2009,
 	journal      = {Journal of Machine Learning Research},
 	volume       = 10,
-	pages        = {2615--2637},
-	doi          = {10.5555/1577069.1755873}
+	pages        = {2615--2637}
 }
 @misc{shodan_its_2021,
 	title        = {"it's the {MMH3} hash of the http.html property. {See}: PyPI mmh3"},
