### 2023-05-31 #5487 @hugovk
Hello! Thanks for citing https://github.com/hugovk/top-pypi-packages!

Please could you adjust my name? "Van Kemenade" is my surname:

```diff
-29 that only 0.23% of the remaining packages in the PyPI ecosystem are more popular (Kemenade
+29 that only 0.23% of the remaining packages in the PyPI ecosystem are more popular (Van Kemenade
 30 et al., 2023). According to PePy, as of May 22, 2023, the total downloads of this
```
```diff
-101 Kemenade, H. van, Si, R., & Dollenstein, Z. (2023). Hugovk/top-pypi-packages: Release
+101 Van Kemenade, H., Si, R., & Dollenstein, Z. (2023). Hugovk/top-pypi-packages: Release
 102 2023.05 (Version 2023.05). Zenodo. https://doi.org/10.5281/zenodo.7882688
```
Thank you!
