### 2023-06-01 #5487 @danielskatz
@hajimes - thanks for your reply.  Due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or so.

Also, you could make the change mentioned above by @hugovk
