+++ paper/paper.bib
@@ -75,15 +75,17 @@
 	isbn         = {978-1-4920-5502-0},
 	edition      = {2nd edition}
 }
-@software{hugo_van_kemenade_2023_7882688,
-	title        = {hugovk/top-pypi-packages: Release 2023.05},
-	author       = {Hugo van Kemenade and Richard Si and Zsolt Dollenstein},
-	year         = 2023,
-	month        = may,
-	publisher    = {Zenodo},
-	doi          = {10.5281/zenodo.7882688},
-	url          = {https://doi.org/10.5281/zenodo.7882688},
-	version      = {2023.05}
+@software{hugo_van_kemenade_2023_7994944,
+  author       = {Van Kemenade, Hugo and
+                  Si, Richard and
+                  Dollenstein, Zsolt},
+  title        = {hugovk/top-pypi-packages: Release 2023.06},
+  month        = jun,
+  year         = 2023,
+  publisher    = {Zenodo},
+  version      = {2023.06},
+  doi          = {10.5281/zenodo.7994944},
+  url          = {https://doi.org/10.5281/zenodo.7994944}
 }
 @inproceedings{kakwani_indicnlpsuite_2020,
 	title        = {{IndicNLPSuite}: {Monolingual} {Corpora}, {Evaluation} {Benchmarks} and {Pre}-trained {Multilingual} {Language} {Models} for {Indian} {Languages}},
