### 2023-06-03 #5487 @hajimes
@hugovk
Hello! I apologize for writing your name incorrectly. I just blindly used the BibTeX entry provided by Zenodo, but I should have checked the entry before citation.

Anyway, thank you so much for providing valuable statistics!

@danielskatz
Thank you for your reply and reminder. I fixed the name of @hugovk and also updated lines 27-31 from

> As of May 1, 2023, mmh3 was being downloaded more than 2 million times per month, and it ranks as the 1,027th most downloaded PyPI package (of around 450,000 projects), showing that only 0.23% of the remaining packages in the PyPI ecosystem are more popular (Kemenade et al., 2023). According to PePy, as of May 22, 2023, the total downloads of this library exceeded 78 millions.

to

> As of June 1, 2023, mmh3 was being downloaded more than 2 million times per month, and it ranks as the 970th most downloaded PyPI package (of around 458,000 projects), showing that only 0.22% of the remaining packages in the PyPI ecosystem are more popular (Van Kemenade et al., 2023). According to PePy, as of June 2, 2023, the total downloads of this library exceeded 79 millions.
