### 2023-06-14 #5487 @hajimes
Perhaps not? (Although if you think he should be, I'll ask Austin Appleby whether he can join as a co-author)

I mean, it's true that the current mmh3 repository contains some code that was ultimately derived from Austin Appleby's original [SMHasher](https://github.com/aappleby/smhasher/tree/master) library (i.e., `murmurhash3.c` and `murmurhash3.h` which were derived from `MurmurHash3.cpp` and `MurmurHash3.h`, respectively).

However, probably a better solution is to exclude these files, use a git submodule to refer to Appleby's repo, and then write a script that converts the original C++ files to more portable C code at compile time. In this way I'll be able to clarify the contribution of each engineer, as well as increase the total amount of my own code.

If you think it's a good idea, I'll refactor mmh3 in a few days or so, though I'm not sure updates after submission are subject to evaluation by the review committee.
