### 2023-07-04 #5487 @hajimes
I sincerely apologize for my late response.

I'll try to implement the modified version of my library soon, hopefully tomorrow, at least within this week.
