### 2023-07-09 hajimes/mmh3#48 @hajimes
Decouple the original code from this repository
Currently I'm working on refactoring the library to decouple files whose large part traces back to the original C++ code (specifically, `murmurhash3.c` and `murmurhash3.h`) from this repository.

The update is to adress the pre-review process of the Journal of Open Source Software (JOSS), whose managing EiC (Daniel S. Katz) thoughtfully pointed out that it was not clear which part of this library was my (and other contributors') own contributions.
https://github.com/openjournals/joss-reviews/issues/5487

I proposed to use a git submodule to refer to Appleby's repository, and then write a script that converts the original C++ files to more portable C code at compile time.

It turns out, however, readability may be degraded to some extent in my current ad hoc implementation, which may also impact on easiness of extension. Solving these issues will be left for future updates. On the other hand, this update will clarify the extent of the authorship of code and solve the license issue #45.
