### 2023-07-10 hajimes/mmh3#48 @hajimes
After several tries, I gave up the above architecture. Because `mmh3module.c` also depends on `murmurhash3.h`, decoupling the header file from my repo completely will hinder future development.

Instead, generated files (`murmurhash3.c` and `murmurhash3.h`) will stay in my repo, rather than being temporary files in the build time. This approach follows the structure of `_hacl` (used by the official `hashlib`) in CPtyhon's repo.
https://github.com/python/cpython/tree/main/Modules/_hacl

It still fulfills the main motivation of this refactoring, i.e., clarifying the extent of the authorship of code.
