### 2023-07-12 02f1066
Fix a misplaced argument in setup.py
