### 2023-07-12 #5487 @danielskatz
@hajimes - Can you let us know here where this is now? I'm unsure from your issue above if you are done or still working, and if you are done, you could summarize here what you did and how things have changed from the POV of understanding your contributions and what JOSS will review.
