### 2023-07-13 4baa149
Refactor C code to be generated from the original
