+++ src/mmh3/mmh3module.c
@@ -11,8 +11,8 @@
 #include <byteswap.h>
 #endif
 
-#include "hashlib.h"
 #include "_mmh3/murmurhash3.h"
+#include "hashlib.h"
 
 #if defined(_MSC_VER)
 typedef signed __int8 int8_t;
