### 2023-07-13 5a1b03d
Fix a mypy error
