### 2023-07-13 #5487 @hajimes
I am very sorry for the repeated delay. As a progress report, I uploaded the result of refactoring to the [feature/refactor-submodule branch](https://github.com/hajimes/mmh3/tree/feature/refactor-submodule) that should be near-final, where [./src/mmh3/_mmh3/refresh.py](https://github.com/hajimes/mmh3/blob/feature/refactor-submodule/src/mmh3/_mmh3/refresh.py) handles the conversion of the original code by Austin Appleby.

I'll recheck my code later and merge it into the master branch, and then explain the changes in detail.

Again, so sorry for causing you trouble.
