+++ src/mmh3/_mmh3/murmurhash3.c
@@ -14,6 +14,11 @@
 // MurmurHash3 was written by Austin Appleby, and is placed in the public
 // domain. The author hereby disclaims copyright to this source code.
 
+// Note - The x86 and x64 versions do _not_ produce the same results, as the
+// algorithms are optimized for their respective platforms. You can still
+// compile and run any of them on any platform, but your performance with the
+// non-native version will be less than optimal.
+
 #include "murmurhash3.h"
 
 //-----------------------------------------------------------------------------
