+++ src/mmh3/_mmh3/README.md
@@ -7,7 +7,7 @@ This directory contains C files that were generated from the
 
 Use the `refresh.py` script in this directory to generate PEP 7-compliant C code from Appleby's original SMHasher project, instead of editing `_mmh3` files manually. Add transformation code to the `refresh.py` script to perform further edits.
 
-After file generation, use `clang-format` (with `.clang-format` in the top directory of `mmh3` project) to format generated code.
+After file generation, use `clang-format` (with `.clang-format` in the top directory of the `mmh3` project) to format the generated code.
 
 ## Local files
 
