### 2023-07-14 24ac0f9
Change the project structure (#48), fix type hints
