+++ src/mmh3/murmurhash3.h
@@ -1,324 +1,0 @@
-/***
- * This file is under MIT <year> Hajime Senuma, just like other files.
- * See LICENSE for details.
- *
- * It was originally written by Austin Appleby in C++ under the public domain,
- * but ported to PEP 7 C for Python 3.6 and later by the mmh3 project.
- *
- * Any issues should be reported to https://github.com/hajimes/mmh3/issues.
- *
- * The following is the original public domain notice by Austin Appleby.
- */
-
-//-----------------------------------------------------------------------------
-// MurmurHash3 was written by Austin Appleby, and is placed in the public
-// domain. The author hereby disclaims copyright to this source code.
-
-#ifndef MURMURHASH3_H_
-#define MURMURHASH3_H_
-
-// To handle 64-bit data; see https://docs.python.org/3/c-api/arg.html
-#ifndef PY_SSIZE_T_CLEAN
-#define PY_SSIZE_T_CLEAN
-#endif
-#include <Python.h>
-
-#if defined(__BYTE_ORDER__) && (__BYTE_ORDER__ == __ORDER_BIG_ENDIAN__)
-#include <byteswap.h>
-#endif
-
-//-----------------------------------------------------------------------------
-// Platform-specific functions and macros
-
-// Microsoft Visual Studio
-
-#if defined(_MSC_VER)
-typedef signed __int8 int8_t;
-typedef signed __int32 int32_t;
-typedef signed __int64 int64_t;
-typedef unsigned __int8 uint8_t;
-typedef unsigned __int32 uint32_t;
-typedef unsigned __int64 uint64_t;
-
-// Other compilers
-
-#else  // defined(_MSC_VER)
-
-#include <stdint.h>
-
-#endif  // !defined(_MSC_VER)
-
-//-----------------------------------------------------------------------------
-// Platform-specific functions and macros
-
-// Microsoft Visual Studio
-
-#if defined(_MSC_VER)
-
-#define FORCE_INLINE __forceinline
-
-#include <stdlib.h>
-
-#define ROTL32(x, y) _rotl(x, y)
-#define ROTL64(x, y) _rotl64(x, y)
-
-#define BIG_CONSTANT(x) (x)
-
-// Other compilers
-
-#else  // defined(_MSC_VER)
-
-#if ((__GNUC__ > 4) || (__GNUC__ == 4 && GNUC_MINOR >= 4))
-/* gcc version >= 4.4 4.1 = RHEL 5, 4.4 = RHEL 6. Don't inline for RHEL 5 gcc
- * which is 4.1*/
-#define FORCE_INLINE inline __attribute__((always_inline))
-#else
-#define FORCE_INLINE
-#endif
-
-static FORCE_INLINE uint32_t
-rotl32(uint32_t x, int8_t r)
-{
-    return (x << r) | (x >> (32 - r));
-}
-
-static FORCE_INLINE uint64_t
-rotl64(uint64_t x, int8_t r)
-{
-    return (x << r) | (x >> (64 - r));
-}
-
-#define ROTL32(x, y) rotl32(x, y)
-#define ROTL64(x, y) rotl64(x, y)
-
-#define BIG_CONSTANT(x) (x##LLU)
-
-#endif  // !defined(_MSC_VER)
-
-//-----------------------------------------------------------------------------
-// Block read - on little-endian machines this is a single load,
[…]
