+++ src/mmh3/_mmh3/refresh.py
@@ -1,0 +1,676 @@
+from __future__ import annotations
+
+import os
+import re
+import textwrap
+from collections.abc import Callable
+
+###
+# Simple classes to handle the transformation of the original code.
+#
+
+
+class MMH3Source:
+    """A data class to represent the original source code of MurmurHash3.
+
+    Lines to be retrieved are hard-coded, as the original code is effectively frozen.
+    """
+
+    def __init__(self, code: str) -> None:
+        self._code_lines = code.split("\n")
+
+    @property
+    def note_comment(self) -> str:
+        return "\n".join(self._code_lines[4:8])
+
+    @property
+    def header_include(self) -> str:
+        return "\n".join(self._code_lines[9:10])
+
+    @property
+    def macros(self) -> str:
+        return "\n".join(self._code_lines[11:50])
+
+    @property
+    def getblock_functions(self) -> str:
+        return "\n".join(self._code_lines[50:64])
+
+    @property
+    def finalization_mixes(self) -> str:
+        return "\n".join(self._code_lines[64:91])
+
+    @property
+    def body(self) -> str:
+        return "\n".join(self._code_lines[91:336])
+
+    @property
+    def finalization_x86_128(self) -> str:
+        return "\n".join(self._code_lines[233:246])
+
+    @property
+    def finalization_x64_128(self) -> str:
+        return "\n".join(self._code_lines[318:329])
+
+    @property
+    def constants_x86_128(self) -> str:
+        return "\n".join(self._code_lines[160:164])
+
+    @property
+    def constants_x64_128(self) -> str:
+        return "\n".join(self._code_lines[263:265])
+
+
+class MMH3Header:
+    """A data class to represent the original header code of MurmurHash3.
+
+    Lines to be retrieved are hard-coded, as the original code is effectively frozen.
+    """
+
+    def __init__(self, code: str) -> None:
+        self._code_lines = code.split("\n")
+
+    @property
+    def header_guards_begin(self) -> str:
+        return "\n".join(self._code_lines[4:7])
+
+    @property
+    def stdint(self) -> str:
+        return "\n".join(self._code_lines[7:26])
+
+    @property
+    def declarations(self) -> str:
+        return "\n".join(self._code_lines[26:36])
+
+    @property
+    def header_guards_end(self) -> str:
+        return "\n".join(self._code_lines[36:37])
+
+
+class MMH3CodeBuilder:
+    def __init__(self) -> None:
+        self._code: list[tuple[str, list[Callable[[str], str]]]] = []
+
+    def add(
+        self, subcode: str, transforms: list[Callable[[str], str]] = []
+    ) -> MMH3CodeBuilder:
+        self._code.append((subcode, transforms))
+        return self
+
+    def build(self) -> str:
+        new_code = ""
+
+        for subcode, transforms in self._code:
+            for tr in transforms:
+                subcode = tr(subcode)
+            new_code += subcode + "\n\n"
+
+        return new_code
+
+
+###
+# The following functions are used to transform the original MurmurHash3 code.
+#
+
+
+def append_python_directives(subcode: str) -> str:
+    """Append Python.h, as well as a macro definition to handle 64-bit data.
+
+    Args:
+        subcode (str): The code to be appended.
+
+    Returns:
+        str: The appended code.
+    """
+    subcode += "\n\n"
+
+    subcode += textwrap.dedent(
+        """\
[…]
