+++ src/mmh3/_mmh3/murmurhash3.c
@@ -296,40 +296,40 @@ murmurhash3_x64_128(const void *key, const Py_ssize_t len, const uint32_t seed,
 
     switch (len & 15) {
         case 15:
-            k2 ^= (uint64_t)tail[14] << 48;
+            k2 ^= ((uint64_t)tail[14]) << 48;
         case 14:
-            k2 ^= (uint64_t)tail[13] << 40;
+            k2 ^= ((uint64_t)tail[13]) << 40;
         case 13:
-            k2 ^= (uint64_t)tail[12] << 32;
+            k2 ^= ((uint64_t)tail[12]) << 32;
         case 12:
-            k2 ^= (uint64_t)tail[11] << 24;
+            k2 ^= ((uint64_t)tail[11]) << 24;
         case 11:
-            k2 ^= (uint64_t)tail[10] << 16;
+            k2 ^= ((uint64_t)tail[10]) << 16;
         case 10:
-            k2 ^= (uint64_t)tail[9] << 8;
+            k2 ^= ((uint64_t)tail[9]) << 8;
         case 9:
-            k2 ^= (uint64_t)tail[8] << 0;
+            k2 ^= ((uint64_t)tail[8]) << 0;
             k2 *= c2;
             k2 = ROTL64(k2, 33);
             k2 *= c1;
             h2 ^= k2;
 
         case 8:
-            k1 ^= (uint64_t)tail[7] << 56;
+            k1 ^= ((uint64_t)tail[7]) << 56;
         case 7:
-            k1 ^= (uint64_t)tail[6] << 48;
+            k1 ^= ((uint64_t)tail[6]) << 48;
         case 6:
-            k1 ^= (uint64_t)tail[5] << 40;
+            k1 ^= ((uint64_t)tail[5]) << 40;
         case 5:
-            k1 ^= (uint64_t)tail[4] << 32;
+            k1 ^= ((uint64_t)tail[4]) << 32;
         case 4:
-            k1 ^= (uint64_t)tail[3] << 24;
+            k1 ^= ((uint64_t)tail[3]) << 24;
         case 3:
-            k1 ^= (uint64_t)tail[2] << 16;
+            k1 ^= ((uint64_t)tail[2]) << 16;
         case 2:
-            k1 ^= (uint64_t)tail[1] << 8;
+            k1 ^= ((uint64_t)tail[1]) << 8;
         case 1:
-            k1 ^= (uint64_t)tail[0] << 0;
+            k1 ^= ((uint64_t)tail[0]) << 0;
             k1 *= c1;
             k1 = ROTL64(k1, 31);
             k1 *= c2;
