### 2023-07-14 ebd1d56
Update the cibuildwheel version
