+++ .github/workflows/wheels.yml
@@ -51,7 +51,7 @@ jobs:
         if: runner.os == 'Linux'
         uses: docker/setup-qemu-action@v2
       - name: Build wheels
-        uses: pypa/cibuildwheel@v2.12.3
+        uses: pypa/cibuildwheel@v2.14.0
         with:
           output-dir: wheelhouse
         env:
