### 2023-07-15 #5487 @hajimes
I finally uploaded the update as v4.0.1 yesterday.

Files that are mostly dependent on the original now reside in
https://github.com/hajimes/mmh3/tree/master/src/mmh3/_mmh3

The project structure is roughly based on the `_hacl` module in CPtyhon's repo, which is used by `hashlib`, a part of the Python standard library.
https://github.com/python/cpython/tree/main/Modules/_hacl

That is,
1. Appleby's original C++ repo is referred by git submodule.
2. `refresh.py` (that is my contribution) transforms original files into new forms to be used by this project.
3. Newly generated C files (concretely, `murmurhash3.c` and `murmurhash3.h`) stay as a part of my repository, rather than are treated as temporary files during the build time. I chose this design because `mmh3module.c` depends on `murmurhash3.h`.

Each transform is represented by a string-to-string function, and the building process of new files is represented by a list of code fragments each of which is accompanied by these transforms.

The transforms can be loosely classified into four categories:
- Transforms that introduce customs specific to Python extensions (e.g., `append_python_directives`, `introduce_py_ssize_t`, ...)
- Transforms to handle more platforms (e.g., `append_byteswap_header`, `transform_getblocks`, `expand_win_stdint_typedefs`, `fix_non_win_force_inline``, ...)
- "Tranforms" that actually append inline functions used by hashers in `mmh3module.c` (e.g., `append_mur_macros`, `generate_hasher_digest_*`, ...)
- Other trivial transforms

I am really thankful to editors and reviewers for giving me an opportunity to rethink the structure of my project.
