### 2023-07-19 #5487 @drvinceknight
@hajimes - I'll be taking this submission on as your editor. Could you take a look a [this list of potential reviewers](https://reviewers.joss.theoj.org/lookup) and identify a few people who would be good candidates to review this submission?
