### 2023-07-23 #5487 @hajimes
@drvinceknight - Thank you, I'm glad you took the time to be my editor!

Considering programming languages (Python and C/C++) and topic areas (high performance computing, natural language processing, machine learning, internet of things, cybersecurity, etc.), I'd suggest the following people as good candidates to review my paper:
- proycon
- mrshu
- jarrah42
- innat
- sergioburdisso
