### 2023-07-24 #5487 @drvinceknight
@proycon @mrshu @jarrah42 would any of you be willing to review this submission for JOSS? The submission under consideration is mmh3: A Python extension for MurmurHash3.
