### 2023-08-14 #5487 @drvinceknight
@innat @sergioburdisso would either of you be willing to review this submission for JOSS? The submission under consideration is mmh3: A Python extension for MurmurHash3.
