### 2023-10-08 #5487 @drvinceknight
@JPenuchot would you be willing to review this submission for JOSS? The submission under consideration is mmh3: A Python extension for MurmurHash3.
