### 2023-12-06 #6124 @editorialbot
[REVIEW]: mmh3: A Python extension for MurmurHash3
**Submitting author:** @hajimes (<a href="http://orcid.org/0000-0001-8542-1768">Hajime Senuma</a>)
**Repository:** https://github.com/hajimes/mmh3
**Branch with paper.md** (empty if default branch): paper
**Version:** 5.0.1
