### 2023-12-07 #6124 @mrshu
@drvinceknight this looks very good to me. The only thing I would identify as "missing" would be the Community guidelines, as I could not find them explicitly spelled out. Other than that, I believe this paper certainly merits acceptance.
