+++ .github/workflows/superlinter.yml
@@ -26,7 +26,7 @@ jobs:
 
       # Runs the Super-Linter action
       - name: Run Super-Linter
-        uses: github/super-linter@v5.7.2
+        uses: super-linter/super-linter@v5.7.2
         env:
           DEFAULT_BRANCH: master
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
