+++ .github/workflows/wheels.yml
@@ -77,7 +77,10 @@ jobs:
         with:
           python-version: '3.12'
       - name: Build sdist
-        run: python3 -m build --sdist
+        run: |
+          python -m pip install --upgrade pip
+          pip install setuptools
+          python setup.py build sdist
       - uses: actions/upload-artifact@v4
         with:
           path: dist/*.tar.gz
