### 2023-12-23 bd9d869
Update configs for artifact actions
