+++ .github/workflows/wheels.yml
@@ -65,6 +65,7 @@ jobs:
           CIBW_TEST_SKIP: '*-win_arm64 *-macosx_arm64'
       - uses: actions/upload-artifact@v4
         with:
+          name: wheels-${{ matrix.os }}-${{ matrix.cibw_archs }}
           path: ./wheelhouse/*.whl
   build_sdist:
     name: Build a source distribution
@@ -98,6 +99,7 @@ jobs:
         with:
           name: artifact
           path: dist
+          merge-multiple: true
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
