+++ .github/workflows/wheels.yml
@@ -65,7 +65,7 @@ jobs:
           CIBW_TEST_SKIP: '*-win_arm64 *-macosx_arm64'
       - uses: actions/upload-artifact@v4
         with:
-          name: wheels-${{ matrix.os }}-${{ matrix.archs }}
+          name: wheels-${{ matrix.os }}-${{ matrix.archs }}-${{ matrix.build }}
           path: ./wheelhouse/*.whl
   build_sdist:
     name: Build a source distribution
