+++ .github/workflows/wheels.yml
@@ -105,8 +105,16 @@ jobs:
           python -m pip install --upgrade pip
           pip install setuptools wheel twine
       - name: Publish
+        if: github.event.inputs.repository == 'pypi'
         env:
           TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
           TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
+        run: |
+          twine upload --repository ${{ github.event.inputs.repository }} dist/*
+      - name: 'Publish to TestPyPI'
+        if: github.event.inputs.repository == 'testpypi'
+        env:
+          TWINE_USERNAME: ${{ secrets.TEST_PYPI_USERNAME }}
+          TWINE_PASSWORD: ${{ secrets.TEST_PYPI_PASSWORD }}
         run: |
           twine upload --repository ${{ github.event.inputs.repository }} dist/*
