+++ .github/workflows/wheels.yml
@@ -86,6 +86,7 @@ jobs:
         with:
           path: dist/*.tar.gz
   publish:
+    if: github.event.inputs.repository != 'dummy'
     name: 'Upload to PyPI/Test PyPI'
     runs-on: ubuntu-22.04
     needs: [build_wheels, build_sdist]
@@ -104,17 +105,17 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install setuptools wheel twine
-      - name: Publish
-        if: github.event.inputs.repository == 'pypi'
-        env:
-          TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
-          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
-        run: |
-          twine upload --repository ${{ github.event.inputs.repository }} dist/*
       - name: 'Publish to TestPyPI'
         if: github.event.inputs.repository == 'testpypi'
         env:
           TWINE_USERNAME: ${{ secrets.TEST_PYPI_USERNAME }}
           TWINE_PASSWORD: ${{ secrets.TEST_PYPI_PASSWORD }}
+        run: |
+          twine upload --repository ${{ github.event.inputs.repository }} dist/*
+      - name: Publish
+        if: github.event.inputs.repository == 'pypi'
+        env:
+          TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
+          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
         run: |
           twine upload --repository ${{ github.event.inputs.repository }} dist/*
