+++ .github/workflows/build.yml
@@ -18,14 +18,14 @@ jobs:
   build:
     strategy:
       matrix:
-        os: [macos-12, windows-2022, ubuntu-22.04]
-        python-version: [3.8, 3.9, "3.10", "3.11"]
+        os: [macOS-13, windows-2022, ubuntu-22.04]
+        python-version: [3.8, 3.9, "3.10", "3.11", "3.12"]
 
     runs-on: ${{ matrix.os }}
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
