### 2024-01-02 1a4b2df
Add tests to check type hints for x64arch arg
