### 2024-01-02 300f8e5
Fix type hints
