### 2024-01-02 3acbd5b
Fix type hints to include the x64arch arg (#51)
