+++ tests/test_mmh3.py
@@ -169,6 +169,11 @@ def test_hash_from_buffer() -> None:
 
 def test_hash_bytes() -> None:
     assert mmh3.hash_bytes("foo") == b"aE\xf5\x01W\x86q\xe2\x87}\xba+\xe4\x87\xaf~"
+    assert (
+        mmh3.hash_bytes("foo", 0, True)
+        == b"aE\xf5\x01W\x86q\xe2\x87}\xba+\xe4\x87\xaf~"
+    )
+    assert mmh3.hash_bytes("foo", 0, x64arch=False) == b"%\x1b|We%\xb6`e%\xb6`e%\xb6`"
     # TODO
 
 
@@ -189,7 +194,18 @@ def test_hash64() -> None:
         8325606756057297185,
         17961889624427075301,
     )
-    # TODO
+    assert mmh3.hash64("foo", signed=False, x64arch=True) == (
+        16316970633193145697,
+        9128664383759220103,
+    )
+    assert mmh3.hash64("foo", signed=False, x64arch=False) == (
+        6968798590592097061,
+        6968798590746895717,
+    )
+    assert mmh3.hash64("foo", 0, True, False) == (
+        16316970633193145697,
+        9128664383759220103,
+    )
 
 
 def test_hash128() -> None:
@@ -201,7 +217,9 @@ def test_hash128() -> None:
     assert (
         mmh3.hash128("foo", 42, signed=True) == -124315475380607080215185174712879655950
     )
-    # TODO
+    assert (
+        mmh3.hash128("foo", 42, True, True) == -124315475380607080215185174712879655950
+    )
 
 
 def test_64bit() -> None:
