### 2024-01-04 05f4c9b
Fix mmh3_86_128 to handle big-endian
