### 2024-01-04 2ba0728
Fix inconsistent results on s390x (#52)
