### 2024-01-05 7af46c7
Revert MANIFEST.in for testing
