### 2024-01-05 b541c18
Remove readme from _mmh3
