+++ MANIFEST.in
@@ -5,7 +5,6 @@ include src/mmh3/py.typed
 include src/mmh3/_mmh3/*.c
 include src/mmh3/_mmh3/*.h
 include src/mmh3/_mmh3/FILE_HEADER
-include src/mmh3/_mmh3/README.md
 include README.md
 include LICENSE
 include CHANGELOG.md
