### 2024-01-05 53660f6
Remove _mmh3/FILE_HEADER from manifest
