+++ MANIFEST.in
@@ -4,7 +4,6 @@ include src/mmh3/*.pyi
 include src/mmh3/py.typed
 include src/mmh3/_mmh3/*.c
 include src/mmh3/_mmh3/*.h
-include src/mmh3/_mmh3/FILE_HEADER
 include README.md
 include LICENSE
 include CHANGELOG.md
