### 2024-01-05 60462e2
Explicity exclude unnecessary files from MANIFT.in
