### 2024-01-05 5e7fbe0
Add exclude to pyproject.toml
