+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "mmh3"
-version = "4.1.0-alpha.6"
+version = "4.1.0-alpha.7"
 description = "Python extension for MurmurHash (MurmurHash3), a set of fast and robust hash functions."
 readme = "README.md"
 license = {file = "LICENSE"}
@@ -41,3 +41,6 @@ Changelog = "https://github.com/hajimes/mmh3/blob/master/CHANGELOG.md"
 [tool.isort]
 profile = "black"
 src_paths = ["src/mmh3/__init__.pyi", "tests"]
+
+[tool.setuptools.packages.find]
+exclude = ["src/mmh3/_mmh3/refresh.py"]
