### 2024-01-05 fcb5180
Add exclude to setup.py
