+++ MANIFEST.in
@@ -10,6 +10,3 @@ include CODE_OF_CONDUCT.md
 include pyproject.toml
 include setup.py
 include MANIFEST.in
-exclude src/mmh3/_mmh3/refresh.py
-exclude src/mmh3/_mmh3/FILE_HEADER
-exclude src/mmh3/_mmh3/README.md
