+++ pyproject.toml
@@ -41,6 +41,3 @@ Changelog = "https://github.com/hajimes/mmh3/blob/master/CHANGELOG.md"
 [tool.isort]
 profile = "black"
 src_paths = ["src/mmh3/__init__.pyi", "tests"]
-
-[tool.setuptools.packages.find]
-exclude = ["src/mmh3/_mmh3/refresh.py"]
