+++ MANIFEST.in
@@ -2,7 +2,6 @@ include src/mmh3/*.c
 include src/mmh3/*.h
 include src/mmh3/*.pyi
 include src/mmh3/py.typed
-include src/mmh3/_mmh3/*
 include README.md
 include LICENSE
 include CHANGELOG.md
