+++ src/mmh3/mmh3module.c
@@ -11,8 +11,8 @@
 #include <byteswap.h>
 #endif
 
-#include "murmurhash3.h"
 #include "hashlib.h"
+#include "murmurhash3.h"
 
 #if defined(_MSC_VER)
 typedef signed __int8 int8_t;
