### 2024-01-06 0916881
Change the project layout for Bazel (#50)
