+++ docs/CONTRIBUTING.md
@@ -1,0 +1,8 @@
+# mmh3 contribbuting guide
+
+Follow our [Code of Conduct](./CODE_OF_CONDUCT.md)/
+
+## New contributor guide
+
+Read [README.md](README.md) to get an overview of the `mmh3` project.
+
