### 2024-01-07 be675f5
Add descriptions to contributing
