+++ docs/CONTRIBUTING.md
@@ -22,9 +22,12 @@ Before submitting a new issue, it's a good idea to check
 As of 4.1.0, the layout of the project is as follows:
 
 - `src/mmh3`
-  - `mmh3module.c`: this file
+  - `mmh3module.c`: the main file that serves as the interface between Python and the MurmurHash3 c implementation.
+  - `murmurhash.c`: implentations of the MurmurHash3 family. Auto-generated from Austin Appleby's original code. DO NOT edit this file manually. See [README in the util directory](../util/README.md) for details.
+  - `murmurhash.h`: headers and macros for MurmurHash3. Auto-gerated from `util/refresh.py`. DO NOT edit this file manually.
+  - `hashlib.h`: taken from [CPython's code base](https://github.com/python/cpython/blob/9ce0f48e918860ffa32751a85b0fe7967723e2e3/Modules/hashlib.h).
 - `util`
-  - `refresh.py`
+  - `refresh.py`: file that generates `src/mmh3/murmurhash.c` and `src/mmh3/murmurhash.h` from the original MurmurHash3 C++ code. Edit this file to modify the contents of these files.
 
 ### Testing
 
@@ -35,9 +38,17 @@ from the project root directory.
 
 #### (Optional) Testing on s390x
 
-When you changed the source in a way which may cause endian issues,
-(...)
+When you changed the source in a way which may cause endian issues, you may want
+to locally test on s390x, the only big-endian platform officially supported by
+Python.
 
-### Pull request
+[*Emulating a big-endian s390x with QEMU*](https://til.simonwillison.net/docker/emulate-s390x-with-qemu)
+by Simon Willison is a good introduction to Docker/QEMU settings for emulating
+s390x.
 
-Finally, [create a pull request](https://github.com/hajimes/mmh3/pulls).
+#### Pull request
+
+Once you've pushed your changes to your fork, you can
+[create a pull request (PR)](https://github.com/hajimes/mmh3/pulls) on the main
+project repository. Please provide a clear and detailed description of your
+changes in the PR, and reference any related issues.
