+++ docs/CONTRIBUTING.md
@@ -1,4 +1,4 @@
-# mmh3 contribbuting guide
+# mmh3 contributing guide
 
 Thank you for your interest in contributing to the `mmh3` project!
 
@@ -13,18 +13,18 @@ simply submitting a bug report or a feature suggestion
 through the [issue tracker](https://github.com/hajimes/mmh3/issues).
 
 Before submitting a new issue, it's a good idea to check
-[known issues on README](https://github.com/hajimes/mmh3#known-issues).
+[Known Issues section on README](https://github.com/hajimes/mmh3#known-issues).
 
-## Maitaining and developing the project
+## Maintaining and developing the project
 
 ### Project structure
 
 As of 4.1.0, the layout of the project is as follows:
 
 - `src/mmh3`
-  - `mmh3module.c`: the main file that serves as the interface between Python and the MurmurHash3 c implementation.
-  - `murmurhash.c`: implentations of the MurmurHash3 family. Auto-generated from Austin Appleby's original code. DO NOT edit this file manually. See [README in the util directory](../util/README.md) for details.
-  - `murmurhash.h`: headers and macros for MurmurHash3. Auto-gerated from `util/refresh.py`. DO NOT edit this file manually.
+  - `mmh3module.c`: the main file that serves as the interface between Python and the MurmurHash3 c implementations.
+  - `murmurhash.c`: implementations of the MurmurHash3 family. Auto-generated from Austin Appleby's original code. DO NOT edit this file manually. See [README in the util directory](../util/README.md) for details.
+  - `murmurhash.h`: headers and macros for MurmurHash3. Auto-generated from `util/refresh.py`. DO NOT edit this file manually.
   - `hashlib.h`: taken from [CPython's code base](https://github.com/python/cpython/blob/9ce0f48e918860ffa32751a85b0fe7967723e2e3/Modules/hashlib.h).
 - `util`
   - `refresh.py`: file that generates `src/mmh3/murmurhash.c` and `src/mmh3/murmurhash.h` from the original MurmurHash3 C++ code. Edit this file to modify the contents of these files.
@@ -38,7 +38,7 @@ from the project root directory.
 
 #### (Optional) Testing on s390x
 
-When you changed the source in a way which may cause endian issues, you may want
+When you have modified the code in a way which may cause endian issues, you may want
 to locally test on s390x, the only big-endian platform officially supported by
 Python.
 
@@ -46,7 +46,7 @@ Python.
 by Simon Willison is a good introduction to Docker/QEMU settings for emulating
 s390x.
 
-#### Pull request
+### Pull request
 
 Once you've pushed your changes to your fork, you can
 [create a pull request (PR)](https://github.com/hajimes/mmh3/pulls) on the main
