### 2024-01-09 #6124 @hajimes
@mrshu
I appreciate your constructive and careful review.

As for the community guideline, which is defined as guidelines to "1) Contribute to the software 2) Report issues or problems with the software 3) Seek support," I presume my library has already satisfied the points 2) and 3) to some extent. As a GitHub project, users may usually assume that the bug tracker is <https://github.com/hajimes/mmh3/issues>, which is also explicitly specified in [pyproject.toml](https://github.com/hajimes/mmh3/blob/master/pyproject.toml). For the "3) Seek support," I think [external tutorials listed in the See Also section of README](https://github.com/hajimes/mmh3#see-also) will answer many of the questions users would have.

However, guidelines to "1) Contribute to the software" were clearly missing in my project. To solve the issue, I added [CONTRIBUTING.md](https://github.com/hajimes/mmh3/blob/master/docs/CONTRIBUTING.md) in the latest release of the library. The document is probably not thorough or comprehensive, but I hope it will be helpful for future developers to a certain degree.

Again, I am grateful to you for taking the time to review my draft.
