### 2024-05-27 #6124 @cassiersg
**Substantial scholarly effort**: The submission is a python wrapper over a C library for a family of hash functions, and does therefore not contain "novel" features. I am pointing this out in case there is any JOSS policy regarding such wrappers. Otherwise, this is a small but high-quality library that has been maintained over many years.

**Statement of need** & **State of the field**: The documentation & paper do not discuss much murmurhash3 itself and how it compares to other hash functions. Since performance is claimed as an important characteristic, I think that a discussion of performance should be included or linked somewhere. This could provide answer such as "How does murmurhash3 compare to other hash functions (xxhash, cityhash, spookyhash, fnv...), as a function of the input size?", "How does mmh3 compare to pure-python implementations of murmurhash3?".

**Functionality documentation**: I did not see any exhaustive online API documentation. It seems that docstrings are of good quality, though. I would recommend setting up a API documentation with sphinx (e.g. on readthedocs).
