### 2024-05-27 #6124 @danielskatz
@cassiersg - In regards to your effort comment

> Substantial scholarly effort: The submission is a python wrapper over a C library for a family of hash functions, and does therefore not contain "novel" features. I am pointing this out in case there is any JOSS policy regarding such wrappers. Otherwise, this is a small but high-quality library that has been maintained over many years.

I can mention that this was a question at the time this was submitted. The author commented on it in https://github.com/openjournals/joss-reviews/issues/5487#issuecomment-1561482722 and then the JOSS editors had a discussion about it, and decided it could proceed to be reviewed.
