### 2024-06-15 #6124 @danielskatz
👋 @hajimes - I just wanted to check on how you are doing?  I know you said above "the middle of June", which has not yet passed. (As track editor, I try to check on reviews in the CSISM track where no progress has been recorded in a 2-week period.)
