+++ benchmark/bench_hash.py
@@ -1,0 +1,196 @@
+import gc
+import hashlib
+import mmh3
+import xxhash
+import time
+import pymmh3
+
+import pandas as pd
+import matplotlib.pyplot as plt
+
+"""
+The logic of the benchmarking functions is roughly based on the following C libary
+https://github.com/Cyan4973/xxHash/tree/dev/tests/bench
+"""
+
+
+TIMELOOP_NANOSEC = 1000000000  # 1 second in nanoseconds
+
+
+class BenchmarkHashFunction:
+    def __init__(self, size, total_microseconds, run_microseconds):
+        self._reset(total_microseconds, run_microseconds)
+
+    def _reset(self, total_microseconds, run_microseconds):
+        if run_microseconds == None:
+            self.run_microseconds = 1
+        else:
+            self.run_microseconds = run_microseconds
+
+        # Run the benchmark function until the time spent is greater than the run budget.
+        # defaults to 1ms (in nanoseconds)
+        self.run_budget_nanoseconds = self.run_microseconds * TIMELOOP_NANOSEC / 1000
+
+        self.fastest_nanoseconds_per_run = float("inf")
+        self.fastest_run_sum_of_return = -1
+        self.number_of_loops = 1
+
+    def run(self, data):
+        run_budget_nonoseconds = self.run_budget_nonoseconds
+
+        run_time_min_nonoseconds = run_budget_nonoseconds / 2
+
+        return mmh3.hash(data)
+
+    def benchmark_function(self, params):
+        result = {}
+
+        clock_start = time.time_ns()
+
+        for i in range(self.number_of_loops):
+            for j in range(params["number_of_blocks"]):
+                b = params["source_buffers"][j]
+                params["destinations"][j] = params["function"](b)
+
+        clock_end = time.time_ns()
+        time_spent = clock_end - clock_start
+
+        result["loop_duration_nanoseconds"] = time_spent
+        result["nanoseconds_per_run"] = time_spent / self.number_of_loops
+
+        return result
+
+    def run_timed_benchmarks(self, params):
+        time_spent = 0
+
+        WOLKLOAD_MULTIPLIER = 10
+
+        while True:
+            run_result = self.benchmark_function(params)
+
+            time_spent += run_result["loop_duration_nanoseconds"]
+
+            if run_result["loop_duration_nanoseconds"] > (
+                self.run_budget_nanoseconds / 50
+            ):
+                fastest_run_ns = min(
+                    self.fastest_nanoseconds_per_run, run_result["nanoseconds_per_run"]
+                )
+                self.number_of_loops = (
+                    int(self.run_budget_nanoseconds / fastest_run_ns) + 1
+                )
+            else:
+                self.number_of_loops *= WOLKLOAD_MULTIPLIER
+
+            if (
+                run_result["loop_duration_nanoseconds"]
+                < self.run_budget_nanoseconds / 2
+            ):
+                continue
+
+            if run_result["nanoseconds_per_run"] < self.fastest_nanoseconds_per_run:
+                self.fastest_nanoseconds_per_run = run_result["nanoseconds_per_run"]
+                self.fastest_run_sum_of_return = run_result["loop_duration_nanoseconds"]
+
+            break
+
+        result = {}
+        result["loop_duration_nanoseconds"] = time_spent
[…]
