+++ docs/requirements.txt
@@ -1,0 +1,49 @@
+#
+# This file is autogenerated by pip-compile with Python 3.12
+# by the following command:
+#
+#    pip-compile docs/requirements.in
+#
+click==8.1.7
+    # via mkdocs
+ghp-import==2.1.0
+    # via mkdocs
+jinja2==3.1.4
+    # via mkdocs
+markdown==3.6
+    # via
+    #   markdown-include
+    #   mkdocs
+markdown-include==0.8.1
+    # via -r docs/requirements.in
+markupsafe==2.1.5
+    # via
+    #   jinja2
+    #   mkdocs
+mergedeep==1.3.4
+    # via
+    #   mkdocs
+    #   mkdocs-get-deps
+mkdocs==1.6.0
+    # via -r docs/requirements.in
+mkdocs-get-deps==0.2.0
+    # via mkdocs
+packaging==24.1
+    # via mkdocs
+pathspec==0.12.1
+    # via mkdocs
+platformdirs==4.2.2
+    # via mkdocs-get-deps
+python-dateutil==2.9.0.post0
+    # via ghp-import
+pyyaml==6.0.1
+    # via
+    #   mkdocs
+    #   mkdocs-get-deps
+    #   pyyaml-env-tag
+pyyaml-env-tag==0.1
+    # via mkdocs
+six==1.16.0
+    # via python-dateutil
+watchdog==4.0.1
+    # via mkdocs
