+++ docs/contributing.md
@@ -1,4 +1,4 @@
-# mmh3 Contributing Guidelines
+# Contributing
 
 Thank you for your interest in contributing to the `mmh3` project!
 
@@ -22,7 +22,7 @@ As of 4.1.0, the layout of the project is as follows:
 
 - `src/mmh3`
   - `mmh3module.c`: the main file that serves as the interface between Python and the MurmurHash3 c implementations.
-  - `murmurhash.c`: implementations of the MurmurHash3 family. Auto-generated from Austin Appleby's original code. DO NOT edit this file manually. See [README in the util directory](../util/README.md) for details.
+  - `murmurhash.c`: implementations of the MurmurHash3 family. Auto-generated from Austin Appleby's original code. DO NOT edit this file manually. See [README in the util directory](https://github.com/hajimes/mmh3/blob/master/util/README.md) for details.
   - `murmurhash.h`: headers and macros for MurmurHash3. Auto-generated from `util/refresh.py`. DO NOT edit this file manually.
   - `hashlib.h`: taken from [CPython's code base](https://github.com/python/cpython/blob/9ce0f48e918860ffa32751a85b0fe7967723e2e3/Modules/hashlib.h).
 - `util`
