+++ docs-old/CODE_OF_CONDUCT.md
@@ -1,3 +1,0 @@
-# Code of Conduct
-
-This project has adopted the ACM Code of Ethics and Professional Conduct ([https://www.acm.org/code-of-ethics](https://www.acm.org/code-of-ethics)) in the version adopted in June 22nd, 2018.
