+++ docs/index.md
@@ -1,1 +1,2 @@
+<!-- markdownlint-disable-next-line MD041 -->
 {!README.md!}
