+++ docs/requirements.txt
@@ -5,41 +5,72 @@
 #    pip-compile docs/requirements.in
 #
 click==8.1.7
-    # via mkdocs
+    # via
+    #   mkdocs
+    #   mkdocstrings
+colorama==0.4.6
+    # via griffe
 ghp-import==2.1.0
     # via mkdocs
+griffe==0.48.0
+    # via mkdocstrings-python
 jinja2==3.1.4
-    # via mkdocs
+    # via
+    #   mkdocs
+    #   mkdocstrings
 markdown==3.6
     # via
     #   markdown-include
     #   mkdocs
+    #   mkdocs-autorefs
+    #   mkdocstrings
+    #   pymdown-extensions
 markdown-include==0.8.1
     # via -r docs/requirements.in
 markupsafe==2.1.5
     # via
     #   jinja2
     #   mkdocs
+    #   mkdocs-autorefs
+    #   mkdocstrings
 mergedeep==1.3.4
     # via
     #   mkdocs
     #   mkdocs-get-deps
 mkdocs==1.6.0
-    # via -r docs/requirements.in
+    # via
+    #   -r docs/requirements.in
+    #   mkdocs-autorefs
+    #   mkdocstrings
+mkdocs-autorefs==1.0.1
+    # via mkdocstrings
 mkdocs-get-deps==0.2.0
     # via mkdocs
+mkdocstrings[python]==0.25.1
+    # via
+    #   -r docs/requirements.in
+    #   mkdocstrings-python
+mkdocstrings-python==1.10.5
+    # via mkdocstrings
+mmh3==4.1.0
+    # via -r docs/requirements.in
 packaging==24.1
     # via mkdocs
 pathspec==0.12.1
     # via mkdocs
 platformdirs==4.2.2
-    # via mkdocs-get-deps
+    # via
+    #   mkdocs-get-deps
+    #   mkdocstrings
+pymdown-extensions==10.8.1
+    # via mkdocstrings
 python-dateutil==2.9.0.post0
     # via ghp-import
 pyyaml==6.0.1
     # via
     #   mkdocs
     #   mkdocs-get-deps
+    #   pymdown-extensions
     #   pyyaml-env-tag
 pyyaml-env-tag==0.1
     # via mkdocs
