+++ .readthedocs.yaml
@@ -15,3 +15,8 @@ mkdocs:
 
 # Build all formats
 formats: all
+
+# Optionally declare the Python requirements required to build your docs
+python:
+   install:
+   - requirements: docs/requirements.txt
