+++ docs/api.md
@@ -1,3 +1,0 @@
-# API Reference
-
-::: mmh3
