+++ docs/index.md
@@ -1,2 +1,0 @@
-<!-- markdownlint-disable-next-line MD041 -->
-{!README.md!}
