+++ docs/requirements.txt
@@ -2,79 +2,68 @@
 # This file is autogenerated by pip-compile with Python 3.12
 # by the following command:
 #
-#    pip-compile docs/requirements.in
+#    pip-compile requirements.in
 #
-click==8.1.7
+alabaster==0.7.16
+    # via sphinx
+babel==2.15.0
+    # via sphinx
+certifi==2024.7.4
+    # via requests
+charset-normalizer==3.3.2
+    # via requests
+docutils==0.21.2
     # via
-    #   mkdocs
-    #   mkdocstrings
-colorama==0.4.6
-    # via griffe
-ghp-import==2.1.0
-    # via mkdocs
-griffe==0.48.0
-    # via mkdocstrings-python
+    #   myst-parser
+    #   sphinx
+idna==3.7
+    # via requests
+imagesize==1.4.1
+    # via sphinx
 jinja2==3.1.4
     # via
-    #   mkdocs
-    #   mkdocstrings
-markdown==3.6
+    #   myst-parser
+    #   sphinx
+markdown-it-py==3.0.0
     # via
-    #   markdown-include
-    #   mkdocs
-    #   mkdocs-autorefs
-    #   mkdocstrings
-    #   pymdown-extensions
-markdown-include==0.8.1
-    # via -r docs/requirements.in
+    #   mdit-py-plugins
+    #   myst-parser
 markupsafe==2.1.5
-    # via
-    #   jinja2
-    #   mkdocs
-    #   mkdocs-autorefs
-    #   mkdocstrings
-mergedeep==1.3.4
-    # via
-    #   mkdocs
-    #   mkdocs-get-deps
-mkdocs==1.6.0
-    # via
-    #   -r docs/requirements.in
-    #   mkdocs-autorefs
-    #   mkdocstrings
-mkdocs-autorefs==1.0.1
-    # via mkdocstrings
-mkdocs-get-deps==0.2.0
-    # via mkdocs
-mkdocstrings[python]==0.25.1
-    # via
-    #   -r docs/requirements.in
-    #   mkdocstrings-python
-mkdocstrings-python==1.10.5
-    # via mkdocstrings
-mmh3==4.1.0
-    # via -r docs/requirements.in
+    # via jinja2
+mdit-py-plugins==0.4.1
+    # via myst-parser
+mdurl==0.1.2
+    # via markdown-it-py
+myst-parser==3.0.1
+    # via -r requirements.in
 packaging==24.1
-    # via mkdocs
-pathspec==0.12.1
-    # via mkdocs
-platformdirs==4.2.2
-    # via
-    #   mkdocs-get-deps
-    #   mkdocstrings
-pymdown-extensions==10.8.1
-    # via mkdocstrings
-python-dateutil==2.9.0.post0
-    # via ghp-import
+    # via sphinx
+pygments==2.18.0
+    # via sphinx
 pyyaml==6.0.1
+    # via myst-parser
+requests==2.32.3
+    # via sphinx
+shibuya==2024.7.13
+    # via -r requirements.in
+snowballstemmer==2.2.0
+    # via sphinx
+sphinx==7.4.5
     # via
-    #   mkdocs
-    #   mkdocs-get-deps
-    #   pymdown-extensions
-    #   pyyaml-env-tag
-pyyaml-env-tag==0.1
-    # via mkdocs
-six==1.16.0
-    # via python-dateutil
-watchdog==4.0.1
-    # via mkdocs
+    #   -r requirements.in
+    #   myst-parser
+    #   shibuya
+sphinxcontrib-applehelp==1.0.8
+    # via sphinx
+sphinxcontrib-devhelp==1.0.6
+    # via sphinx
[…]
