+++ docs/index.rst
@@ -1,0 +1,25 @@
+mmh3 documentation
+==================
+mmh3 is a Python extension for `MurmurHash (MurmurHash3) <https://en.wikipedia.org/wiki/MurmurHash>`_, a set of fast and robust non-cryptographic hash functions invented by Austin Appleby.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: User Guideline
+
+   Quickstart<readme_link>
+   api
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Project documentation
+
+   contributing
+   changelog
+   CODE_OF_CONDUCT
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
