+++ docs/conf.py
@@ -9,23 +9,23 @@
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'mmh3'
-copyright = '2011-2024, Hajime Senuma'
-author = 'Hajime Senuma'
+project = "mmh3"
+copyright = "2011-2024, Hajime Senuma"
+author = "Hajime Senuma"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ['sphinx.ext.autodoc', 'myst_parser']
+extensions = ["sphinx.ext.autodoc", "myst_parser"]
 
-templates_path = ['_templates']
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+templates_path = ["_templates"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'shibuya'
-html_static_path = ['_static']
+html_theme = "shibuya"
+html_static_path = ["_static"]
 html_theme_options = {
-    'github_url': 'https://github.com/hajimes/mmh3',
+    "github_url": "https://github.com/hajimes/mmh3",
 }
