+++ docs/readme_link.md
@@ -1,2 +1,3 @@
+<!-- markdownlint-disable -->
 ```{include} ../README.md
 ```
