### 2024-07-20 1455737
Remove mkdocs settings
