+++ mkdocs.yml
@@ -1,32 +1,0 @@
-site_name: mmh3
-docs_dir: docs
-repo_url: https://github.com/hajimes/mmh3
-theme:
-  name: readthedocs
-  highlightjs: true
-  hljs_languages:
-      - shell
-  collapse_navigation: true
-  navigation_depth: 3
-nav:
-  - Home: index.md
-  - api.md
-  - contributing.md
-  - changelog.md
-  - CODE_OF_CONDUCT.md
-plugins:
-  - search
-  - mkdocstrings:
-      handlers:
-        # See: https://mkdocstrings.github.io/python/usage/
-        python:
-          options:
-            docstring_style: sphinx
-            preload_modules:
-            - 'mmh3'
-            load_external_modules: true
-            filters:
-            - "!^_"
-markdown_extensions:
-  - markdown_include.include:
-      base_path: .
