+++ .readthedocs.yaml
@@ -20,3 +20,5 @@ formats: all
 python:
    install:
    - requirements: docs/requirements.txt
+   - method: pip
+     path: .
