+++ .readthedocs.yaml
@@ -15,7 +15,9 @@ sphinx:
   configuration: docs/conf.py
 
 # Build all formats
-formats: all
+formats:
+  - pdf
+  - epub
 
 # Optionally declare the Python requirements required to build your docs
 python:
