+++ docs/readme_link.md
@@ -1,0 +1,3 @@
+<!-- markdownlint-disable -->
+```{include} ../README.md
+```
