+++ docs/requirements.in
@@ -1,0 +1,3 @@
+myst-parser
+shibuya
+sphinx

[…]
