### 2024-07-21 64c7e45
Add a doc page for benchmarks
