+++ docs/benchmarks.md
@@ -1,0 +1,11 @@
+# Benchmarks
+
+```{figure} images/throughput_small_inputs.png
+:alt: Throughput for small inputs
+:align: center
+
+Figure 1: Throughput for small inputs
+```
+
+As you see, the throughput of `mmh3` is significantly
+higher than that of `hashlib` (`md5` and `sha1`) for small inputs.
