+++ src/mmh3/mmh3module.c
@@ -479,30 +479,19 @@ MMH3Hasher32_get_name(PyObject *self, void *closure)
 }
 
 static PyGetSetDef MMH3Hasher32_getsetters[] = {
-    {"digest_size", (getter)MMH3Hasher32_get_digest_size, NULL, NULL, NULL},
-    {"block_size", (getter)MMH3Hasher32_get_block_size, NULL, NULL, NULL},
-    {"name", (getter)MMH3Hasher32_get_name, NULL, NULL, NULL},
+    {"digest_size", (getter)MMH3Hasher32_get_digest_size, NULL,
+     "number of bytes in this hashes output", NULL},
+    {"block_size", (getter)MMH3Hasher32_get_block_size, NULL,
+     "number of bytes of the internal block of this algorithm", NULL},
+    {"name", (getter)MMH3Hasher32_get_name, NULL,
+     "the hash algorithm being used by this object", NULL},
     {NULL} /* Sentinel */
 };
 
 PyDoc_STRVAR(
     MMH3Hasher32Type_doc,
     "An mmh3_32 is an object used to calculate the murmurhash3_x86_32 hash\n"
-    "of a string of information.\n"
-    "\n"
-    "Methods:\n"
-    "\n"
-    "update(input) -- updates the current digest with an additional string\n"
-    "digest() -- return the current digest value\n"
-    "sintdigest() -- return the current digest as a 32 bit signed integer\n"
-    "uintdigest() -- return the current digest as a 32 bit unsigned integer\n"
-    "copy() -- return a copy of the current object\n"
-    "\n"
-    "Attributes:\n"
-    "\n"
-    "name -- the hash algorithm being used by this object\n"
-    "digest_size -- number of bytes in this hashes output\n"
-    "block_size -- number of bytes of the internal block of this algorithm");
+    "of a string of information.");
 
 static PyTypeObject MMH3Hasher32Type = {
     PyVarObject_HEAD_INIT(NULL, 0).tp_name = "mmh3.mmh3_32",
@@ -801,35 +790,19 @@ MMH3Hasher128x64_get_name(PyObject *self, void *closure)
 }
 
 static PyGetSetDef MMH3Hasher128x64_getsetters[] = {
-    {"digest_size", (getter)MMH3Hasher128x64_get_digest_size, NULL, NULL,
-     NULL},
-    {"block_size", (getter)MMH3Hasher128x64_get_block_size, NULL, NULL, NULL},
-    {"name", (getter)MMH3Hasher128x64_get_name, NULL, NULL, NULL},
+    {"digest_size", (getter)MMH3Hasher128x64_get_digest_size, NULL,
+     "number of bytes in this hashes output", NULL},
+    {"block_size", (getter)MMH3Hasher128x64_get_block_size, NULL,
+     "number of bytes of the internal block of this algorithm", NULL},
+    {"name", (getter)MMH3Hasher128x64_get_name, NULL,
+     "the hash algorithm being used by this object", NULL},
     {NULL} /* Sentinel */
 };
 
 PyDoc_STRVAR(
     MMH3Hasher128x64Type_doc,
     "An mmh3_x64_128 is an object used to calculate the murmurhash3_x64_128\n"
-    "hash of a string of information.\n"
-    "\n"
-    "Methods:\n"
-    "\n"
-    "update(input) -- updates the current digest with an additional string\n"
-    "digest() -- return the current digest value\n"
-    "sintdigest() -- return the current digest as a 128 bit signed integer\n"
-    "uintdigest() -- return the current digest as a 128 bit unsigned integer\n"
-    "stupledigest() -- return the current digest as a tuple of two 64 bit\n"
-    "                  signed integers\n"
[…]
