### 2024-08-15 #6124 @hajimes
Dear @danielskatz,
I deeply apologize for the repeated delay.

As for the progress, actually a large part of the work was already done three weeks ago.
I have completed one of the two suggestions from the second reviewer (introducing an RTD-style doc), and pushed it to the main branch.
For the other suggestion (benchmarking), I wrote a base program (following [xxhash's implementation](https://github.com/Cyan4973/xxHash/tree/dev/tests/bench)) and pushed it into a non-main branch. Since the code can generate a comparison graph for one performance metric, the remaining task is to write code for generating the other three variations of performance metrics and provide explanations for these graphs.

I return to my home country Japan today, so I'll try to finish the update in the next few days.
