### 2024-08-22 ea50371
Add throughput benchmarking for large inputs
