### 2024-08-22 a02db3d
Fix type hints
