### 2024-08-24 fc536c3
Add argparse modes
