+++ .github/workflows/wheels.yml
@@ -92,7 +92,7 @@ jobs:
     needs: [build_wheels, build_sdist]
     steps:
       - name: Set up Python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: '3.12'
       - name: Set up built items
