### 2024-08-24 042aa2a
Bump super-linter/super-linter from 5.7.2 to 7.0.0
Bumps [super-linter/super-linter](https://github.com/super-linter/super-linter) from 5.7.2 to 7.0.0.
- [Release notes](https://github.com/super-linter/super-linter/releases)
- [Changelog](https://github.com/super-linter/super-linter/blob/main/CHANGELOG.md)
- [Commits](https://github.com/super-linter/super-linter/compare/v5.7.2...v7.0.0)

---
updated-dependencies:
- dependency-name: super-linter/super-linter
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
