### 2024-08-25 d6e4a3f
Bump super-linter to 7.0.0; w/o PYLINT, CHECKOV
