+++ src/mmh3/mmh3module.c
@@ -4,6 +4,7 @@
 #endif
 
 #include <Python.h>
+
 #include <stdio.h>
 #include <string.h>
 
@@ -22,9 +23,9 @@ typedef unsigned __int8 uint8_t;
 typedef unsigned __int32 uint32_t;
 typedef unsigned __int64 uint64_t;
 // Other compilers
-#else  // defined(_MSC_VER)
+#else // defined(_MSC_VER)
 #include <stdint.h>
-#endif  // defined(_MSC_VER)
+#endif // defined(_MSC_VER)
 
 #define MMH3_32_DIGESTSIZE 4
 #define MMH3_128_DIGESTSIZE 16
@@ -35,62 +36,56 @@ typedef unsigned __int64 uint64_t;
 //-----------------------------------------------------------------------------
 // One shot functions
 
-PyDoc_STRVAR(mmh3_hash_doc,
-             "hash(key[, seed=0, signed=True]) -> 32-bit int\n\n"
-             "Return a hash value as a 32-bit integer. "
-             "Calculated by the MurmurHash3_x86_32 algorithm.");
+PyDoc_STRVAR(mmh3_hash_doc, "hash(key[, seed=0, signed=True]) -> 32-bit int\n\n"
+                            "Return a hash value as a 32-bit integer. "
+                            "Calculated by the MurmurHash3_x86_32 algorithm.");
 
-static PyObject *
-mmh3_hash(PyObject *self, PyObject *args, PyObject *keywds)
-{
-    const char *target_str;
-    Py_ssize_t target_str_len;
-    uint32_t seed = 0;
-    int32_t result[1];
-    long long_result = 0;
-    unsigned char is_signed = 1;
+static PyObject *mmh3_hash(PyObject *self, PyObject *args, PyObject *keywds) {
+  const char *target_str;
+  Py_ssize_t target_str_len;
+  uint32_t seed = 0;
+  int32_t result[1];
+  long long_result = 0;
+  unsigned char is_signed = 1;
 
-    static char *kwlist[] = {(char *)"key", (char *)"seed", (char *)"signed",
-                             NULL};
+  static char *kwlist[] = {(char *)"key", (char *)"seed", (char *)"signed",
+                           NULL};
 
 #ifndef _MSC_VER
 #if __LONG_WIDTH__ == 64 || defined(__APPLE__)
-    static uint64_t mask[] = {0x0ffffffff, 0xffffffffffffffff};
+  static uint64_t mask[] = {0x0ffffffff, 0xffffffffffffffff};
 #endif
 #endif
 
-    if (!PyArg_ParseTupleAndKeywords(args, keywds, "s#|IB", kwlist,
-                                     &target_str, &target_str_len, &seed,
-                                     &is_signed)) {
-        return NULL;
-    }
+  if (!PyArg_ParseTupleAndKeywords(args, keywds, "s#|IB", kwlist, &target_str,
+                                   &target_str_len, &seed, &is_signed)) {
+    return NULL;
+  }
 
-    murmurhash3_x86_32(target_str, target_str_len, seed, result);
+  murmurhash3_x86_32(target_str, target_str_len, seed, result);
 
 #if defined(_MSC_VER)
-    /* for Windows envs */
-    long_result = result[0];
-    if (is_signed == 1) {
-        return PyLong_FromLong(long_result);
-    }
-    else {
-        return PyLong_FromUnsignedLong(long_result);
-    }
-#else  // defined(_MSC_VER)
-    /* for standard envs */
-#if __LONG_WIDTH__ == 64 || defined(__APPLE__)
-    long_result = result[0] & mask[is_signed];
+  /* for Windows envs */
[…]
