+++ src/mmh3/murmurhash3.h
@@ -43,11 +43,11 @@ typedef unsigned __int64 uint64_t;
 
 // Other compilers
 
-#else  // defined(_MSC_VER)
+#else // defined(_MSC_VER)
 
 #include <stdint.h>
 
-#endif  // !defined(_MSC_VER)
+#endif // !defined(_MSC_VER)
 
 //-----------------------------------------------------------------------------
 // Platform-specific functions and macros
@@ -67,7 +67,7 @@ typedef unsigned __int64 uint64_t;
 
 // Other compilers
 
-#else  // defined(_MSC_VER)
+#else // defined(_MSC_VER)
 
 #if ((__GNUC__ > 4) || (__GNUC__ == 4 && GNUC_MINOR >= 4))
 /* gcc version >= 4.4 4.1 = RHEL 5, 4.4 = RHEL 6. Don't inline for RHEL 5 gcc
@@ -77,16 +77,12 @@ typedef unsigned __int64 uint64_t;
 #define FORCE_INLINE
 #endif
 
-static FORCE_INLINE uint32_t
-rotl32(uint32_t x, int8_t r)
-{
-    return (x << r) | (x >> (32 - r));
+static FORCE_INLINE uint32_t rotl32(uint32_t x, int8_t r) {
+  return (x << r) | (x >> (32 - r));
 }
 
-static FORCE_INLINE uint64_t
-rotl64(uint64_t x, int8_t r)
-{
-    return (x << r) | (x >> (64 - r));
+static FORCE_INLINE uint64_t rotl64(uint64_t x, int8_t r) {
+  return (x << r) | (x >> (64 - r));
 }
 
 #define ROTL32(x, y) rotl32(x, y)
@@ -94,29 +90,25 @@ rotl64(uint64_t x, int8_t r)
 
 #define BIG_CONSTANT(x) (x##LLU)
 
-#endif  // !defined(_MSC_VER)
+#endif // !defined(_MSC_VER)
 
 //-----------------------------------------------------------------------------
 // Block read - if your platform needs to do endian-swapping or can only
 // handle aligned reads, do the conversion here
 
-static FORCE_INLINE uint32_t
-getblock32(const uint32_t *p, Py_ssize_t i)
-{
+static FORCE_INLINE uint32_t getblock32(const uint32_t *p, Py_ssize_t i) {
 #if defined(__BYTE_ORDER__) && (__BYTE_ORDER__ == __ORDER_BIG_ENDIAN__)
-    return bswap_32(p[i]);
+  return bswap_32(p[i]);
 #else
-    return p[i];
+  return p[i];
 #endif
 }
 
-static FORCE_INLINE uint64_t
-getblock64(const uint64_t *p, Py_ssize_t i)
-{
+static FORCE_INLINE uint64_t getblock64(const uint64_t *p, Py_ssize_t i) {
 #if defined(__BYTE_ORDER__) && (__BYTE_ORDER__ == __ORDER_BIG_ENDIAN__)
-    return bswap_64(p[i]);
+  return bswap_64(p[i]);
 #else
-    return p[i];
+  return p[i];
 #endif
 }
 
@@ -124,108 +116,95 @@ getblock64(const uint64_t *p, Py_ssize_t i)
 // Building blocks for multiply and rotate (MUR) operations.
 // Names are taken from Google Guava's implementation
 
-static FORCE_INLINE uint32_t
-mixK1(uint32_t k1)
-{
-    const uint32_t c1 = 0xcc9e2d51;
[…]
