### 2024-08-26 0da19ba
Re-enable VALIDATE_PYTHON_PYLINT
