+++ pyproject.toml
@@ -40,4 +40,15 @@ Changelog = "https://github.com/hajimes/mmh3/blob/master/docs/changelog.md"
 
 [tool.isort]
 profile = "black"
-src_paths = ["src/mmh3/__init__.pyi", "tests"]
+src_paths = ["src/mmh3/__init__.pyi", "util", "tests"]
+
+[tool.pylint.main]
+# Use multiple processes to speed up Pylint.
+# The value 0 specifies the number of processors to be auto-detected.
+# This setting can be found in the template file of super-linter 7.0.0.
+jobs=0
+
+[tool.pylint."messages control"]
+# import-error: An error tricky to resolve, especially on super-linter.
+# wrong-import-order: Respect isort's import order.
+disable = ["import-error", "wrong-import-order"]
