### 2024-08-26 c9dd7cd
Re-enable PEP 7-compliant .clang-format
