+++ .markdown-lint.yml
@@ -1,6 +1,3 @@
-# Default state for all rules
-default: true
-
 # MD024/no-duplicate-heading : Multiple headings with the same content : https://github.com/DavidAnson/markdownlint/blob/v0.34.0/doc/md024.md
 MD024:
   # Only check sibling headings (default is false)
