+++ docs/CODE_OF_CONDUCT.md
@@ -1,3 +1,5 @@
 # Code of Conduct
 
-This project adheres the ACM Code of Ethics and Professional Conduct ([https://www.acm.org/code-of-ethics](https://www.acm.org/code-of-ethics)), as specified in the version adopted in June 22nd, 2018.
+This project adheres the ACM Code of Ethics and Professional Conduct
+([https://www.acm.org/code-of-ethics](https://www.acm.org/code-of-ethics)), as
+specified in the version adopted in June 22nd, 2018.
