+++ README.md
@@ -9,13 +9,23 @@
 [![Total Downloads](https://static.pepy.tech/badge/mmh3)](https://pepy.tech/project/mmh3?versions=*&versions=4.*&versions=3.*&versions=2.*)
 [![Recent Downloads](https://static.pepy.tech/badge/mmh3/month)](https://pepy.tech/project/mmh3?versions=*&versions=4.*&versions=3.*&versions=2.*)
 
-mmh3 is a Python extension for [MurmurHash (MurmurHash3)](https://en.wikipedia.org/wiki/MurmurHash), a set of fast and robust non-cryptographic hash functions invented by Austin Appleby.
+mmh3 is a Python extension for
+[MurmurHash (MurmurHash3)](https://en.wikipedia.org/wiki/MurmurHash), a set of
+fast and robust non-cryptographic hash functions invented by Austin Appleby.
 
-Combined with probabilistic techniques like a [Bloom filter](https://en.wikipedia.org/wiki/Bloom_filter), [MinHash](https://en.wikipedia.org/wiki/MinHash), and [feature hashing](https://en.wikipedia.org/wiki/Feature_hashing), mmh3 allows you to develop high-performance systems in fields such as data mining, machine learning, and natural language processing.
+Combined with probabilistic techniques like a
+[Bloom filter](https://en.wikipedia.org/wiki/Bloom_filter),
+[MinHash](https://en.wikipedia.org/wiki/MinHash), and
+[feature hashing](https://en.wikipedia.org/wiki/Feature_hashing), mmh3 allows
+you to develop high-performance systems in fields such as data mining, machine
+learning, and natural language processing.
 
-Another common use of mmh3 is to [calculate favicon hashes](https://gist.github.com/yehgdotnet/b9dfc618108d2f05845c4d8e28c5fc6a) used by [Shodan](https://www.shodan.io), the world's first IoT search engine.
+Another common use of mmh3 is to
+[calculate favicon hashes](https://gist.github.com/yehgdotnet/b9dfc618108d2f05845c4d8e28c5fc6a)
+used by [Shodan](https://www.shodan.io), the world's first IoT search engine.
 
-This page offers a quick start guide. For more detailed information, see the [documentation](https://mmh3.readthedocs.io/en/latest/).
+This page offers a quick start guide. For more detailed information, see the
+[documentation](https://mmh3.readthedocs.io/en/latest/).
 
 ## How to use
 
@@ -58,9 +68,11 @@ Other functions:
 b'V\x8f}\xad\x8eNM\xa84\x07FU\x9c\xc4\xcc\x8e'
 ```
 
-Beware that `hash64` returns **two** values, because it uses the 128-bit version of MurmurHash3 as its backend.
+Beware that `hash64` returns **two** values, because it uses the 128-bit version
+of MurmurHash3 as its backend.
 
-`hash_from_buffer` hashes byte-likes without memory copying. The method is suitable when you hash a large memory-view such as `numpy.ndarray`.
+`hash_from_buffer` hashes byte-likes without memory copying. The method is
+suitable when you hash a large memory-view such as `numpy.ndarray`.
 
 ```shell
 >>> mmh3.hash_from_buffer(numpy.random.rand(100))
@@ -69,7 +81,9 @@ Beware that `hash64` returns **two** values, because it uses the 128-bit version
 3812874078
 ```
 
-`hash64`, `hash128`, and `hash_bytes` have the third argument for architecture optimization (keyword arg: `x64arch`). Use True for x64 and False for x86 (default: True):
+`hash64`, `hash128`, and `hash_bytes` have the third argument for architecture
[…]
