+++ docs/changelog.md
@@ -3,7 +3,8 @@
 All notable changes to this project will be documented here.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
-This project has adhered to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) since version 3.0.0.
+This project has adhered to
+[Semantic Versioning](https://semver.org/spec/v2.0.0.html) since version 3.0.0.
 
 ## [[Unreleased]]
 
@@ -13,7 +14,8 @@ This project has adhered to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Change the format of the changelog to conform to the Keep a Changelog standard.
+- Change the format of the changelog to conform to the Keep a Changelog
+  standard.
 
 ## [4.1.0] - 2024-01-09
 
@@ -23,9 +25,11 @@ This project has adhered to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fix issues with Bazel by changing the directory structure of the project (<https://github.com/hajimes/mmh3/issues/50>).
+- Fix issues with Bazel by changing the directory structure of the project
+  (<https://github.com/hajimes/mmh3/issues/50>).
 - Fix incorrect type hints (<https://github.com/hajimes/mmh3/issues/51>).
-- Fix invalid results on s390x when the arg `x64arch` of `hash64` or `hash_bytes` is set to `False` (<https://github.com/hajimes/mmh3/issues/52>).
+- Fix invalid results on s390x when the arg `x64arch` of `hash64` or
+  `hash_bytes` is set to `False` (<https://github.com/hajimes/mmh3/issues/52>).
 
 ## [4.0.1] - 2023-07-14
 
@@ -43,9 +47,12 @@ The major version bump is due to the backward incompatible changes.
 
 ### Added
 
-- Add experimental support for `hashlib`-compliant hasher classes (<https://github.com/hajimes/mmh3/issues/39>). Note that they are not yet fully tuned for performance.
+- Add experimental support for `hashlib`-compliant hasher classes
+  (<https://github.com/hajimes/mmh3/issues/39>). Note that they are not yet
+  fully tuned for performance.
 - Add support for type hints (<https://github.com/hajimes/mmh3/issues/44>).
-- Add wheels for more platforms (`musllinux`, `s390x`, `win_arm64`, and `macosx_universal2`).
+- Add wheels for more platforms (`musllinux`, `s390x`, `win_arm64`, and
+  `macosx_universal2`).
 - Add a code of conduct (the ACM Code of Ethics and Professional Conduct).
 
 ### Changed
@@ -56,29 +63,43 @@ The major version bump is due to the backward incompatible changes.
 
 - Drop support for Python 3.7, as it will reach the end of life on 2023-06-27.
 - Backward incompatible changes:
-  - A hash function now returns the same value under big-endian platforms as that under little-endian ones (<https://github.com/hajimes/mmh3/issues/47>).
-  - Remove the `__version__` constant from the module (<https://github.com/hajimes/mmh3/issues/42>). Use `importlib.metadata` instead.
+  - A hash function now returns the same value under big-endian platforms as
+    that under little-endian ones (<https://github.com/hajimes/mmh3/issues/47>).
+  - Remove the `__version__` constant from the module
+    (<https://github.com/hajimes/mmh3/issues/42>). Use `importlib.metadata`
+    instead.
 
 ## [3.1.0] - 2023-03-24
 
 ### Added
 
[…]
