+++ util/README.md
@@ -5,12 +5,14 @@ This directory contains C files that were generated from the
 
 ## Updating \_mmh3
 
-Try `git submodule update --init` to fetch Appleby's original SMHasher project as a github submodule.
-Then, run the `refresh.py` script to generate PEP 7-compliant C code from the original project, instead of editing `murmurhash3.*` files manually.
-Add transformation code to the `refresh.py` script to perform further edits.
+Try `git submodule update --init` to fetch Appleby's original SMHasher project
+as a github submodule. Then, run the `refresh.py` script to generate PEP
+7-compliant C code from the original project, instead of editing `murmurhash3.*`
+files manually. Add transformation code to the `refresh.py` script to perform
+further edits.
 
-After file generation, use `clang-format` to format the generated code.
-Try `clang-format -i src/mmh3/*.{c,h}` from the project's top-level directory.
+After file generation, use `clang-format` to format the generated code. Try
+`clang-format -i src/mmh3/*.{c,h}` from the project's top-level directory.
 
 ## Local files
 
