### 2024-08-26 6f840d6
Fix minor typos in README.md
