+++ docs/changelog.md
@@ -11,6 +11,9 @@ This project has adhered to
 ### Added
 
 - (planned: Add support for Python 3.13)
+- Add Read the Docs documentation (<https://github.com/hajimes/mmh3/issues/54>).
+- (planned: Document benchmark results
+  (<https://github.com/hajimes/mmh3/issues/53>)).
 
 ### Changed
 
