+++ docs/conf.py
@@ -1,5 +1,5 @@
 # pylint: disable=C0114,C0103
-# # Configuration file for the Sphinx documentation builder.
+# Configuration file for the Sphinx documentation builder.
 #
 # For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
