+++ docs/changelog_link.md
@@ -1,0 +1,5 @@
+<!-- markdownlint-disable -->
+
+```{include} ../CHANGELOG.md
+
+```
