### 2024-08-26 23f08ff
Adopt the Keep a Changelog format (#63)
