+++ CHANGELOG.md
@@ -1,0 +1,237 @@
+# Changelog
+
+All notable changes to this project will be documented here.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
+This project has adhered to
+[Semantic Versioning](https://semver.org/spec/v2.0.0.html) since version 3.0.0.
+
+## [Unreleased]
+
+### Added
+
+- (planned: Add support for Python 3.13)
+- Add Read the Docs documentation (<https://github.com/hajimes/mmh3/issues/54>).
+- (planned: Document benchmark results
+  (<https://github.com/hajimes/mmh3/issues/53>)).
+
+### Changed
+
+- Change the format of the changelog to conform to the Keep a Changelog
+  standard.
+
+## [4.1.0] - 2024-01-09
+
+### Added
+
+- Add support for Python 3.12.
+
+### Fixed
+
+- Fix issues with Bazel by changing the directory structure of the project
+  (<https://github.com/hajimes/mmh3/issues/50>).
+- Fix incorrect type hints (<https://github.com/hajimes/mmh3/issues/51>).
+- Fix invalid results on s390x when the arg `x64arch` of `hash64` or
+  `hash_bytes` is set to `False` (<https://github.com/hajimes/mmh3/issues/52>).
+
+## [4.0.1] - 2023-07-14
+
+### Changed
+
+- Refactor the project structure (<https://github.com/hajimes/mmh3/issues/48>).
+
+### Fixed
+
+- Fix incorrect type hints.
+
+## [4.0.0] - 2023-05-22
+
+The major version bump is due to the backward incompatible changes.
+
+### Added
+
+- Add experimental support for `hashlib`-compliant hasher classes
+  (<https://github.com/hajimes/mmh3/issues/39>). Note that they are not yet
+  fully tuned for performance.
+- Add support for type hints (<https://github.com/hajimes/mmh3/issues/44>).
+- Add wheels for more platforms (`musllinux`, `s390x`, `win_arm64`, and
+  `macosx_universal2`).
+- Add a code of conduct (the ACM Code of Ethics and Professional Conduct).
+
+### Changed
+
+- Switch license from CC0 to MIT (<https://github.com/hajimes/mmh3/issues/43>).
+
+### Removed
+
+- Drop support for Python 3.7, as it will reach the end of life on 2023-06-27.
+- Backward incompatible changes:
+  - A hash function now returns the same value under big-endian platforms as
+    that under little-endian ones (<https://github.com/hajimes/mmh3/issues/47>).
+  - Remove the `__version__` constant from the module
+    (<https://github.com/hajimes/mmh3/issues/42>). Use `importlib.metadata`
+    instead.
+
+## [3.1.0] - 2023-03-24
+
+### Added
+
+- Add support for Python 3.10 and 3.11. Thanks
+  [wouter bolsterlee](https://github.com/wbolster) and
+  [Dušan Nikolić](https://github.com/n-dusan)!
+- Add support for 32-bit architectures such as `i686` and `armv7l`. From now on,
+  `hash` and `hash_from_buffer` on these architectures will generate the same
+  hash values as those on other environments. Thanks
+  [Danil Shein](https://github.com/dshein-alt)!
+- In relation to the above, `manylinux2014_i686` wheels are now available.
+- Support for hashing huge data (>16GB). Thanks
+  [arieleizenberg](https://github.com/arieleizenberg)!
+
[…]
