+++ README.md
@@ -129,7 +129,7 @@ b'\x82_n\xdd \xac\xb6j\xef\x99\xb1e\xc4\n\xc9\xfd'
 
 ## Changelog
 
-See [Changelog](https://mmh3.readthedocs.io/en/latest/changelog.html) for the
+See [Changelog](https://mmh3.readthedocs.io/en/latest/changelog_link.html) for the
 complete changelog.
 
 ### [Unreleased]
