### 2024-08-27 78aee6a
Relocate document in util dir to CONTRIBUTING
