+++ docs/index.rst
@@ -8,13 +8,13 @@ mmh3 is a Python extension for `MurmurHash (MurmurHash3) <https://en.wikipedia.o
 
    Quickstart<readme_link>
    api
+   Changelog<changelog_link>
 
 .. toctree::
    :maxdepth: 2
    :caption: Project documentation
 
    CONTRIBUTING
-   Changelog<changelog_link>
    CODE_OF_CONDUCT
 
 Indices and tables
