+++ .github/workflows/test-wheels.yml
@@ -8,12 +8,7 @@ permissions: {}
 
 jobs:
   publish:
-    runs-on: ubuntu-22.04
     permissions:
       contents: read
       packages: read
-    steps:
-      - name: "Test wheels build without uploading"
-        uses: ./.github/actions/wheels
-        with:
-          pypi-repository: "none"
+    uses: ./.github/workflows/wheels.yml
