+++ .github/workflows/upload-pypi.yml
@@ -8,14 +8,12 @@ permissions: {}
 
 jobs:
   publish:
-    runs-on: ubuntu-22.04
     permissions:
       contents: read
       packages: read
-    steps:
-      - name: "Upload to PyPI"
-        uses: ./.github/actions/wheels
-        with:
-          pypi-repository: "pypi"
-          pypi-username: ${{ secrets.PYPI_USERNAME }}
-          pypi-password: ${{ secrets.PYPI_PASSWORD }}
+    uses: ./.github/workflows/wheels.yml
+    with:
+      pypi-repository: "pypi"
+    secrets:
+      pypi-username: ${{ secrets.PYPI_USERNAME }}
+      pypi-password: ${{ secrets.PYPI_PASSWORD }}
