### 2024-08-27 2fd8955
Re-enable VALIDATE_CHECKOV (#62)
