+++ .github/workflows/build.yml
@@ -26,7 +26,7 @@ jobs:
 
     strategy:
       matrix:
-        os: [macOS-13, windows-2022, ubuntu-22.04]
+        os: [macOS-14, windows-2022, ubuntu-22.04]
         python-version: [3.8, 3.9, "3.10", "3.11", "3.12"]
 
     runs-on: ${{ matrix.os }}
