+++ .github/workflows/wheels.yml
@@ -33,9 +33,9 @@ jobs:
             archs: ARM64
           - os: macOS-13
             archs: x86_64
-          - os: macOS-13
+          - os: macOS-14
             archs: arm64
-          - os: macOS-13
+          - os: macOS-14
             archs: universal2
     steps:
       - name: Checkout
@@ -52,14 +52,13 @@ jobs:
         with:
           output-dir: wheelhouse
         env:
-          CIBW_BUILD: "{cp38,cp39,cp310,cp311,cp312}-${{ matrix.build }}*"
+          CIBW_BUILD: "{cp38,cp39,cp310,cp311,cp312,cp313}-${{ matrix.build }}*"
           CIBW_SKIP: cp38-win_arm64
           CIBW_ARCHS: ${{ matrix.archs }}
           CIBW_BUILD_FRONTEND: "build"
           CIBW_TEST_REQUIRES: "pytest"
           CIBW_TEST_COMMAND: "pytest {project}"
-          # Until the day Apple silicon instances are available on GH Actions
-          CIBW_TEST_SKIP: "*-win_arm64 *-macosx_arm64"
+          CIBW_TEST_SKIP: "*-win_arm64 cp38-macosx_*:arm64"
       - uses: actions/upload-artifact@v4.3.6
         with:
           name: Wheel-${{ matrix.os }}-${{ matrix.build }}${{ matrix.archs }}
