### 2024-08-27 256b7cb
Fix typo & update CHANGELOG
