+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This project has adhered to
 
 ### Added
 
-- (planned: Add support for Python 3.13)
+- Add support for Python 3.13.
 - Add Read the Docs documentation (<https://github.com/hajimes/mmh3/issues/54>).
 - (planned: Document benchmark results
   (<https://github.com/hajimes/mmh3/issues/53>)).
