### 2024-08-27 188f012
Fix some lint errors
