### 2024-08-28 f21dc1b
Fix some pylint warnings
