### 2024-08-28 5c8baa6
Rename benchmark class
