+++ benchmark/benchmark.py
@@ -6,10 +6,12 @@ https://github.com/Cyan4973/xxHash/tree/dev/tests/bench
 
 """
 
+from __future__ import annotations
+
 import gc
 import hashlib
 import time
-from typing import Any, Callable, Dict, List
+from typing import Any, Callable, Dict, Final, List
 
 import matplotlib.pyplot as plt
 import mmh3
@@ -28,24 +30,35 @@ class Benchmarker:
 
     # pylint: disable=too-few-public-methods
 
-    def __init__(self, total_microseconds, run_microseconds):
-        if run_microseconds is None:
-            self.run_microseconds = 1
-        else:
-            self.run_microseconds = run_microseconds
+    SIZE_TO_HASH_PER_ROUND: Final[int] = 200000
+    HASH_ROUNDS_MAX: Final[int] = 1000
+
+    MARGIN_FOR_LATENCY: Final[int] = 1024
+
+    WOLKLOAD_MULTIPLIER: Final[int] = 10
+
+    def __init__(self, round_budget_ms: int, total_budget_ms: int):
+        """Initializes the Benchmarker.
+
+        Args:
+            round_budget_ms: The minimum time for each round to spend running
+                the hash function in milliseconds.
+            total_budget_ms: The total minimum time to spend benchmarking in
+                milliseconds.
+        """
 
         # Run the benchmark function
         # until the time spent is greater than the run budget.
         # defaults to 1ms (in nanoseconds)
-        self.run_budget_nanoseconds = self.run_microseconds * TIMELOOP_NANOSEC / 1000
+        self.run_budget_nanoseconds = round_budget_ms * TIMELOOP_NANOSEC / 1000
 
-        self.total_microseconds = total_microseconds
+        self.total_budget_ms = total_budget_ms
 
         self.fastest_nanoseconds_per_run = float("inf")
         self.fastest_run_sum_of_return = -1
         self.number_of_loops = 1
 
-    def _warmup(self, params: Dict[str, Any]) -> None:
+    def _warmup(self, destinations: list[int]) -> None:
         """Warm up the CPU by running the benchmark function.
 
         The current implmentation follows the logic of the following C code:
@@ -64,21 +77,26 @@ class Benchmarker:
             params: The parameters for the benchmark function.
         """
 
-        for i in range(len(params["destinations"])):
-            params["destinations"][i] = 0xE5
+        for i in range(len(destinations)):
+            destinations[i] = 0xE5
 
-    def _benchmark_function(self, params: Dict[str, Any]) -> Dict[str, Any]:
+    def _benchmark_function(
+        self,
+        f: Callable,
+        number_of_blocks: int,
+        source_buffers: list[bytearray],
+        destinations: list[int],
+    ) -> Dict[str, Any]:
         result = {}
 
-        self._warmup(params)
-        target_function = params["function"]
+        self._warmup(destinations)
 
         clock_start = time.time_ns()
 
         for _ in range(self.number_of_loops):
-            for i in range(params["number_of_blocks"]):
-                b = params["source_buffers"][i]
-                params["destinations"][i] = target_function(b)
+            for i in range(number_of_blocks):
+                b = source_buffers[i]
+                destinations[i] = f(b)
 
         clock_end = time.time_ns()
         time_spent = clock_end - clock_start
@@ -88,7 +106,7 @@ class Benchmarker:
 
         return result
 
-    def run_timed_benchmarks(self, params: Dict[str, Any]) -> Dict[str, Any]:
[…]
