+++ pyproject.toml
@@ -31,7 +31,7 @@ classifiers = [
 dev = [
   "mypy == 1.11.2",
   "pytest == 8.3.2",
-  "matplotlib == 3.9.1",
+  "matplotlib == 3.9.2",
   "pandas == 2.2.2",
   "pymmh3 == 0.0.5",
   "xxhash == 3.4.1"
