### 2024-08-29 3641adf
Upgrade pip in the test workflow
