+++ .github/workflows/build.yml
@@ -38,6 +38,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
+          python -m pip install --upgrade pip
           pip install .
           pip install ".[dev]"
       - name: Test with pytest
