+++ pyproject.toml
@@ -9,6 +9,7 @@ description = "Python extension for MurmurHash (MurmurHash3), a set of fast and 
 readme = "README.md"
 license = {file = "LICENSE"}
 keywords = ["utility", "hash", "MurmurHash"]
+requires-python = ">=3.8"
 authors = [
   {name = "Hajime Senuma", email="hajime.senuma@gmail.com"}
 ]
@@ -31,10 +32,12 @@ classifiers = [
 dev = [
   "mypy == 1.11.2",
   "pytest == 8.3.2",
-  "matplotlib == 3.9.2",
   "pandas == 2.2.2",
   "pymmh3 == 0.0.5",
   "xxhash == 3.4.1"
+]
+dev-plot = [
+  "matplotlib >= 3.7",
 ]
 
 [project.urls]
