### 2024-08-29 dfa6eb3
Fix type hints
