### 2024-08-29 0f3b70b
Omit redundant line in CHANGELOG
