### 2024-08-30 7f9b71f
Move constants
