### 2024-08-30 875023d
Refactor benchmark
