+++ .github/workflows/build.yml
@@ -40,7 +40,7 @@ jobs:
         run: |
           pip install --upgrade pip
           pip install .
-          pip install ".[dev]"
+          pip install ".[test]"
       - name: Test with pytest
         run: python -m pytest
       - name: Test type hints with mypy
