+++ pyproject.toml
@@ -29,15 +29,15 @@ classifiers = [
 ]
 
 [project.optional-dependencies]
-dev = [
+test = [
   "mypy == 1.11.2",
-  "pytest == 8.3.2",
+  "pytest == 8.3.2"
+]
+dev = [
   "pandas >= 2.0.0",
   "pymmh3 == 0.0.5",
-  "xxhash == 3.4.1"
-]
-dev-plot = [
-  "matplotlib >= 3.7",
+  "xxhash == 3.4.1",
+  "matplotlib >= 3.7"
 ]
 
 [project.urls]
