+++ .github/workflows/build.yml
@@ -21,8 +21,6 @@ jobs:
     permissions:
       contents: read
       packages: read
-      # To report GitHub Actions status checks
-      statuses: write
 
     strategy:
       matrix:
@@ -38,6 +36,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
+          pip install --upgrade pip
           pip install .
           pip install ".[test]"
       - name: Test with pytest
