+++ pyproject.toml
@@ -29,8 +29,14 @@ classifiers = [
 
 [project.optional-dependencies]
 test = [
-  "mypy >= 1.0",
-  "pytest >= 7.0.0"
+  "mypy >= 1.11.1",
+  "pytest == 8.3.2"
+]
+dev = [
+  "pandas >= 2.0.0",
+  "pymmh3 == 0.0.5",
+  "xxhash == 3.4.1",
+  "matplotlib >= 3.7"
 ]
 
 [project.urls]
