+++ docs/requirements.txt
@@ -34,8 +34,8 @@ mdit-py-plugins==0.4.1
     # via myst-parser
 mdurl==0.1.2
     # via markdown-it-py
-myst-parser==3.0.1
-    # via -r requirements.in
+myst-parser==4.0.0
+    # via -r docs/requirements.in
 packaging==24.1
     # via sphinx
 pygments==2.18.0
@@ -45,12 +45,12 @@ pyyaml==6.0.1
 requests==2.32.3
     # via sphinx
 shibuya==2024.7.13
-    # via -r requirements.in
+    # via -r docs/requirements.in
 snowballstemmer==2.2.0
     # via sphinx
 sphinx==7.4.5
     # via
-    #   -r requirements.in
+    #   -r docs/requirements.in
     #   myst-parser
     #   shibuya
 sphinxcontrib-applehelp==1.0.8
