### 2024-08-30 0ab03d3
Bump shibuya from 2024.7.13 to 2024.8.27
Bumps [shibuya](https://github.com/lepture/shibuya) from 2024.7.13 to 2024.8.27.
- [Release notes](https://github.com/lepture/shibuya/releases)
- [Changelog](https://github.com/lepture/shibuya/blob/master/docs/changelog.rst)
- [Commits](https://github.com/lepture/shibuya/compare/2024.7.13...2024.8.27)

---
updated-dependencies:
- dependency-name: shibuya
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
