### 2024-08-30 64dda40
Add Read the Docs dependencies to pyproject.toml
