+++ docs/requirements.txt
@@ -1,69 +1,0 @@
-#
-# This file is autogenerated by pip-compile with Python 3.12
-# by the following command:
-#
-#    pip-compile requirements.in
-#
-alabaster==0.7.16
-    # via sphinx
-babel==2.15.0
-    # via sphinx
-certifi==2024.7.4
-    # via requests
-charset-normalizer==3.3.2
-    # via requests
-docutils==0.21.2
-    # via
-    #   myst-parser
-    #   sphinx
-idna==3.7
-    # via requests
-imagesize==1.4.1
-    # via sphinx
-jinja2==3.1.4
-    # via
-    #   myst-parser
-    #   sphinx
-markdown-it-py==3.0.0
-    # via
-    #   mdit-py-plugins
-    #   myst-parser
-markupsafe==2.1.5
-    # via jinja2
-mdit-py-plugins==0.4.1
-    # via myst-parser
-mdurl==0.1.2
-    # via markdown-it-py
-myst-parser==3.0.1
-    # via -r requirements.in
-packaging==24.1
-    # via sphinx
-pygments==2.18.0
-    # via sphinx
-pyyaml==6.0.1
-    # via myst-parser
-requests==2.32.3
-    # via sphinx
-shibuya==2024.7.13
-    # via -r requirements.in
-snowballstemmer==2.2.0
-    # via sphinx
-sphinx==7.4.5
-    # via
-    #   -r requirements.in
-    #   myst-parser
-    #   shibuya
-sphinxcontrib-applehelp==1.0.8
-    # via sphinx
-sphinxcontrib-devhelp==1.0.6
-    # via sphinx
-sphinxcontrib-htmlhelp==2.0.5
-    # via sphinx
-sphinxcontrib-jsmath==1.0.1
-    # via sphinx
-sphinxcontrib-qthelp==1.0.7
-    # via sphinx
-sphinxcontrib-serializinghtml==1.1.10
-    # via sphinx
-urllib3==2.2.2
-    # via requests
