+++ docs/requirements.in
@@ -1,3 +1,0 @@
-myst-parser
-shibuya
-sphinx
