+++ pyproject.toml
@@ -41,7 +41,7 @@ dev = [
 docs = [
   "myst-parser == 4.0.0",
   "shibuya == 2024.8.27",
-  "sphinx == 7.4.7"
+  "sphinx == 8.0.2"
 ]
 
 [project.urls]
