### 2024-08-31 a34c652
Fix CHANGELOG url in pyproject.toml
