+++ .github/workflows/benchmark.yml
@@ -1,0 +1,29 @@
+---
+name: Benchmarking hash functions
+
+on:
+  workflow_dispatch:
+
+permissions: {}
+
+jobs:
+  benchmark:
+    permissions:
+      contents: read
+      packages: read
+    runs-on: ubuntu-22.04
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
+      - name: Install dependencies
+        run: |
+          pip install --upgrade pip
+          pip install .
+          pip install ".[dev]"
+      - name: Benchmark hash functions
+        run: |
+          python -m pyperf system tune
+          python becnhmark/bechmark.py
