+++ pyproject.toml
@@ -37,6 +37,7 @@ dev = [
   "matplotlib == 3.7.3",
   "pandas == 2.0.3",
   "pymmh3 == 0.0.5",
+  "pyperf == 2.7.0",
   "xxhash == 3.5.0"
 ]
 docs = [
