### 2024-09-02 9d40e4a
Add expedient on cond to bench.yml
