### 2024-09-02 b967258
Fix python version in benchmark workflow
