+++ .github/workflows/benchmark.yml
@@ -20,7 +20,8 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.12"
+          # pandas 2.0.3 does not have wheels for Python 3.12
+          python-version: "3.11"
       - name: Install dependencies
         run: |
           pip install --upgrade pip
