### 2024-09-02 4512aa6
Add sudo install for pyperf
