### 2024-09-02 c6a9bbc
Add pipe to dev/null
