+++ .github/workflows/benchmark.yml
@@ -1,5 +1,5 @@
 ---
-name: Benchmarking hash functions
+name: Benchmark
 
 on:
   push:
@@ -34,6 +34,6 @@ jobs:
           lscpu -a -e
           lsmod | grep msr
           sudo python -m pip install pyperf
-          sudo python -m pyperf system tune
+          sudo python -m pyperf system tune >/dev/null
       - name: Benchmark hash functions
         run: python becnhmark/bechmark.py
