+++ .github/workflows/benchmark.yml
@@ -14,7 +14,7 @@ jobs:
     permissions:
       contents: read
       packages: read
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-20.04
     steps:
       - uses: actions/checkout@v4
       - name: Set up Python
