### 2024-09-02 5950521
Change gha-permissions
