+++ .github/workflows/benchmark.yml
@@ -7,14 +7,9 @@ on:
       - feature/introduce-pyperf
   workflow_dispatch:
 
-permissions: {}
-
 jobs:
   benchmark:
-    permissions:
-      contents: read
-      packages: read
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-22.04
     steps:
       - uses: actions/checkout@v4
       - name: Set up Python
