+++ .github/workflows/wheels.yml
@@ -59,7 +59,7 @@ jobs:
           CIBW_TEST_REQUIRES: "pytest"
           CIBW_TEST_COMMAND: "pytest {project}"
           CIBW_TEST_SKIP: "*-win_arm64 cp38-macosx_*:arm64"
-      - uses: actions/upload-artifact@v4.3.6
+      - uses: actions/upload-artifact@v4.4.0
         with:
           name: Wheel-${{ matrix.os }}-${{ matrix.build }}${{ matrix.archs }}
           path: ./wheelhouse/*.whl
@@ -78,7 +78,7 @@ jobs:
           python -m pip install --upgrade pip
           pip install setuptools
           python setup.py build sdist
-      - uses: actions/upload-artifact@v4.3.6
+      - uses: actions/upload-artifact@v4.4.0
         with:
           path: dist/*.tar.gz
   publish:
