### 2024-09-02 0c15491
Bump shibuya from 2024.8.27 to 2024.8.30 (#71)
Bumps [shibuya](https://github.com/lepture/shibuya) from 2024.8.27 to 2024.8.30.
- [Release notes](https://github.com/lepture/shibuya/releases)
- [Changelog](https://github.com/lepture/shibuya/blob/master/docs/changelog.rst)
- [Commits](https://github.com/lepture/shibuya/compare/2024.8.27...2024.8.30)

---
updated-dependencies:
- dependency-name: shibuya
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@u
[…]
