### 2024-09-02 a2916ca
Fix bench function name
