### 2024-09-03 06a59ce
Improve benchmarking logic
