### 2024-09-03 7f893fb
Remove wrong lines
