### 2024-09-03 2466c40
Add hash functions to benchmark
