### 2024-09-03 100d7ec
Remove push condition from benchmark workflow
