### 2024-09-03 cfcc40e
Bump pandas from 2.0.3 to 2.2.2
Bumps [pandas](https://github.com/pandas-dev/pandas) from 2.0.3 to 2.2.2.
- [Release notes](https://github.com/pandas-dev/pandas/releases)
- [Commits](https://github.com/pandas-dev/pandas/compare/v2.0.3...v2.2.2)

---
updated-dependencies:
- dependency-name: pandas
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
