+++ pyproject.toml
@@ -35,7 +35,7 @@ test = [
 ]
 dev = [
   "matplotlib == 3.7.3",
-  "pandas == 2.0.3",
+  "pandas == 2.2.2",
   "pymmh3 == 0.0.5",
   "pyperf == 2.7.0",
   "xxhash == 3.5.0"
