### 2024-09-03 abb4f64
Bump matplotlib from 3.7.3 to 3.9.2
Bumps [matplotlib](https://github.com/matplotlib/matplotlib) from 3.7.3 to 3.9.2.
- [Release notes](https://github.com/matplotlib/matplotlib/releases)
- [Commits](https://github.com/matplotlib/matplotlib/compare/v3.7.3...v3.9.2)

---
updated-dependencies:
- dependency-name: matplotlib
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
