+++ .github/workflows/draft-pdf.yml
@@ -1,8 +1,22 @@
-on: [push]
+---
+name: Draft Paper
+
+on:
+  push:
+    branches:
+      - master
+      - paper
+
+permissions: {}
 
 jobs:
   paper:
+    permissions:
+      contents: read
+      packages: read
+
     runs-on: ubuntu-latest
+
     name: Paper Draft
     steps:
       - name: Checkout
