### 2024-09-03 ddbd80b
Update paper, bib, and workflow
