+++ paper/paper.bib
@@ -75,17 +75,17 @@
 	isbn         = {978-1-4920-5502-0},
 	edition      = {2nd edition}
 }
-@software{hugo_van_kemenade_2023_7994944,
+@software{hugo_van_kemenade_2024_13624792,
   author       = {Van Kemenade, Hugo and
                   Si, Richard and
                   Dollenstein, Zsolt},
-  title        = {hugovk/top-pypi-packages: Release 2023.06},
-  month        = jun,
-  year         = 2023,
+  title        = {hugovk/top-pypi-packages: Release 2024.09},
+  month        = sep,
+  year         = 2024,
   publisher    = {Zenodo},
-  version      = {2023.06},
-  doi          = {10.5281/zenodo.7994944},
-  url          = {https://doi.org/10.5281/zenodo.7994944}
+  version      = {2024.09},
+  doi          = {10.5281/zenodo.13624792},
+  url          = {https://doi.org/10.5281/zenodo.13624792}
 }
 @inproceedings{kakwani_indicnlpsuite_2020,
 	title        = {{IndicNLPSuite}: {Monolingual} {Corpora}, {Evaluation} {Benchmarks} and {Pre}-trained {Multilingual} {Language} {Models} for {Indian} {Languages}},
