+++ .github/workflows/draft-pdf.yml
@@ -20,7 +20,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4.1.7
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
@@ -28,7 +28,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: paper/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4.4.0
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
