### 2024-09-03 974d9e1
Omit pymmh3 from benchmark workflow
