+++ .github/workflows/benchmark.yml
@@ -14,17 +14,7 @@ jobs:
 
     strategy:
       matrix:
-        hash:
-          [
-            mmh3_32,
-            mmh3_128,
-            xxh_32,
-            xxh_64,
-            xxh3_64,
-            xxh3_128,
-            md5,
-            sha1,
-          ]
+        hash: [mmh3_32, mmh3_128, xxh_32, xxh_64, xxh3_64, xxh3_128, md5, sha1]
 
     runs-on: ubuntu-22.04
     steps:
