### 2024-09-03 3831f2c
Revise benchmark plotting
