+++ pyproject.toml
@@ -33,17 +33,19 @@ test = [
   "mypy == 1.11.2",
   "pytest == 8.3.2"
 ]
-dev = [
-  "matplotlib == 3.7.3",
-  "pandas == 2.0.3",
-  "pymmh3 == 0.0.5",
-  "pyperf == 2.7.0",
-  "xxhash == 3.5.0"
-]
 docs = [
   "myst-parser == 4.0.0",
   "shibuya == 2024.8.30",
   "sphinx == 8.0.2"
+]
+benchmark = [
+  "pymmh3 == 0.0.5",
+  "pyperf == 2.7.0",
+  "xxhash == 3.5.0"
+]
+plot = [
+  "matplotlib == 3.9.2",
+  "pandas == 2.2.2"
 ]
 
 [project.urls]
