### 2024-09-04 7baa944
Enhance benchmark workflow
