+++ .github/workflows/benchmark.yml
@@ -14,19 +14,7 @@ jobs:
 
     strategy:
       matrix:
-        hash:
-          [
-            mmh3_32,
-            mmh3_128,
-            xxh_32,
-            xxh_64,
-            xxh3_64,
-            xxh3_128,
-            pymmh3_32,
-            pymmh3_128,
-            md5,
-            sha1,
-          ]
+        hash: [mmh3_32, mmh3_128, xxh_32, xxh_64, xxh3_64, xxh3_128, md5, sha1]
 
     runs-on: ubuntu-22.04
     steps:
@@ -35,13 +23,12 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          # pandas 2.0.3 does not provide wheels for Python 3.12
-          python-version: "3.11"
+          python-version: "3.12"
       - name: Install dependencies
         run: |
           pip install --upgrade pip
           pip install .
-          pip install ".[dev]"
+          pip install ".[benchmark]"
       - name: Benchmark hash functions
         run: |
           mkdir var
