+++ benchmark/plot_graph.py
@@ -21,9 +21,10 @@ DIGEST_SIZES = {
     "sha1": hashlib.sha1().digest_size,
 }
 
+BANDWIDTH_FILE_NAME = "bandwidth.png"
+BANDWIDTH_SMALL_FILE_NAME = "bandwidth_small.png"
 LATENCY_FILE_NAME = "latency.png"
 LATENCY_SMALL_FILE_NAME = "latency_small.png"
-OUTPUT_BANDWIDTH_FILE_NAME = "output_bandwidth.png"
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
@@ -31,8 +32,8 @@ if __name__ == "__main__":
     parser.add_argument("filenames", nargs="+")
     args = parser.parse_args()
 
-    data_result: dict[str, list[float]] = {}
-    data_digest_size_result: dict[str, list[float]] = {}
+    result_latency: dict[str, list[float]] = {}
+    result_bandwidth: dict[str, list[float]] = {}
     index: list[int] = []
 
     for file_name in args.filenames:
@@ -40,47 +41,58 @@ if __name__ == "__main__":
         base_name = os.path.basename(file_name)
         hash_name = os.path.splitext(base_name)[0]
 
-        data_result[hash_name] = []
-        data_digest_size_result[hash_name] = []
+        result_bandwidth[hash_name] = []
+        result_latency[hash_name] = []
         index = []
 
         for bench_name in suite.get_benchmark_names():
             bench = suite.get_benchmark(bench_name)
             data_size = int(bench_name.split(" ")[0])
             index.append(data_size)
-            latency_ns = bench.median() * 1000 * 1000 * 1000
-            data_result[hash_name].append(latency_ns)
-            data_digest_size_result[hash_name].append(
-                (DIGEST_SIZES[hash_name] / bench.median()) / (1024 * 1024)
-            )
+            latency_seconds = bench.median()
 
-    df_latency = pd.DataFrame(data_result, index=index)
-    df_latency = df_latency[DIGEST_SIZES.keys()]
-    df_bandwidth = pd.DataFrame(
-        data_digest_size_result, index=df_latency.index / (1024 * 1024)
-    )
+            result_bandwidth[hash_name].append(
+                DIGEST_SIZES[hash_name] / latency_seconds
+            )
+            result_latency[hash_name].append(latency_seconds)
+
+    df_bandwidth = pd.DataFrame(result_bandwidth, index=index)
     df_bandwidth = df_bandwidth[DIGEST_SIZES.keys()]
+    df_latency = pd.DataFrame(result_latency, index=index)
+    df_latency = df_latency[DIGEST_SIZES.keys()]
 
     plt.rcParams["figure.dpi"] = 72 * 3
 
     plt.figure()
 
-    df_large_data = df_latency.copy()
-    df_large_data = df_large_data / 1000 / 1000
-    df_large_data.index = df_latency.index / (1024 * 1024)
-    df_large_data.plot(xlabel="Input size (MB)", ylabel="Latency (ms)")
+    df_bandwidth_all = df_bandwidth / 1024
+    df_bandwidth_all.index = df_bandwidth_all.index / (1024 * 1024)
+    df_bandwidth_all.plot(
+        xlabel="Input size (MB)", ylabel="Output bandwidth (kB/s)", logy=True
+    )
+    plt.savefig(os.path.join(args.output_dir, BANDWIDTH_FILE_NAME))
+
+    df_bandwidth_small = df_bandwidth.copy()
+    df_bandwidth_small = df_bandwidth_small.drop(columns=["md5", "sha1"])
[…]
