### 2024-09-04 #6124 @hajimes
Progress report: I have completed the benchmarking part. After consideration, I chose to use [pyperf](https://github.com/psf/pyperf) as the microbenchmarking backend (rather than reinventing the wheel) and obtained results in [this GitHub Actions workflow](https://github.com/hajimes/mmh3/actions/runs/10686405688).

Next, I will try to prepare documentation for the benchmark results, revise the paper, and then write my response for the reviewer as soon as possible.  I sincerely appreciate your understanding and patience.
