+++ benchmark/benchmark.py
@@ -3,7 +3,6 @@
 import hashlib
 import itertools
 import time
-from collections import OrderedDict
 from collections.abc import Callable
 from typing import Final
 
@@ -94,17 +93,18 @@ def add_cmdline_args(cmd: list, args) -> None:
     cmd.extend(("--test-buffer-size-max", str(args.test_buffer_size_max)))
 
 
-HASHES = OrderedDict()
-HASHES["mmh3_32"] = mmh3.hash
-HASHES["mmh3_128"] = mmh3.hash_bytes
-HASHES["xxh_32"] = xxhash.xxh32_digest
-HASHES["xxh_64"] = xxhash.xxh64_digest
-HASHES["xxh3_64"] = xxhash.xxh3_64_digest
-HASHES["xxh3_128"] = xxhash.xxh3_128_digest
-HASHES["pymmh3_32"] = pymmh3.hash
-HASHES["pymmh3_128"] = pymmh3.hash128
-HASHES["md5"] = lambda ba: hashlib.md5(ba).digest()
-HASHES["sha1"] = lambda ba: hashlib.sha1(ba).digest()
+HASHES = {
+    "mmh3_32": mmh3.hash,
+    "mmh3_128": mmh3.hash_bytes,
+    "xxh_32": xxhash.xxh32_digest,
+    "xxh_64": xxhash.xxh64_digest,
+    "xxh3_64": xxhash.xxh3_64_digest,
+    "xxh3_128": xxhash.xxh3_128_digest,
+    "pymmh3_32": pymmh3.hash,
+    "pymmh3_128": pymmh3.hash128,
+    "md5": lambda ba: hashlib.md5(ba).digest(),
+    "sha1": lambda ba: hashlib.sha1(ba).digest(),
+}
 
 if __name__ == "__main__":
     runner = pyperf.Runner(add_cmdline_args=add_cmdline_args)
