### 2024-09-04 bac8485
Add experimental figures to paper
