+++ paper/paper.md
@@ -130,14 +130,13 @@ Yan Collet [@collet_xxhash_2012].
 
 # Benchmarks
 
-![Output bandwidth]{label="bandwidth"}(../docs/_static/bandwidth.png)
+![Output bandwidth {label="bandwidth"}](../docs/_static/bandwidth.png)
 
 Figure \ref{bandwidth} shows the output bandwidth of...
 
 Figure \autoref{bandwidth} shows the output bandwidth of...
 
 ![Output bandwidth for small inputs](../docs/_static/bandwidth_small.png)
-
 
 # Acknowledgements
 
