### 2024-09-05 9660b9e
Enhance hash_bytes
