### 2024-09-05 ba23958
Specify only major versions for basic gh-actions
