+++ docs/CONTRIBUTING.md
@@ -1,25 +1,26 @@
 # Contributing
 
-Thank you for your interest in contributing to the `mmh3` project!
+Thank you for your interest in contributing to the `mmh3` project. We
+appreciate your support and look forward to your contributions.
 
-Read [README](https://github.com/hajimes/mmh3/blob/master/README.md) to get an
+Please read [README](https://github.com/hajimes/mmh3/blob/master/README.md) to get an
 overview of the `mmh3` project, and follow our
 [Code of Conduct](./CODE_OF_CONDUCT) (ACM Code of Ethics and Professional
 Conduct).
 
-## Issues
+## Submitting issues
 
-You can contribute to our project by simply submitting a bug report or a feature
-suggestion through the [issue tracker](https://github.com/hajimes/mmh3/issues).
+We welcome your contributions, whether it's submitting a bug report or
+suggesting a new feature through the
+[issue tracker](https://github.com/hajimes/mmh3/issues).
 
-Before submitting a new issue, it's a good idea to check
-[Known Issues section on README](https://github.com/hajimes/mmh3#known-issues).
+Before creating a new issue, please check the
+[Known Issues section in README](https://github.com/hajimes/mmh3#known-issues)
+to see if the problem has already been noted.
 
-## Maintaining and developing the project
+## Project structure
 
-### Project structure
-
-As of 4.1.0, the layout of the project is as follows:
+As of version 4.2.0, the project layout is structured as follows:
 
 - `src/mmh3`
   - `mmh3module.c`: the main file that serves as the interface between Python
@@ -36,14 +37,26 @@ As of 4.1.0, the layout of the project is as follows:
   - `refresh.py`: file that generates `src/mmh3/murmurhash.c` and
     `src/mmh3/murmurhash.h` from the original MurmurHash3 C++ code. Edit this
     file to modify the contents of these files.
+- `benchmark`
+  - `benchmark.py`: script to run benchmarks.
+  - `plot_graph.py`: script to plot benchmark results.
+- `docs`: project documentation directory
+- `.github/workflows`: GitHub Actions workflows
 
-### Testing
+## Testing
 
 Before submitting your changes, make sure to run the project's tests to ensure
-that everything is working as expected. At least you should run `pytest` and
-`mypy --strict tests` from the project root directory.
+that everything is working as expected.
 
-#### (Optional) Testing on s390x
+Try
+
+```bash
+$ pip install ".[test]"
+$ pytest
+$ mypy --strict tests
+```
+
+### (Optional) Testing on s390x
 
 When you have modified the code in a way which may cause endian issues, you may
 want to locally test on s390x, the only big-endian platform officially supported
@@ -53,7 +66,7 @@ by Python.
 by Simon Willison is a good introduction to Docker/QEMU settings for emulating
 s390x.
 
-### Pull request
+## Pull request
 
 Once you've pushed your changes to your fork, you can
 [create a pull request (PR)](https://github.com/hajimes/mmh3/pulls) on the main
@@ -73,23 +86,81 @@ The idea of the subproject directory loosely follows the
 ### Updating mmh3 C code
 
 Try `git submodule update --init` to fetch Appleby's original SMHasher project
-as a git submodule. Then, run the `refresh.py` script to generate PEP
+as a git submodule. Then, run `python util/refresh.py` to generate PEP
 7-compliant C code from the original project, instead of editing `murmurhash3.*`
 files manually.
 
-Add transformation code to the `refresh.py` script to perform
[…]
