### 2024-09-06 bc49995
Update plotting logic
