### 2024-09-06 4e6535b
Fix type hints of plot_graph
