### 2024-09-06 2d292f8
Fix type hints
