+++ benchmark/plot_graph.py
@@ -3,12 +3,47 @@
 import argparse
 import hashlib
 import os
+from typing import TypeVar
 
 import matplotlib.pyplot as plt
 import mmh3
 import pandas as pd
 import pyperf
 import xxhash
+
+T = TypeVar("T")
+
+
+def pad_with_nan(data: dict[T, list[float]]) -> dict[T, list[float]]:
+    """Pad the data with NaN values to make the length of all lists equal.
+
+    Args:
+        data: The data to pad.
+
+    Returns:
+        The padded data.
+    """
+
+    max_len = max(len(v) for v in data.values())
+    for k, v in data.items():
+        data[k] = v + [float("nan")] * (max_len - len(v))
+
+    return data
+
+
+def ordered_intersection(list1: list[T], list2: list[T]) -> list[T]:
+    """Return the intersection of two lists in the order of the first list.
+
+    Args:
+        list1: The first list.
+        list2: The second list.
+
+    Returns:
+        The intersection of the two lists in the order of the first list.
+    """
+
+    return [item for item in list1 if item in list2]
+
 
 DIGEST_SIZES = {
     "mmh3_32": mmh3.mmh3_32().digest_size,
@@ -19,11 +54,14 @@ DIGEST_SIZES = {
     "xxh3_128": xxhash.xxh3_128().digest_size,
     "md5": hashlib.md5().digest_size,
     "sha1": hashlib.sha1().digest_size,
+    "pymmh3_32": mmh3.mmh3_32().digest_size,
+    "pymmh3_128": mmh3.mmh3_x64_128().digest_size,
 }
 
+BANDWIDTH_FILE_NAME = "bandwidth.png"
+BANDWIDTH_SMALL_FILE_NAME = "bandwidth_small.png"
 LATENCY_FILE_NAME = "latency.png"
 LATENCY_SMALL_FILE_NAME = "latency_small.png"
-OUTPUT_BANDWIDTH_FILE_NAME = "output_bandwidth.png"
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
@@ -31,8 +69,8 @@ if __name__ == "__main__":
     parser.add_argument("filenames", nargs="+")
     args = parser.parse_args()
 
-    data_result: dict[str, list[float]] = {}
-    data_digest_size_result: dict[str, list[float]] = {}
+    result_latency: dict[str, list[float]] = {}
+    result_bandwidth: dict[str, list[float]] = {}
     index: list[int] = []
 
     for file_name in args.filenames:
@@ -40,47 +78,64 @@ if __name__ == "__main__":
         base_name = os.path.basename(file_name)
         hash_name = os.path.splitext(base_name)[0]
 
-        data_result[hash_name] = []
-        data_digest_size_result[hash_name] = []
+        result_bandwidth[hash_name] = []
+        result_latency[hash_name] = []
         index = []
 
         for bench_name in suite.get_benchmark_names():
             bench = suite.get_benchmark(bench_name)
             data_size = int(bench_name.split(" ")[0])
             index.append(data_size)
-            latency_ns = bench.median() * 1000 * 1000 * 1000
-            data_result[hash_name].append(latency_ns)
-            data_digest_size_result[hash_name].append(
-                (DIGEST_SIZES[hash_name] / bench.median()) / (1024 * 1024)
-            )
+            latency_seconds = bench.median()
 
-    df_latency = pd.DataFrame(data_result, index=index)
-    df_latency = df_latency[DIGEST_SIZES.keys()]
-    df_bandwidth = pd.DataFrame(
[…]
