### 2024-09-06 d81c67c
Use function inlining
