+++ setup.py
@@ -3,7 +3,7 @@
 from setuptools import Extension, setup
 
 mmh3module = Extension(
-    "mmh3", sources=["./src/mmh3/mmh3module.c", "./src/mmh3/murmurhash3.c"]
+    "mmh3", sources=["./src/mmh3/mmh3module.c"]
 )
 
 setup(
