+++ src/mmh3/murmurhash3.h
@@ -309,17 +309,343 @@ digest_x86_128_impl(uint32_t h1, uint32_t h2, uint32_t h3, uint32_t h4,
 #endif
 }
 
+static FORCE_INLINE void
+murmurhash3_x86_32(const void *key, Py_ssize_t len, uint32_t seed, void *out)
+{
+    const uint8_t *data = (const uint8_t *)key;
+    const Py_ssize_t nblocks = len / 4;
+
+    uint32_t h1 = seed;
+
+    const uint32_t c1 = 0xcc9e2d51;
+    const uint32_t c2 = 0x1b873593;
+
+    //----------
+    // body
+
+    const uint32_t *blocks = (const uint32_t *)(data + nblocks * 4);
+
+    for (Py_ssize_t i = -nblocks; i; i++) {
+        uint32_t k1 = getblock32(blocks, i);
+
+        k1 *= c1;
+        k1 = ROTL32(k1, 15);
+        k1 *= c2;
+
+        h1 ^= k1;
+        h1 = ROTL32(h1, 13);
+        h1 = h1 * 5 + 0xe6546b64;
+    }
+
+    //----------
+    // tail
+
+    const uint8_t *tail = (const uint8_t *)(data + nblocks * 4);
+
+    uint32_t k1 = 0;
+
+    switch (len & 3) {
+        case 3:
+            k1 ^= tail[2] << 16;
+        case 2:
+            k1 ^= tail[1] << 8;
+        case 1:
+            k1 ^= tail[0];
+            k1 *= c1;
+            k1 = ROTL32(k1, 15);
+            k1 *= c2;
+            h1 ^= k1;
+    };
+
+    //----------
+    // finalization
+
+    h1 ^= len;
+
+    h1 = fmix32(h1);
+
+    *(uint32_t *)out = h1;
+}
+
 //-----------------------------------------------------------------------------
 
-void
-murmurhash3_x86_32(const void *key, Py_ssize_t len, uint32_t seed, void *out);
+static FORCE_INLINE void
+murmurhash3_x86_128(const void *key, const Py_ssize_t len, uint32_t seed,
+                    void *out)
+{
+    const uint8_t *data = (const uint8_t *)key;
+    const Py_ssize_t nblocks = len / 16;
 
-void
-murmurhash3_x86_128(const void *key, Py_ssize_t len, uint32_t seed, void *out);
+    uint32_t h1 = seed;
+    uint32_t h2 = seed;
+    uint32_t h3 = seed;
+    uint32_t h4 = seed;
 
-void
-murmurhash3_x64_128(const void *key, Py_ssize_t len, uint32_t seed, void *out);
+    const uint32_t c1 = 0x239b961b;
+    const uint32_t c2 = 0xab0e9789;
+    const uint32_t c3 = 0x38b34ae5;
+    const uint32_t c4 = 0xa1e38b93;
+
+    //----------
+    // body
+
+    const uint32_t *blocks = (const uint32_t *)(data + nblocks * 16);
+
+    for (Py_ssize_t i = -nblocks; i; i++) {
+        uint32_t k1 = getblock32(blocks, i * 4 + 0);
+        uint32_t k2 = getblock32(blocks, i * 4 + 1);
+        uint32_t k3 = getblock32(blocks, i * 4 + 2);
+        uint32_t k4 = getblock32(blocks, i * 4 + 3);
+
+        k1 *= c1;
+        k1 = ROTL32(k1, 15);
+        k1 *= c2;
+        h1 ^= k1;
+
+        h1 = ROTL32(h1, 19);
[…]
