+++ setup.py
@@ -2,9 +2,7 @@
 
 from setuptools import Extension, setup
 
-mmh3module = Extension(
-    "mmh3", sources=["./src/mmh3/mmh3module.c"]
-)
+mmh3module = Extension("mmh3", sources=["./src/mmh3/mmh3module.c"])
 
 setup(
     ext_modules=[mmh3module],
