### 2024-09-07 8f4b693
Function inlining only for mmh3_32
