+++ src/mmh3/murmurhash3.h
@@ -309,8 +309,10 @@ digest_x86_128_impl(uint32_t h1, uint32_t h2, uint32_t h3, uint32_t h4,
 #endif
 }
 
+//-----------------------------------------------------------------------------
+
 static FORCE_INLINE void
-murmurhash3_x86_32(const void *key, Py_ssize_t len, uint32_t seed, void *out)
+murmurhash3_x86_32(const void *key, const Py_ssize_t len, const uint32_t seed, const void *out)
 {
     const uint8_t *data = (const uint8_t *)key;
     const Py_ssize_t nblocks = len / 4;
@@ -369,283 +371,12 @@ murmurhash3_x86_32(const void *key, Py_ssize_t len, uint32_t seed, void *out)
 
 //-----------------------------------------------------------------------------
 
-static FORCE_INLINE void
-murmurhash3_x86_128(const void *key, const Py_ssize_t len, uint32_t seed,
-                    void *out)
-{
-    const uint8_t *data = (const uint8_t *)key;
-    const Py_ssize_t nblocks = len / 16;
+void
+murmurhash3_x86_128(const void *key, const Py_ssize_t len, const uint32_t seed, const void *out);
 
-    uint32_t h1 = seed;
-    uint32_t h2 = seed;
-    uint32_t h3 = seed;
-    uint32_t h4 = seed;
-
-    const uint32_t c1 = 0x239b961b;
-    const uint32_t c2 = 0xab0e9789;
-    const uint32_t c3 = 0x38b34ae5;
-    const uint32_t c4 = 0xa1e38b93;
-
-    //----------
-    // body
-
-    const uint32_t *blocks = (const uint32_t *)(data + nblocks * 16);
-
-    for (Py_ssize_t i = -nblocks; i; i++) {
-        uint32_t k1 = getblock32(blocks, i * 4 + 0);
-        uint32_t k2 = getblock32(blocks, i * 4 + 1);
-        uint32_t k3 = getblock32(blocks, i * 4 + 2);
-        uint32_t k4 = getblock32(blocks, i * 4 + 3);
-
-        k1 *= c1;
-        k1 = ROTL32(k1, 15);
-        k1 *= c2;
-        h1 ^= k1;
-
-        h1 = ROTL32(h1, 19);
-        h1 += h2;
-        h1 = h1 * 5 + 0x561ccd1b;
-
-        k2 *= c2;
-        k2 = ROTL32(k2, 16);
-        k2 *= c3;
-        h2 ^= k2;
-
-        h2 = ROTL32(h2, 17);
-        h2 += h3;
-        h2 = h2 * 5 + 0x0bcaa747;
-
-        k3 *= c3;
-        k3 = ROTL32(k3, 17);
-        k3 *= c4;
-        h3 ^= k3;
-
-        h3 = ROTL32(h3, 15);
-        h3 += h4;
-        h3 = h3 * 5 + 0x96cd1c35;
-
-        k4 *= c4;
-        k4 = ROTL32(k4, 18);
-        k4 *= c1;
-        h4 ^= k4;
-
-        h4 = ROTL32(h4, 13);
-        h4 += h1;
-        h4 = h4 * 5 + 0x32ac3b17;
-    }
-
-    //----------
-    // tail
-
-    const uint8_t *tail = (const uint8_t *)(data + nblocks * 16);
-
-    uint32_t k1 = 0;
-    uint32_t k2 = 0;
-    uint32_t k3 = 0;
-    uint32_t k4 = 0;
-
-    switch (len & 15) {
-        case 15:
-            k4 ^= tail[14] << 16;
-        case 14:
[…]
