+++ setup.py
@@ -2,7 +2,9 @@
 
 from setuptools import Extension, setup
 
-mmh3module = Extension("mmh3", sources=["./src/mmh3/mmh3module.c"])
+mmh3module = Extension(
+    "mmh3", sources=["./src/mmh3/mmh3module.c", "./src/mmh3/murmurhash3.c"]
+)
 
 setup(
     ext_modules=[mmh3module],
