### 2024-09-07 9bf6c42
Eliminate branching in interfaces to mmh3_128
