+++ docs/benchmarks.md
@@ -42,9 +42,9 @@ Figure 3: Latency. Smaller is better.
 Figure 4: Latency for small data. Smaller is better.
 ```
 
-## Microbenchmarking Techniques and Tools
+## Notes on Python Microbenchmarking
 
-**Microbenchmarking** is a specialized type of benchmarking to measure the
+**Microbenchmarking** is a type of benchmarking aimed at measuring the
 performance of a small piece of code. It is crucial to compare the performance
 of different implementations of a group of related algorithms or to measure the
 impact of a change in the code. Although conceptually simple, microbenchmarking
@@ -66,7 +66,7 @@ performance dashboard for CPython.
 benchmark. It is designed to be user-friendly, offering a simple interface for
 benchmarking by allowing users to write benchmarks that integrate seamlessly
 with unit tests. It is also used as the primary backend for the Python
-environment in <code>[github-action-benchmark]</code>, a GitHub Action for
+environment in [github-action-benchmark], a GitHub Action for
 continuous benchmarking.
 
 **airspeed velocity** (`asv`) is a tool that supports continuous benchmarking
@@ -82,6 +82,68 @@ Each tool has its strengths and weaknesses. Both `pytest-benchmark` and
 `pyperf` requires more manual coding and setup for benchmarking, it offers
 greater flexiblity and can be applied to a broader range of scenarios.
 
+### Clock resolution
+
+In Python, the most basic way to measure time is by using the `time.time()`
+function. However, this function may not provide the highest resolution
+available on the system. In addition, it doesn't ensure monotonicity, meaning
+the time could move backward if the system clock has been set back. As a
+result, the difference between two calls might even be negative.
+
+````{danger}
+```python
+import time
+
+def bench_my_function_1() -> float:
+    t0 = time.time()
+
+    my_function()
+
+    # In fractional seconds
+    return time.time() - t0
+```
+````
+
+For higher resolution, you can use `time.perf_counter()`, introduced in
+Python 3.3. While it may not have much use in a single call, it provides the
+highest available resolution and is monotonic, making it ideal for benchmarking
+tasks.
+
+```python
+import time
+
+def bench_my_function_2() -> float:
+    t0 = time.perf_counter()
+
+    my_function()
+
+    # In fractional seconds
+    return time.perf_counter() - t0
+```
+
+Since Python 3.7, `time.perf_counter_ns()` is available, returning the time
+as an integer in nanoseconds.
+
+```python
+import time
+
+def bench_my_function_3() -> int:
+    t0 = time.perf_counter_ns()
+
+    my_function()
+
+    # In nanoseconds
+    return time.perf_counter_ns() - t0
+```
+
+If you're implementing your benchmarking logic from scratch, it is recommended
+to use `time.perf_counter_ns()` to avoid the rounding errors that can occur with
+with the float-based `time.perf_counter()`. However, the [bench_time_func()]
+function of pyperf 2.7.0 expects the return value to be the differences of
+`time.perf_counter()`, as the library predates the introduction of
+`time.perf_counter_ns()`. Choose the appropriate function depending on your
+approach.
+
 ### Iterator pre-instantiation
 
 In its simplest form, mircrobenchmarking a function involves writing a loop
@@ -89,6 +151,11 @@ that calls the function multiple times and measures the total time taken.
 This total is then divided by the number of iterations to calculate the average
 execution time per call.
 
+However, when microbenchmarking a small function in an interpreted language
+like Python, even the time taken to instantiate the iterator (`range(n)`) can be
[…]
