+++ pyproject.toml
@@ -59,6 +59,8 @@ profile = "black"
 src_paths = ["src/mmh3/__init__.pyi", "util", "tests", "benchmark", "docs"]
 
 [tool.pylint.main]
+ignore-paths = '^docs/examples/.*$'
+
 # Use multiple processes to speed up Pylint.
 # The value 0 specifies the number of processors to be auto-detected.
 # This setting can be found in the template file of super-linter 7.0.0.
