### 2024-09-09 7ea6f98
Check gha-ubuntu os settings
