+++ .github/workflows/check-kernel-configs.yml
@@ -1,0 +1,42 @@
+---
+name: Check Kernel Configs
+
+on: # yamllint disable-line rule:truthy
+  push:
+    branches:
+      - master
+      - feature/**
+  pull_request:
+    types:
+      - opened
+      - synchronize
+      - reopened
+
+permissions: {}
+
+jobs:
+  build:
+    permissions:
+      contents: read
+      packages: read
+
+    runs-on: ubuntu-22.04
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python 3.12
+        uses: actions/setup-python@v5
+        with:
+          python-version: 3.12
+      - name: check kernel configs
+        run: |
+          cat /sys/devices/system/cpu/isolated
+          cat /sys/devices/system/cpu/nohz_full
+          cat /proc/irq/default_smp_affinity
+          cat /proc/irq/0/smp_affinity
+          taskset -c 1,3 python3 -m timeit 'sum(range(10**7))'
+          python benchmark/test_os_affinity.py
+          taskset -c 1,3 python benchmark/test_os_affinity.py
+          lscpu --extended
+          lscpu -a -e
+          cat /proc/sys/kernel/perf_event_max_sample_rate
+          cat /sys/devices/system/cpu/intel_pstate/no_turbo
