+++ .github/workflows/check-kernel-configs.yml
@@ -34,7 +34,7 @@ jobs:
           cat /proc/irq/default_smp_affinity
           cat /proc/irq/0/smp_affinity
           taskset -c 1,3 python3 -m timeit 'sum(range(10**7))'
-          python benchmark/test_os_affinity.py
+          python benchmark/print_os_affinity.py
           taskset -c 1,3 python benchmark/print_os_affinity.py
           lscpu --extended
           lscpu -a -e
