### 2024-09-09 d2aef41
Fix kernel check
