+++ .github/workflows/check-kernel-configs.yml
@@ -15,7 +15,7 @@ on: # yamllint disable-line rule:truthy
 permissions: {}
 
 jobs:
-  build:
+  tune:
     permissions:
       contents: read
       packages: read
@@ -29,14 +29,14 @@ jobs:
           python-version: 3.12
       - name: check kernel configs
         run: |
-          cat /sys/devices/system/cpu/isolated
-          cat /sys/devices/system/cpu/nohz_full
-          cat /proc/irq/default_smp_affinity
-          cat /proc/irq/0/smp_affinity
+          echo -n "cpu/isolated: "; cat /sys/devices/system/cpu/isolated
+          echo -n "cpu/nohz_full: "; cat /sys/devices/system/cpu/nohz_full
+          echo -n "default_smp_affinity "; cat /proc/irq/default_smp_affinity
           taskset -c 1,3 python3 -m timeit 'sum(range(10**7))'
-          python benchmark/print_os_affinity.py
-          taskset -c 1,3 python benchmark/print_os_affinity.py
+          python benchmark/check_cpu_affinity.py
+          taskset -c 1,3 python benchmark/check_cpu_affinity.py
           lscpu --extended
           lscpu -a -e
-          cat /proc/sys/kernel/perf_event_max_sample_rate
-          cat /sys/devices/system/cpu/intel_pstate/no_turbo
+          echo -n "perf_event_max_sample_rate: "; cat /proc/sys/kernel/perf_event_max_sample_rate
+          echo 1 | sudo tee /proc/sys/kernel/perf_event_max_sample_rate
+          echo -n "perf_event_max_sample_rate now: "; cat /proc/sys/kernel/perf_event_max_sample_rate
