### 2024-09-09 921addb
More checks for gha-ubuntu
