+++ .github/workflows/check-kernel-configs.yml
@@ -41,6 +41,8 @@ jobs:
           echo 1 | sudo tee /proc/sys/kernel/perf_event_max_sample_rate > /dev/null
           echo -n "perf_event_max_sample_rate now updates to: "; cat /proc/sys/kernel/perf_event_max_sample_rate
           echo -n "randomize_va_space: "; cat /proc/sys/kernel/randomize_va_space
+          sudo modprobe intel_pstate
+          sudo modprobe acpi_cpufreq
           echo -n "cpu0/cpufreq/scaling_min_freq: "; cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_min_freq
           echo -n "cpu0/cpufreq/scaling_available_frequencies: "; cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_available_frequencies
           echo -n "cpu0/cpufreq/cpuinfo_min_freq: "; cat /sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_min_freq
