+++ .github/workflows/check-kernel-configs.yml
@@ -43,6 +43,8 @@ jobs:
           echo -n "randomize_va_space: "; cat /proc/sys/kernel/randomize_va_space
           sudo modprobe intel_pstate
           sudo modprobe acpi_cpufreq
+          sudo modprobe cpufreq_userspace
+          sudo modprobe cpufreq_ondemand
           echo -n "cpu0/cpufreq/scaling_min_freq: "; cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_min_freq
           echo -n "cpu0/cpufreq/scaling_available_frequencies: "; cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_available_frequencies
           echo -n "cpu0/cpufreq/cpuinfo_min_freq: "; cat /sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_min_freq
