### 2024-09-09 012456c
change irq affinity
