+++ pyproject.toml
@@ -59,14 +59,14 @@ profile = "black"
 src_paths = ["src/mmh3/__init__.pyi", "util", "tests", "benchmark", "docs"]
 
 [tool.pylint.MASTER]
-ignore-paths = "^docs/examples/.*$"
+ignore-paths = "^docs/.*$"
 
 [tool.pylint.main]
 # Use multiple processes to speed up Pylint.
 # The value 0 specifies the number of processors to be auto-detected.
 # This setting can be found in the template file of super-linter 7.0.0.
 jobs=0
-ignore-paths = "^docs/examples/.*$"
+ignore-paths = "^docs/.*$"
 
 [tool.pylint."messages control"]
 # import-error: An error tricky to resolve, especially on super-linter.
