+++ .github/workflows/check-kernel-configs.yml
@@ -31,7 +31,9 @@ jobs:
         run: |
           echo -n "cpu/isolated: "; cat /sys/devices/system/cpu/isolated
           echo -n "cpu/nohz_full: "; cat /sys/devices/system/cpu/nohz_full
-          echo -n "default_smp_affinity "; cat /proc/irq/default_smp_affinity
+          echo -n "default_smp_affinity: "; cat /proc/irq/default_smp_affinity
+          echo 3 | sudo tee /proc/irq/default_smp_affinity > /dev/null
+          echo -n "default_smp_affinity updated to: "; cat /proc/irq/default_smp_affinity
           taskset -c 1,3 python3 -m timeit 'sum(range(10**7))'
           python benchmark/check_cpu_affinity.py
           taskset -c 1,3 python benchmark/check_cpu_affinity.py
@@ -39,15 +41,5 @@ jobs:
           lscpu -a -e
           echo -n "perf_event_max_sample_rate: "; cat /proc/sys/kernel/perf_event_max_sample_rate
           echo 1 | sudo tee /proc/sys/kernel/perf_event_max_sample_rate > /dev/null
-          echo -n "perf_event_max_sample_rate now updates to: "; cat /proc/sys/kernel/perf_event_max_sample_rate
+          echo -n "perf_event_max_sample_rate now updated to: "; cat /proc/sys/kernel/perf_event_max_sample_rate
           echo -n "randomize_va_space: "; cat /proc/sys/kernel/randomize_va_space
-          sudo modprobe intel_pstate
-          sudo modprobe acpi_cpufreq
-          sudo modprobe cpufreq_userspace
-          sudo modprobe cpufreq_ondemand
-          echo -n "cpu0/cpufreq/scaling_min_freq: "; cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_min_freq
-          echo -n "cpu0/cpufreq/scaling_available_frequencies: "; cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_available_frequencies
-          echo -n "cpu0/cpufreq/cpuinfo_min_freq: "; cat /sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_min_freq
-          echo -n "cpu3/cpufreq/scaling_min_freq: "; cat /sys/devices/system/cpu/cpu3/cpufreq/scaling_min_freq
-          echo -n "cpu3/cpufreq/scaling_available_frequencies: "; cat /sys/devices/system/cpu/cpu3/cpufreq/scaling_available_frequencies
-          echo -n "cpu3/cpufreq/cpuinfo_min_freq: "; cat /sys/devices/system/cpu/cpu3/cpufreq/cpuinfo_min_freq
