### 2024-09-09 68fea6f
Add numactl checks
