+++ .github/workflows/check-kernel-configs.yml
@@ -47,4 +47,5 @@ jobs:
           which irqbalance
           sudo systemctl stop irqbalance
           systemctl status irqbalance
+          sudo apt install numactl
           numactl --hardware
