+++ pyproject.toml
@@ -58,15 +58,12 @@ Changelog = "https://github.com/hajimes/mmh3/blob/master/CHANGELOG.md"
 profile = "black"
 src_paths = ["src/mmh3/__init__.pyi", "util", "tests", "benchmark", "docs"]
 
-[tool.pylint.MASTER]
-ignore-paths = "^docs/.*$"
-
 [tool.pylint.main]
 # Use multiple processes to speed up Pylint.
 # The value 0 specifies the number of processors to be auto-detected.
 # This setting can be found in the template file of super-linter 7.0.0.
-jobs=0
-ignore-paths = "^docs/.*$"
+jobs = 0
+ignore-paths = ["docs"]
 
 [tool.pylint."messages control"]
 # import-error: An error tricky to resolve, especially on super-linter.
