+++ .github/workflows/check-kernel-configs.yml
@@ -46,4 +46,5 @@ jobs:
           systemctl status irqbalance
           which irqbalance
           sudo systemctl stop irqbalance
-          systemctl status irqbalance
+          sudo apt install numactl
+          numactl --hardware
