### 2024-09-10 2b755f2
Deserialize benchmark
