+++ .github/workflows/benchmark.yml
@@ -11,12 +11,11 @@ jobs:
     permissions:
       contents: read
       packages: read
-
-    strategy:
-      matrix:
-        hash: [mmh3_32, mmh3_128, xxh_32, xxh_64, xxh3_64, xxh3_128, md5, sha1]
-
     runs-on: ubuntu-22.04
+    env:
+      HASH_LIST: ("mmh3_32" "mmh3_128" "xxh_32" "xxh_64" "xxh3_64"\
+            "xxh3_128" "md5" "sha1")
+      BENCHMARK_MAX_SIZE: 2097152
     steps:
       - name: Checkout
         uses: actions/checkout@v4
@@ -32,30 +31,12 @@ jobs:
       - name: Benchmark hash functions
         run: |
           mkdir var
-          python benchmark/benchmark.py -o var/${{ matrix.hash }}.json \
-            --test-hash ${{ matrix.hash }} --test-buffer-size-max 134217728
-      - name: Upload artifacts
-        uses: actions/upload-artifact@v4
-        with:
-          name: benchmark-result-${{ matrix.hash }}
-          path: var/${{ matrix.hash }}.json
-  postprocess:
-    name: Postprocess benchmark results
-    runs-on: ubuntu-22.04
-    needs: benchmark
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - uses: actions/download-artifact@v4
-        with:
-          path: ./artifacts
-      - name: Zip benchmark results
-        run: |
-          mkdir var
-          mkdir var/benchmark_all_results
-          cp artifacts/*/*.json var/benchmark_all_results
+          for hash_name in "${HASH_LIST[@]}"; do
+            python benchmark/benchmark.py -o var/${hash_name}.json \
+              --test-hash ${hash_name} --test-buffer-size-max ${BENCHMARK_MAX_SIZE}
+          done
       - name: Upload artifacts
         uses: actions/upload-artifact@v4
         with:
           name: benchmark-results
-          path: var/benchmark_all_results
+          path: var
