+++ .github/workflows/benchmark.yml
@@ -29,8 +29,8 @@ jobs:
       - name: Benchmark hash functions
         run: |
           mkdir var
-          hash_list = ("mmh3_32" "mmh3_128" "xxh_32" "xxh_64" "xxh3_64"\
-            "xxh3_128" "md5" "sha1")
+          declare -a hash_list=("mmh3_32" "mmh3_128" "xxh_32" "xxh_64" \
+             "xxh3_64" "xxh3_128" "md5" "sha1")
           for hash_name in "${hash_list[@]}"; do
             python benchmark/benchmark.py -o var/${hash_name}.json \
               --test-hash ${hash_name} \
