### 2024-09-10 9f648d8
Add more references; change graphs
