### 2024-09-10 b8a702f
Refine benchmarking and plot_graph
