+++ .github/workflows/check-kernel-configs.yml
@@ -1,48 +1,0 @@
----
-name: Check Kernel Configs
-
-on: # yamllint disable-line rule:truthy
-  push:
-    branches:
-      - master
-      - feature/**
-  pull_request:
-    types:
-      - opened
-      - synchronize
-      - reopened
-
-permissions: {}
-
-jobs:
-  tune:
-    permissions:
-      contents: read
-      packages: read
-
-    runs-on: ubuntu-22.04
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python 3.12
-        uses: actions/setup-python@v5
-        with:
-          python-version: 3.12
-      - name: check kernel configs
-        run: |
-          echo -n "cpu/isolated: "; cat /sys/devices/system/cpu/isolated
-          echo -n "cpu/nohz_full: "; cat /sys/devices/system/cpu/nohz_full
-          echo -n "default_smp_affinity: "; cat /proc/irq/default_smp_affinity
-          echo 3 | sudo tee /proc/irq/default_smp_affinity > /dev/null
-          echo -n "default_smp_affinity updated to: "; cat /proc/irq/default_smp_affinity
-          taskset -c 1,3 python3 -m timeit 'sum(range(10**7))'
-          python benchmark/check_cpu_affinity.py
-          taskset -c 1,3 python benchmark/check_cpu_affinity.py
-          lscpu --extended
-          lscpu -a -e
-          echo -n "perf_event_max_sample_rate: "; cat /proc/sys/kernel/perf_event_max_sample_rate
-          echo 1 | sudo tee /proc/sys/kernel/perf_event_max_sample_rate > /dev/null
-          echo -n "perf_event_max_sample_rate now updated to: "; cat /proc/sys/kernel/perf_event_max_sample_rate
-          echo -n "randomize_va_space: "; cat /proc/sys/kernel/randomize_va_space
-          systemctl status irqbalance
-          which irqbalance
-          sudo systemctl stop irqbalance
