### 2024-09-10 bff23fa
Deparallelize and tune benchmark workflow
