+++ .github/workflows/superlinter.yml
@@ -50,3 +50,15 @@ jobs:
           VALIDATE_JSCPD: false
           # Suppressed because it even accuses book titles
           VALIDATE_NATURAL_LANGUAGE: false
+          # Suppressed because it does not honor the ignore-paths option
+          VALIDATE_PYTHON_PYLINT: false
+      # super-linter 7 does not honor the ignore-paths option of pylint
+      # so we run pylint separately
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: 3.12
+      - name: Run pylint
+        run: |
+          pip install pylint
+          pylint --recursive=y .
