### 2024-09-10 bcf3c1c
Remove redundant line in CONTRIBUTING
