### 2024-09-11 5922225
Add bytes returning functions for mutable input
