+++ tests/test_mmh3.py
@@ -54,10 +54,6 @@ def test_hash() -> None:
         "The quick brown fox jumps over the lazy dog", u32_to_s32(0x9747B28C)
     ) == u32_to_s32(0x2FA826CD)
 
-    assert mmh3.hash(
-        "The quick brown fox jumps over the lazy dog", u32_to_s32(0x9747B28C)
-    ) == u32_to_s32(0x2FA826CD)
-
 
 def test_hash_unsigned() -> None:
     assert mmh3.hash("foo", signed=False) == 4138058784
@@ -235,6 +231,113 @@ def test_hash128() -> None:
     assert mmh3.hash128("", 123, False, False) == 0x26F3E79926F3E79926F3E799FEDC5245
 
 
+def test_mmh3_32_digest() -> None:
+    assert mmh3.mmh3_32_digest(b"") == b"\0\0\0\0"
+    assert mmh3.mmh3_32_digest(b"", 0) == b"\0\0\0\0"
+    assert mmh3.mmh3_32_digest(b"", seed=0) == b"\0\0\0\0"
+    assert mmh3.mmh3_32_digest(b"\x21\x43\x65\x87", 0) == (0xF55B516B).to_bytes(
+        4, "little"
+    )
+    assert mmh3.mmh3_32_digest(b"\x21\x43\x65\x87", u32_to_s32(0x5082EDEE)) == (
+        0x2362F9DE
+    ).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"\x21\x43\x65", 0) == (0x7E4A8634).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"\x21\x43", 0) == (0xA0F7B07A).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"\x21", 0) == (0x72661CF4).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"\xff\xff\xff\xff", 0) == (0x76293B50).to_bytes(
+        4, "little"
+    )
+    assert mmh3.mmh3_32_digest(b"\x00\x00\x00\x00", 0) == (0x2362F9DE).to_bytes(
+        4, "little"
+    )
+    assert mmh3.mmh3_32_digest(b"\x00\x00\x00", 0) == (0x85F0B427).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"\x00\x00", 0) == (0x30F4C306).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"\x00", 0) == (0x514E28B7).to_bytes(4, "little")
+
+    assert mmh3.mmh3_32_digest("aaaa", 0x9747B28C) == (0x5A97808A).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest("aaa", 0x9747B28C) == (0x283E0130).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest("aa", 0x9747B28C) == (0x5D211726).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest("a", 0x9747B28C) == (0x7FA09EA6).to_bytes(4, "little")
+
+    assert mmh3.mmh3_32_digest("abcd", 0x9747B28C) == (0xF0478627).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest("abc", 0x9747B28C) == (0xC84A62DD).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest("ab", 0x9747B28C) == (0x74875592).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest("a", 0x9747B28C) == (0x7FA09EA6).to_bytes(4, "little")
+
[…]
