+++ docs/api.rst
@@ -1,6 +1,9 @@
 API Reference
 ================
 
+.. caution::
+  This reference contains functions that are not yet released in the stable version.
+
 .. automodule:: mmh3
    :members:
    :undoc-members:
