### 2024-09-11 77f6f28
Add intdigests for mmh3_32
