+++ tests/test_mmh3.py
@@ -293,6 +293,141 @@ def test_mmh3_32_digest() -> None:
     assert mmh3.mmh3_32_digest(v[1:3], 0x9747B28C) == (0x5D211726).to_bytes(4, "little")
 
 
+def test_mmh3_sintdigest() -> None:
+    assert mmh3.mmh3_32_sintdigest("foo") == -156908512
+    assert mmh3.mmh3_32_sintdigest(b"foo") == -156908512
+    assert mmh3.mmh3_32_sintdigest(bytearray(b"foo")) == -156908512
+    assert mmh3.mmh3_32_sintdigest(memoryview(b"foobar")[0:3]) == -156908512
+
+    # Test vectors devised by Ian Boyd
+    # https://stackoverflow.com/a/31929528
+    assert mmh3.mmh3_32_sintdigest(b"", seed=0) == 0
+    assert mmh3.mmh3_32_sintdigest(b"", seed=1) == 0x514E28B7
+    assert mmh3.mmh3_32_sintdigest(b"", seed=u32_to_s32(0xFFFFFFFF)) == u32_to_s32(
+        0x81F16F39
+    )
+    assert mmh3.mmh3_32_sintdigest(b"\x21\x43\x65\x87", 0) == u32_to_s32(0xF55B516B)
+    assert mmh3.mmh3_32_sintdigest(
+        b"\x21\x43\x65\x87", u32_to_s32(0x5082EDEE)
+    ) == u32_to_s32(0x2362F9DE)
+    assert mmh3.mmh3_32_sintdigest(b"\x21\x43\x65", 0) == u32_to_s32(0x7E4A8634)
+    assert mmh3.mmh3_32_sintdigest(b"\x21\x43", 0) == u32_to_s32(0xA0F7B07A)
+    assert mmh3.mmh3_32_sintdigest(b"\x21", 0) == u32_to_s32(0x72661CF4)
+    assert mmh3.mmh3_32_sintdigest(b"\xff\xff\xff\xff", 0) == u32_to_s32(0x76293B50)
+    assert mmh3.mmh3_32_sintdigest(b"\x00\x00\x00\x00", 0) == u32_to_s32(0x2362F9DE)
+    assert mmh3.mmh3_32_sintdigest(b"\x00\x00\x00", 0) == u32_to_s32(0x85F0B427)
+    assert mmh3.mmh3_32_sintdigest(b"\x00\x00", 0) == u32_to_s32(0x30F4C306)
+    assert mmh3.mmh3_32_sintdigest(b"\x00", 0) == u32_to_s32(0x514E28B7)
+
+    assert mmh3.mmh3_32_sintdigest("aaaa", u32_to_s32(0x9747B28C)) == u32_to_s32(
+        0x5A97808A
+    )
+    assert mmh3.mmh3_32_sintdigest("aaa", u32_to_s32(0x9747B28C)) == u32_to_s32(
+        0x283E0130
+    )
+    assert mmh3.mmh3_32_sintdigest("aa", u32_to_s32(0x9747B28C)) == u32_to_s32(
+        0x5D211726
+    )
+    assert mmh3.mmh3_32_sintdigest("a", u32_to_s32(0x9747B28C)) == u32_to_s32(
+        0x7FA09EA6
+    )
+
+    assert mmh3.mmh3_32_sintdigest("abcd", u32_to_s32(0x9747B28C)) == u32_to_s32(
+        0xF0478627
+    )
+    assert mmh3.mmh3_32_sintdigest("abc", u32_to_s32(0x9747B28C)) == u32_to_s32(
+        0xC84A62DD
+    )
[…]
