### 2024-09-11 a2dee1f
Add intdigests for 128-bit algs
