+++ src/mmh3/mmh3module.c
@@ -295,7 +295,7 @@ mmh3_mmh3_32_digest(PyObject *self, PyObject *args, PyObject *keywds)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
-    const char result[MMH3_32_DIGESTSIZE];
+    char result[MMH3_32_DIGESTSIZE];
 
     static char *kwlist[] = {(char *)"key", (char *)"seed", NULL};
 
@@ -311,8 +311,7 @@ mmh3_mmh3_32_digest(PyObject *self, PyObject *args, PyObject *keywds)
     ((uint32_t *)result)[0] = bswap_32(((uint32_t *)result)[0]);
 #endif
 
-    return PyBytes_FromStringAndSize((unsigned char *)result,
-                                     MMH3_32_DIGESTSIZE);
+    return PyBytes_FromStringAndSize((char *)result, MMH3_32_DIGESTSIZE);
 }
 
 PyDoc_STRVAR(
@@ -477,6 +476,62 @@ mmh3_mmh3_x64_128_uintdigest(PyObject *self, PyObject *args, PyObject *keywds)
     return retval;
 }
 
+PyDoc_STRVAR(mmh3_mmh3_x64_128_stupledigest_doc,
+             "mmh3_x64_128_stupledigest(key[, seed=0]) -> tuple[int, int]\n\n"
+             "Return a hash value from a memory buffer as a tuple of two "
+             "64-bit signed integers. "
+             "Calculated by the MurmurHash3_x64_128 algorithm. ");
+
+static PyObject *
+mmh3_mmh3_x64_128_stupledigest(PyObject *self, PyObject *args,
+                               PyObject *keywds)
+{
+    Py_buffer target_buf;
+    uint32_t seed = 0;
+    uint64_t result[2];
+
+    static char *kwlist[] = {(char *)"key", (char *)"seed", NULL};
+
+    if (!PyArg_ParseTupleAndKeywords(args, keywds, "s*|I", kwlist, &target_buf,
+                                     &seed)) {
+        return NULL;
+    }
+
+    murmurhash3_x64_128(target_buf.buf, target_buf.len, seed, result);
+    PyBuffer_Release(&target_buf);
+
+    PyObject *retval = Py_BuildValue("LL", result[0], result[1]);
+    return retval;
+}
+
+PyDoc_STRVAR(mmh3_mmh3_x64_128_utupledigest_doc,
+             "mmh3_x64_128_utupledigest(key[, seed=0]) -> tuple[int, int]\n\n"
+             "Return a hash value from a memory buffer as a tuple of two "
+             "64-bit unsigned integers. "
+             "Calculated by the MurmurHash3_x64_128 algorithm. ");
+
+static PyObject *
+mmh3_mmh3_x64_128_utupledigest(PyObject *self, PyObject *args,
+                               PyObject *keywds)
+{
+    Py_buffer target_buf;
+    uint32_t seed = 0;
+    uint64_t result[2];
+
+    static char *kwlist[] = {(char *)"key", (char *)"seed", NULL};
+
+    if (!PyArg_ParseTupleAndKeywords(args, keywds, "s*|I", kwlist, &target_buf,
+                                     &seed)) {
+        return NULL;
+    }
+
+    murmurhash3_x64_128(target_buf.buf, target_buf.len, seed, result);
+    PyBuffer_Release(&target_buf);
+
+    PyObject *retval = Py_BuildValue("KK", result[0], result[1]);
+    return retval;
+}
+
 PyDoc_STRVAR(mmh3_mmh3_x86_128_digest_doc,
[…]
