### 2024-09-11 3b1fff4
Add funcs for non-immutables w/o copying (#75)
