+++ src/mmh3/mmh3module.c
@@ -124,6 +124,8 @@ mmh3_hash_from_buffer(PyObject *self, PyObject *args, PyObject *keywds)
 
     murmurhash3_x86_32(target_buf.buf, target_buf.len, seed, result);
 
+    PyBuffer_Release(&target_buf);
+
 #if defined(_MSC_VER)
     /* for Windows envs */
     long_result = result[0];
@@ -282,6 +284,421 @@ mmh3_hash_bytes(PyObject *self, PyObject *args, PyObject *keywds)
     return PyBytes_FromStringAndSize((char *)result, MMH3_128_DIGESTSIZE);
 }
 
+PyDoc_STRVAR(mmh3_mmh3_32_digest_doc,
+             "mmh3_32_digest(key[, seed=0]) -> bytes\n\n"
+             "Return a hash value from a memory buffer as bytes. "
+             "Calculated by the MurmurHash3_x86_32 algorithm. "
+             "Designed for large memory-views such as numpy arrays.");
+
+static PyObject *
+mmh3_mmh3_32_digest(PyObject *self, PyObject *args, PyObject *keywds)
+{
+    Py_buffer target_buf;
+    uint32_t seed = 0;
+    char result[MMH3_32_DIGESTSIZE];
+
+    static char *kwlist[] = {(char *)"key", (char *)"seed", NULL};
+
+    if (!PyArg_ParseTupleAndKeywords(args, keywds, "s*|I", kwlist, &target_buf,
+                                     &seed)) {
+        return NULL;
+    }
+
+    murmurhash3_x86_32(target_buf.buf, target_buf.len, seed, result);
+    PyBuffer_Release(&target_buf);
+
+#if defined(__BYTE_ORDER__) && (__BYTE_ORDER__ == __ORDER_BIG_ENDIAN__)
+    ((uint32_t *)result)[0] = bswap_32(((uint32_t *)result)[0]);
+#endif
+
+    return PyBytes_FromStringAndSize((char *)result, MMH3_32_DIGESTSIZE);
+}
+
+PyDoc_STRVAR(
+    mmh3_mmh3_32_sintdigest_doc,
+    "mmh3_32_sintdigest(key[, seed=0]) -> 32-bit int\n\n"
+    "Return a hash value from a memory buffer as a signed 32-bit integer. "
+    "Calculated by the MurmurHash3_x86_32 algorithm. ");
+
+static PyObject *
+mmh3_mmh3_32_sintdigest(PyObject *self, PyObject *args, PyObject *keywds)
+{
+    Py_buffer target_buf;
+    uint32_t seed = 0;
+    int32_t result[1];
+
+    static char *kwlist[] = {(char *)"key", (char *)"seed", NULL};
+
+    if (!PyArg_ParseTupleAndKeywords(args, keywds, "s*|I", kwlist, &target_buf,
+                                     &seed)) {
+        return NULL;
+    }
+
+    murmurhash3_x86_32(target_buf.buf, target_buf.len, seed, result);
+    PyBuffer_Release(&target_buf);
+
+    return PyLong_FromLong(result[0]);
+}
+
+PyDoc_STRVAR(
+    mmh3_mmh3_32_uintdigest_doc,
+    "mmh3_32_uintdigest(key[, seed=0]) -> 32-bit int\n\n"
+    "Return a hash value from a memory buffer as an unsigned 32-bit integer. "
+    "Calculated by the MurmurHash3_x86_32 algorithm. ");
+
+static PyObject *
+mmh3_mmh3_32_uintdigest(PyObject *self, PyObject *args, PyObject *keywds)
+{
+    Py_buffer target_buf;
+    uint32_t seed = 0;
+    uint32_t result[1];
+
+    static char *kwlist[] = {(char *)"key", (char *)"seed", NULL};
+
[…]
