+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This project has adhered to
 
 - Fix a reference leak in the `hash_from_buffer()` function
   ([#75](https://github.com/hajimes/mmh3/issues/75)).
+- Fix type hints.
 
 ## [4.1.0] - 2024-01-09
 
