### 2024-09-12 cf6f8a8
Fix and improve type hints (#76)
