+++ pyproject.toml
@@ -36,7 +36,8 @@ test = [
 docs = [
   "myst-parser == 4.0.0",
   "shibuya == 2024.8.30",
-  "sphinx == 8.0.2"
+  "sphinx == 8.0.2",
+  "sphinx-copybutton == 0.5.2."
 ]
 benchmark = [
   "pymmh3 == 0.0.5",
