+++ src/mmh3/mmh3module.c
@@ -37,16 +37,21 @@ typedef unsigned __int64 uint64_t;
 
 PyDoc_STRVAR(
     mmh3_hash_doc,
-    "hash(key: bytes | str, seed: int = 0, signed: bool = True) -> int\n\n"
-    "Return a hash value as a 32-bit integer.\n\n"
-    "Calculated by the MurmurHash3_x86_32 algorithm.\n\n"
+    "hash(key[, seed=0, signed=True]) -> int\n"
+    "\n"
+    "Return a hash as a 32-bit integer.\n"
+    "\n"
+    "Calculated by the MurmurHash3_x86_32 algorithm.\n"
+    "\n"
     "Args:\n"
-    "    key: The input data to hash.\n"
-    "    seed: The seed value.\n\n"
-    "    signed: If True, return a signed integer. Otherwise, return an "
-    "unsigned.\n\n"
+    "    key (bytes | str): The input data to hash.\n"
+    "    seed (int): The seed value. Must be an integer in the range [0, "
+    "0xFFFFFFFF].\n"
+    "    signed (bool): If True, return a signed integer. Otherwise, return "
+    "an unsigned integer.\n"
+    "\n"
     "Returns:\n"
-    "    The hash value as a 32-bit integer.\n\n");
+    "    int: The hash value as a 32-bit integer.\n");
 
 static PyObject *
 mmh3_hash(PyObject *self, PyObject *args, PyObject *keywds)
@@ -101,11 +106,24 @@ mmh3_hash(PyObject *self, PyObject *args, PyObject *keywds)
 #endif  // defined(_MSC_VER)
 }
 
-PyDoc_STRVAR(mmh3_hash_from_buffer_doc,
-             "hash_from_buffer(key[, seed=0, signed=True]) -> 32-bit int\n\n"
-             "Return a hash value from a memory buffer as a 32-bit integer. "
-             "Calculated by the MurmurHash3_x86_32 algorithm. "
-             "Designed for large memory-views such as numpy arrays.");
+PyDoc_STRVAR(
+    mmh3_hash_from_buffer_doc,
+    "hash_from_buffer(key[, seed=0, signed=True]) -> int\n"
+    "\n"
+    "Return a hash from a memory buffer as a 32-bit integer.\n"
+    "\n"
+    "Calculated by the MurmurHash3_x86_32 algorithm. Designed for large "
+    "memory-views such as numpy arrays.\n"
+    "\n"
+    "Args:\n"
+    "    key (Buffer): The input data to hash.\n"
+    "    seed (int): The seed value. Must be an integer in the range [0, "
+    "0xFFFFFFFF].\n"
+    "    signed (bool): If True, return a signed integer. Otherwise, return "
+    "an unsigned integer.\n"
+    "\n"
+    "Returns:\n"
+    "    int: The hash value as a 32-bit integer.\n");
 
 static PyObject *
 mmh3_hash_from_buffer(PyObject *self, PyObject *args, PyObject *keywds)
@@ -162,12 +180,24 @@ mmh3_hash_from_buffer(PyObject *self, PyObject *args, PyObject *keywds)
 
 PyDoc_STRVAR(
     mmh3_hash64_doc,
-    "hash64(key[, seed=0, x64arch=True, signed=True]) -> (64-bit int, "
-    "64-bit "
-    "int)\n\n"
-    "Return a tuple of two 64 bit integers given an input string. "
-    "Calculated by the MurmurHash3_x{64, 86}_128 algorithm. Optimized "
-    "for the x64 bit architecture when x64arch=True, otherwise for x86.");
[…]
