+++ pyproject.toml
@@ -37,7 +37,7 @@ docs = [
   "myst-parser == 4.0.0",
   "shibuya == 2024.8.30",
   "sphinx == 8.0.2",
-  "sphinx-copybutton == 0.5.2."
+  "sphinx-copybutton == 0.5.2"
 ]
 benchmark = [
   "pymmh3 == 0.0.5",
