### 2024-09-13 c7e195a
Improve docstring further
