+++ src/mmh3/mmh3module.c
@@ -110,13 +110,13 @@ PyDoc_STRVAR(
     mmh3_hash_from_buffer_doc,
     "hash_from_buffer(key[, seed=0, signed=True]) -> int\n"
     "\n"
-    "Return a hash from a memory buffer as a 32-bit integer.\n"
+    "Return a hash for the buffer as a 32-bit integer.\n"
     "\n"
     "Calculated by the MurmurHash3_x86_32 algorithm. Designed for large "
     "memory-views such as numpy arrays.\n"
     "\n"
     "Args:\n"
-    "    key (Buffer): The input data to hash.\n"
+    "    key (Buffer | str): The bufer to hash.\n"
     "    seed (int): The seed value. Must be an integer in the range [0, "
     "0xFFFFFFFF].\n"
     "    signed (bool): If True, return a signed integer. Otherwise, return "
@@ -233,19 +233,22 @@ mmh3_hash64(PyObject *self, PyObject *args, PyObject *keywds)
 
 PyDoc_STRVAR(
     mmh3_hash128_doc,
-    "hash128(key[, seed=0, x64arch=True, signed=False]]) -> int\n\n"
+    "hash128(key[, seed=0, x64arch=True, signed=False]]) -> int\n"
+    "\n"
     "Return a hash as a 128-bit integer.\n\n"
-    "Calculated by the MurmurHash3_x{64, 86}_128 algorithm.\n\n"
+    "Calculated by the MurmurHash3_x{64, 86}_128 algorithm.\n"
+    "\n"
     "Args:\n"
     "    key (bytes | str): The input data to hash.\n"
     "    seed (int): The seed value. Must be an integer in the range [0, "
-    "0xFFFFFFFF]. \n\n"
+    "0xFFFFFFFF].\n"
     "    x64arch (bool): If True, use an algorithm optimized for 64-bit "
-    "architecture. Otherwise, use one optimized for 32-bit architecture.\n\n"
+    "architecture. Otherwise, use one optimized for 32-bit architecture.\n"
     "    signed (bool): If True, return a signed integer. Otherwise, return "
-    "an unsigned integer.\n\n"
+    "an unsigned integer.\n"
+    "\n"
     "Returns:\n"
-    "    int: The hash value as a 128-bit integer.\n\n");
+    "    int: The hash value as a 128-bit integer.\n");
 
 static PyObject *
 mmh3_hash128(PyObject *self, PyObject *args, PyObject *keywds)
@@ -292,8 +295,10 @@ mmh3_hash128(PyObject *self, PyObject *args, PyObject *keywds)
 
 PyDoc_STRVAR(
     mmh3_hash_bytes_doc,
-    "hash_bytes(key[, seed=0, x64arch=True]) -> bytes\n\n"
-    "Return a 16-byte hash of the bytes type.\n\n"
+    "hash_bytes(key[, seed=0, x64arch=True]) -> bytes\n"
+    "\n"
+    "Return a 16-byte hash of the ``bytes`` type.\n"
+    "\n"
     "Args:\n"
     "    key (bytes | str): The input data to hash.\n"
     "    seed (int): The seed value. Must be an integer in the range [0, "
@@ -302,8 +307,8 @@ PyDoc_STRVAR(
     "architecture. Otherwise, use one optimized for 32-bit architecture.\n"
     "\n"
     "Returns:\n"
-    "    bytes: The hash value as the bytes type with a length of 16 bytes "
-    "(128 bits).\n");
[…]
