### 2024-09-13 b3a1f78
Add section explanations for API Reference
