### 2024-09-13 c34c627
Refine docstrings and improve API reference (#77)
