+++ docs/benchmarks.md
@@ -1,0 +1,412 @@
+# Benchmarks
+
+## Settings
+
+- Python environment
+  - CPython 3.12.5 (64-bit)
+- Ubuntu 22.04 instance on GitHub Actions
+- Hash libraries
+  - mmh3 4.2.0: dev version
+  - xxhash 3.5.0
+- pyperf 2.7.0
+
+## Results
+
+### Graphs
+
+```{figure} _static/bandwidth.png
+:alt: Latency
+:align: center
+
+Figure 1: Output bandwidth. Larger is better. The y-axis is logscale.
+```
+
+```{figure} _static/bandwidth_small.png
+:alt: Latency
+:align: center
+
+Figure 2: Output bandwidth for small data. Larger is better.
+```
+
+```{figure} _static/latency.png
+:alt: Latency
+:align: center
+
+Figure 3: Latency. Smaller is better.
+```
+
+```{figure} _static/latency_small.png
+:alt: Latency
+:align: center
+
+Figure 4: Latency for small data. Smaller is better.
+```
+
+## Notes on Python Microbenchmarking
+
+**Microbenchmarking** is a type of benchmarking aimed at measuring the
+performance of a small piece of code. It is crucial to compare the performance
+of different implementations of a group of related algorithms or to measure the
+impact of a change in the code. Although conceptually simple, microbenchmarking
+can be challenging in practice, particularly in Python, a scripting language.
+
+### Tools
+
+There are three popular libraries for microbenchmarking in Python: [pyperf],
+[pytest-benchmark], and [airspeed velocity] (also known as `asv`).
+
+**pyperf** is a general-purpose Python performance benchmarking tool, focused
+on providing reliable and accurate benchmarks with detailed reporting and
+statistical analysis. It is the primary backend for [pyperformance], an
+authoritative benchmark suite for Python interpreters, which in turn is used as
+the backend of [Python Speed Center](https://speed.python.org), the official
+performance dashboard for CPython.
+
+**pytest-benchmark** is a plugin for pytest that provides a fixture to
+benchmark. It is designed to be user-friendly, offering a simple interface for
+benchmarking by allowing users to write benchmarks that integrate seamlessly
+with unit tests. It is also used as the primary backend for the Python
+environment in [github-action-benchmark], a GitHub Action for
+continuous benchmarking.
+
+**airspeed velocity** (`asv`) is a tool that supports continuous benchmarking
+for indiviual projects, with built-in native support for a web interface.
+It also features a command-line interface similart to Git, making it to manage
+the history of benchmarks. Both
+[NumPy](https://numpy.org/devdocs/benchmarking.html) and
+[SciPy](https://docs.scipy.org/doc/scipy/dev/contributor/benchmarking.html)
+use asv as their benchmarking tool.
+
+Each tool has its strengths and weaknesses. Both `pytest-benchmark` and
+`asv` are tailored for specific use cases and excel in those contexts. While
+`pyperf` requires more manual coding and setup for benchmarking, it offers
+greater flexiblity and can be applied to a broader range of scenarios.
+
+### Clock resolution
+
+In Python, the most basic way to measure time is by using the `time.time()`
+function. However, this function may not provide the highest resolution
+available on the system. In addition, it doesn't ensure monotonicity, meaning
+the time could move backward if the system clock has been set back. As a
+result, the difference between two calls might even be negative.
+
+````{danger}
+```python
+import time
+
+
+def bench_my_function_1() -> float:
+    t0 = time.time()
+
+    my_function()
+
+    # In fractional seconds
+    return time.time() - t0
+```
+````
+
[…]
