+++ pyproject.toml
@@ -64,7 +64,8 @@ ignore-paths = [
   "^build",
   "^venv",
   "^.venv",
-  "^src/mmh3/__init__.pyi"
+  "^src/mmh3/__init__.pyi",
+  "^docs/examples/"
 ]
 # Use multiple processes to speed up Pylint.
 # The value 0 specifies the number of processors to be auto-detected.
