### 2024-09-14 20a4ab5
Add command line argument for bench type
