### 2024-09-14 d39ef3a
Add random-sized benchmarking methods
