+++ benchmark/benchmark.py
@@ -2,6 +2,8 @@
 
 import hashlib
 import itertools
+import math
+import random
 import time
 from collections.abc import Callable
 from typing import Final
@@ -14,19 +16,6 @@ import xxhash
 K1: Final[int] = 0b1001111000110111011110011011000110000101111010111100101010000111
 K2: Final[int] = 0b1100001010110010101011100011110100100111110101001110101101001111
 MASK: Final[int] = 0xFFFFFFFFFFFFFFFF
-
-HASHES = {
-    "mmh3_32": mmh3.mmh3_32_digest,
-    "mmh3_128": mmh3.mmh3_x64_128_digest,
-    "xxh_32": xxhash.xxh32_digest,
-    "xxh_64": xxhash.xxh64_digest,
-    "xxh3_64": xxhash.xxh3_64_digest,
-    "xxh3_128": xxhash.xxh3_128_digest,
-    "md5": lambda ba: hashlib.md5(ba).digest(),
-    "sha1": lambda ba: hashlib.sha1(ba).digest(),
-    "pymmh3_32": pymmh3.hash,
-    "pymmh3_128": pymmh3.hash128,
-}
 
 
 def init_buffer(ba: bytearray) -> bytearray:
@@ -50,8 +39,24 @@ def init_buffer(ba: bytearray) -> bytearray:
     return ba
 
 
+def generate_size(size: int, p: float) -> int:
+    """Generate a random size for a buffer.
+
+    Args:
+        size: The size of the buffer to hash.
+        p: The percentage of the buffer size to vary.
+
+    Returns:
+        The random size of the buffer.
+    """
+    lower = math.ceil(size * (1 - p))
+    upper = math.floor(size * (1 + p))
+
+    return random.randint(lower, upper)
+
+
 def perf_hash(loops: int, f: Callable, size: int) -> float:
-    """Benchmark the mmh3 hash function.
+    """Benchmark a hash function.
 
     Args:
         loops: The number of outer loops to run.
@@ -62,6 +67,9 @@ def perf_hash(loops: int, f: Callable, size: int) -> float:
         The time taken to hash the buffer in fractional seconds.
     """
     # pylint: disable=too-many-locals
+
+    if size <= 0:
+        raise ValueError("size must be greater than 0")
 
     range_it = itertools.repeat(None, loops)
 
@@ -77,7 +85,7 @@ def perf_hash(loops: int, f: Callable, size: int) -> float:
     data6 = bytes(data[6 : size + 6])
     data7 = bytes(data[7 : size + 7])
     data8 = bytes(data[8 : size + 8])
-    data9 = bytes(data[8 : size + 9])
+    data9 = bytes(data[9 : size + 9])
 
     t0 = time.perf_counter()
     for _ in range_it:
@@ -95,6 +103,114 @@ def perf_hash(loops: int, f: Callable, size: int) -> float:
     return time.perf_counter() - t0
 
 
+def perf_hash_random(loops: int, f: Callable, size: int) -> float:
+    """Benchmark a hash function with varying data sizes.
+
+    Args:
+        loops: The number of outer loops to run.
+        f: The hash function to benchmark
+        size: The size of the buffer to hash.
+
+    Returns:
+        The time taken to hash the buffer in fractional seconds.
+    """
+    # pylint: disable=too-many-locals
+
+    if size <= 0:
+        raise ValueError("size must be greater than 0")
+
+    range_it = itertools.repeat(None, loops)
+    random.seed(42)
+    inner_loops = 10
+    extra_size = 255
+
+    data = bytearray(size + extra_size)
+    data = init_buffer(data)
+
[…]
