### 2024-09-15 ac16ab1
Introduce tox.ini for project automation
