+++ .github/workflows/build.yml
@@ -38,7 +38,7 @@ jobs:
         run: |
           pip install --upgrade pip
           pip install .
-          pip install ".[test]"
+          pip install ".[test,type]"
       - name: Test with pytest
         run: python -m pytest
       - name: Test type hints with mypy
