### 2024-09-15 cac73ac
Introduce tox to automate tasks (#78)
* Introduce tox.ini for project automation

* Update documentation
