+++ tox.ini
@@ -1,0 +1,54 @@
+[tox]
+requires =
+    tox>=4
+envlist = lint, type, py{38,39,310,311,312}
+
+[testenv]
+description = run unit tests
+commands_pre =
+    pip install ".[test]"
+commands =
+    pytest
+
+[testenv:lint]
+description = run linters with formatting
+allowlist_externals =
+    find
+    npx
+commands_pre =
+    pip install ".[lint]"
+commands =
+    black .
+    isort .
+    find ./src/mmh3 -name '*.[ch]' -exec clang-format -i {} +
+    npx prettier --write .
+    pylint --recursive=y .
+    npx markdownlint --config .markdown-lint.yml \
+      --ignore-path .gitignore **/*.md
+
+[testenv:type]
+description = run type checks
+commands_pre =
+    pip install ".[type]"
+commands =
+    mypy --strict tests
+
+[testenv:docs]
+description = run documentation build
+allowlist_externals =
+    make
+commands_pre =
+    pip install ".[docs]"
+commands =
+    make -C docs html
+
+[testenv:build-cfiles]
+allowlist_externals =
+    find
+    git
+commands_pre =
+    pip install ".[lint]"
+commands =
+    git submodule update --init
+    python util/refresh.py
+    find ./src/mmh3 -name '*.[ch]' -exec clang-format -i {} +
