### 2024-09-15 5ea7c39
Migrate setup.py/MANIFEST.in to pyproject.toml
