### 2024-09-15 a22d117
Modernize sdist building
