+++ .github/workflows/wheels.yml
@@ -76,8 +76,8 @@ jobs:
       - name: Build sdist
         run: |
           python -m pip install --upgrade pip
-          pip install setuptools
-          python setup.py build sdist
+          pip install setuptools build
+          python -m build --sdist
       - uses: actions/upload-artifact@v4
         with:
           path: dist/*.tar.gz
