### 2024-09-15 8cd3725
Migrate setup.py/MANIFEST.in to toml file (#79)
