+++ pyproject.toml
@@ -1,5 +1,6 @@
 [build-system]
-requires = ["setuptools >= 61.0.0", "wheel"]
+# setuptools >= 74.1.0 required to build C extensions via pyproject.toml
+requires = ["setuptools >= 74.1.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -64,6 +65,11 @@ Repository = "https://github.com/hajimes/mmh3"
 Changelog = "https://github.com/hajimes/mmh3/blob/master/CHANGELOG.md"
 "Bug Tracker" = "https://github.com/hajimes/mmh3/issues"
 
+[tool.setuptools]
+ext-modules = [
+  {name = "mmh3", sources=["./src/mmh3/mmh3module.c", "./src/mmh3/murmurhash3.c"]}
+]
+
 [tool.isort]
 profile = "black"
 src_paths = ["src/mmh3/__init__.pyi", "util", "tests", "benchmark", "docs"]
@@ -82,4 +88,7 @@ ignore-paths = [
 jobs = 0
 # import-error: An error tricky to resolve, especially on super-linter.
 # wrong-import-order: Respect isort's import order.
-disable = ["import-error", "wrong-import-order"]
+disable = [
+  "import-error",
+  "wrong-import-order"
+]
