+++ MANIFEST.in
@@ -1,10 +1,0 @@
-include src/mmh3/*.c
-include src/mmh3/*.h
-include src/mmh3/*.pyi
-include src/mmh3/py.typed
-include README.md
-include LICENSE
-include CHANGELOG.md
-include pyproject.toml
-include setup.py
-include MANIFEST.in
