### 2024-09-16 3d8bd8f
Fix docstring (#80)
