+++ src/mmh3/mmh3module.c
@@ -392,7 +392,7 @@ PyDoc_STRVAR(
     "\n"
     "Return a hash for the buffer as a 32-bit signed integer.\n"
     "\n"
-    "Calculated by the MurmurHash3_x86_32 algorithm. "
+    "Calculated by the MurmurHash3_x86_32 algorithm.\n"
     "\n"
     "Args:\n"
     "    key (Buffer | str): The input buffer to hash.\n"
@@ -430,7 +430,7 @@ PyDoc_STRVAR(
     "\n"
     "Return a hash for the buffer as a 32-bit unsigned integer.\n"
     "\n"
-    "Calculated by the MurmurHash3_x86_32 algorithm. "
+    "Calculated by the MurmurHash3_x86_32 algorithm.\n"
     "\n"
     "Args:\n"
     "    key (Buffer | str): The input buffer to hash.\n"
@@ -512,7 +512,7 @@ PyDoc_STRVAR(
     "\n"
     "Return a hash for the buffer as a 128-bit signed integer.\n"
     "\n"
-    "Calculated by the MurmurHash3_x64_128 algorithm. "
+    "Calculated by the MurmurHash3_x64_128 algorithm.\n"
     "\n"
     "Args:\n"
     "    key (Buffer | str): The input buffer to hash.\n"
@@ -564,7 +564,7 @@ PyDoc_STRVAR(
     "\n"
     "Return a hash for the buffer as a 128-bit unsigned integer.\n"
     "\n"
-    "Calculated by the MurmurHash3_x64_128 algorithm. "
+    "Calculated by the MurmurHash3_x64_128 algorithm.\n"
     "\n"
     "Args:\n"
     "    key (Buffer | str): The input buffer to hash.\n"
