### 2024-09-16 439ecf5
Add data param to hasher __init__
