+++ src/mmh3/mmh3module.c
@@ -360,7 +360,7 @@ PyDoc_STRVAR(
     "    bytes: The hash value as the ``bytes`` type with a length of 4 bytes "
     "(32 bits).\n"
     "\n"
-    ".. versionadded:: 4.2.0\n");
+    ".. versionadded:: 5.0.0\n");
 
 static PyObject *
 mmh3_mmh3_32_digest(PyObject *self, PyObject *args, PyObject *keywds)
@@ -402,7 +402,7 @@ PyDoc_STRVAR(
     "Returns:\n"
     "    int: The hash value as a 32-bit signed integer.\n"
     "\n"
-    ".. versionadded:: 4.2.0\n");
+    ".. versionadded:: 5.0.0\n");
 
 static PyObject *
 mmh3_mmh3_32_sintdigest(PyObject *self, PyObject *args, PyObject *keywds)
@@ -440,7 +440,7 @@ PyDoc_STRVAR(
     "Returns:\n"
     "    int: The hash value as a 32-bit unsigned integer.\n"
     "\n"
-    ".. versionadded:: 4.2.0\n");
+    ".. versionadded:: 5.0.0\n");
 
 static PyObject *
 mmh3_mmh3_32_uintdigest(PyObject *self, PyObject *args, PyObject *keywds)
@@ -479,7 +479,7 @@ PyDoc_STRVAR(
     "    bytes: The hash value as the ``bytes`` type with a length of 16 "
     "bytes (128 bits).\n"
     "\n"
-    ".. versionadded:: 4.2.0\n");
+    ".. versionadded:: 5.0.0\n");
 
 static PyObject *
 mmh3_mmh3_x64_128_digest(PyObject *self, PyObject *args, PyObject *keywds)
@@ -522,7 +522,7 @@ PyDoc_STRVAR(
     "Returns:\n"
     "    int: The hash value as a 128-bit signed integer.\n"
     "\n"
-    ".. versionadded:: 4.2.0\n");
+    ".. versionadded:: 5.0.0\n");
 
 static PyObject *
 mmh3_mmh3_x64_128_sintdigest(PyObject *self, PyObject *args, PyObject *keywds)
@@ -574,7 +574,7 @@ PyDoc_STRVAR(
     "Returns:\n"
     "    int: The hash value as a 128-bit unsigned integer.\n"
     "\n"
-    ".. versionadded:: 4.2.0\n");
+    ".. versionadded:: 5.0.0\n");
 
 static PyObject *
 mmh3_mmh3_x64_128_uintdigest(PyObject *self, PyObject *args, PyObject *keywds)
@@ -627,7 +627,7 @@ PyDoc_STRVAR(
     "    tuple[int, int]: The hash value as a tuple of two 64-bit signed "
     "integers.\n"
     "\n"
-    ".. versionadded:: 4.2.0\n");
+    ".. versionadded:: 5.0.0\n");
 
 static PyObject *
 mmh3_mmh3_x64_128_stupledigest(PyObject *self, PyObject *args,
@@ -669,7 +669,7 @@ PyDoc_STRVAR(
     "    tuple[int, int]: The hash value as a tuple of two 64-bit unsigned "
     "integers.\n"
     "\n"
-    ".. versionadded:: 4.2.0\n");
+    ".. versionadded:: 5.0.0\n");
 
 static PyObject *
 mmh3_mmh3_x64_128_utupledigest(PyObject *self, PyObject *args,
@@ -710,7 +710,7 @@ PyDoc_STRVAR(
     "    bytes: The hash value as the ``bytes`` type with a length of 16 "
     "bytes (128 bits).\n"
     "\n"
-    ".. versionadded:: 4.2.0\n");
+    ".. versionadded:: 5.0.0\n");
 
 static PyObject *
[…]
