+++ pyproject.toml
@@ -31,7 +31,7 @@ classifiers = [
 
 [project.optional-dependencies]
 test = [
-  "pytest == 8.3.2",
+  "pytest == 8.3.3",
   "pytest-sugar == 1.0.0"
 ]
 lint = [
