### 2024-09-16 b109158
Add data param to mmh3_x86_128 __init__
