### 2024-09-16 ce573f9
Add buffer argument to hasher constructors (#83)
