### 2024-09-16 f951e16
Use y* instead of u* for buffer-input funcs
