+++ tests/test_mmh3.py
@@ -234,7 +234,6 @@ def test_hash128() -> None:
 def test_mmh3_32_digest() -> None:
     assert mmh3.mmh3_32_digest(b"") == b"\0\0\0\0"
     assert mmh3.mmh3_32_digest(b"", 0) == b"\0\0\0\0"
-    assert mmh3.mmh3_32_digest(b"", seed=0) == b"\0\0\0\0"
     assert mmh3.mmh3_32_digest(b"\x21\x43\x65\x87", 0) == (0xF55B516B).to_bytes(
         4, "little"
     )
@@ -254,17 +253,21 @@ def test_mmh3_32_digest() -> None:
     assert mmh3.mmh3_32_digest(b"\x00\x00", 0) == (0x30F4C306).to_bytes(4, "little")
     assert mmh3.mmh3_32_digest(b"\x00", 0) == (0x514E28B7).to_bytes(4, "little")
 
-    assert mmh3.mmh3_32_digest("aaaa", 0x9747B28C) == (0x5A97808A).to_bytes(4, "little")
-    assert mmh3.mmh3_32_digest("aaa", 0x9747B28C) == (0x283E0130).to_bytes(4, "little")
-    assert mmh3.mmh3_32_digest("aa", 0x9747B28C) == (0x5D211726).to_bytes(4, "little")
-    assert mmh3.mmh3_32_digest("a", 0x9747B28C) == (0x7FA09EA6).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"aaaa", 0x9747B28C) == (0x5A97808A).to_bytes(
+        4, "little"
+    )
+    assert mmh3.mmh3_32_digest(b"aaa", 0x9747B28C) == (0x283E0130).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"aa", 0x9747B28C) == (0x5D211726).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"a", 0x9747B28C) == (0x7FA09EA6).to_bytes(4, "little")
 
-    assert mmh3.mmh3_32_digest("abcd", 0x9747B28C) == (0xF0478627).to_bytes(4, "little")
-    assert mmh3.mmh3_32_digest("abc", 0x9747B28C) == (0xC84A62DD).to_bytes(4, "little")
-    assert mmh3.mmh3_32_digest("ab", 0x9747B28C) == (0x74875592).to_bytes(4, "little")
-    assert mmh3.mmh3_32_digest("a", 0x9747B28C) == (0x7FA09EA6).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"abcd", 0x9747B28C) == (0xF0478627).to_bytes(
+        4, "little"
+    )
+    assert mmh3.mmh3_32_digest(b"abc", 0x9747B28C) == (0xC84A62DD).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"ab", 0x9747B28C) == (0x74875592).to_bytes(4, "little")
+    assert mmh3.mmh3_32_digest(b"a", 0x9747B28C) == (0x7FA09EA6).to_bytes(4, "little")
 
-    assert mmh3.mmh3_32_digest("Hello, world!", 0x9747B28C) == (0x24884CBA).to_bytes(
+    assert mmh3.mmh3_32_digest(b"Hello, world!", 0x9747B28C) == (0x24884CBA).to_bytes(
         4, "little"
     )
 
@@ -272,17 +275,17 @@ def test_mmh3_32_digest() -> None:
         0xD58063C1
     ).to_bytes(4, "little")
 
[…]
