+++ src/mmh3/mmh3module.c
@@ -37,7 +37,7 @@ typedef unsigned __int64 uint64_t;
 
 PyDoc_STRVAR(
     mmh3_hash_doc,
-    "hash(key[, seed=0, signed=True]) -> int\n"
+    "hash(key, seed=0, signed=True) -> int\n"
     "\n"
     "Return a hash as a 32-bit integer.\n"
     "\n"
@@ -108,7 +108,7 @@ mmh3_hash(PyObject *self, PyObject *args, PyObject *keywds)
 
 PyDoc_STRVAR(
     mmh3_hash_from_buffer_doc,
-    "hash_from_buffer(key[, seed=0, signed=True]) -> int\n"
+    "hash_from_buffer(key, seed=0, signed=True) -> int\n"
     "\n"
     "Return a hash for the buffer as a 32-bit integer.\n"
     "\n"
@@ -116,14 +116,19 @@ PyDoc_STRVAR(
     "memory-views such as numpy arrays.\n"
     "\n"
     "Args:\n"
-    "    key (Buffer | str): The bufer to hash.\n"
+    "    key (Buffer | str): The bufer to hash. String inputs are also\n"
+    "        supported and are automatically converted to `bytes` using\n"
+    "        UTF-8 encoding before hashing.\n"
     "    seed (int): The seed value. Must be an integer in the range [0, "
     "0xFFFFFFFF].\n"
     "    signed (bool): If True, return a signed integer. Otherwise, return "
     "an unsigned integer.\n"
     "\n"
     "Returns:\n"
-    "    int: The hash value as a 32-bit integer.\n");
+    "    int: The hash value as a 32-bit integer.\n"
+    "\n"
+    ".. deprecated:: 5.0.0\n"
+    "    Use ``mmh3_32_sintdigest()`` or ``mmh3_32_uintdigest()`` instead.\n");
 
 static PyObject *
 mmh3_hash_from_buffer(PyObject *self, PyObject *args, PyObject *keywds)
@@ -180,7 +185,7 @@ mmh3_hash_from_buffer(PyObject *self, PyObject *args, PyObject *keywds)
 
 PyDoc_STRVAR(
     mmh3_hash64_doc,
-    "hash64(key[, seed=0, x64arch=True, signed=True]) -> tuple[int, int]\n"
+    "hash64(key, seed=0, x64arch=True, signed=True) -> tuple[int, int]\n"
     "\n"
     "Return a hash as a tuple of two 64-bit integers.\n"
     "\n"
@@ -233,7 +238,7 @@ mmh3_hash64(PyObject *self, PyObject *args, PyObject *keywds)
 
 PyDoc_STRVAR(
     mmh3_hash128_doc,
-    "hash128(key[, seed=0, x64arch=True, signed=False]]) -> int\n"
+    "hash128(key, seed=0, x64arch=True, signed=False) -> int\n"
     "\n"
     "Return a hash as a 128-bit integer.\n\n"
     "Calculated by the MurmurHash3_x{64, 86}_128 algorithm.\n"
@@ -295,7 +300,7 @@ mmh3_hash128(PyObject *self, PyObject *args, PyObject *keywds)
 
 PyDoc_STRVAR(
     mmh3_hash_bytes_doc,
-    "hash_bytes(key[, seed=0, x64arch=True]) -> bytes\n"
+    "hash_bytes(key, seed=0, x64arch=True) -> bytes\n"
     "\n"
     "Return a 16-byte hash of the ``bytes`` type.\n"
     "\n"
@@ -343,16 +348,19 @@ mmh3_hash_bytes(PyObject *self, PyObject *args, PyObject *keywds)
[…]
