+++ tox.ini
@@ -42,7 +42,7 @@ commands_pre =
 commands =
     make -C docs html
 
-[testenv:build-cfiles]
+[testenv:build_cfiles]
 allowlist_externals =
     find
     git

[…]
