+++ src/mmh3/mmh3module.c
@@ -371,7 +371,7 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_32_digest(PyObject *self, PyObject *args, PyObject *keywds)
+mmh3_mmh3_32_digest(PyObject *self, PyObject *args)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
@@ -410,7 +410,7 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_32_sintdigest(PyObject *self, PyObject *args, PyObject *keywds)
+mmh3_mmh3_32_sintdigest(PyObject *self, PyObject *args)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
@@ -445,7 +445,7 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_32_uintdigest(PyObject *self, PyObject *args, PyObject *keywds)
+mmh3_mmh3_32_uintdigest(PyObject *self, PyObject *args)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
@@ -481,7 +481,7 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_x64_128_digest(PyObject *self, PyObject *args, PyObject *keywds)
+mmh3_mmh3_x64_128_digest(PyObject *self, PyObject *args)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
@@ -521,7 +521,7 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_x64_128_sintdigest(PyObject *self, PyObject *args, PyObject *keywds)
+mmh3_mmh3_x64_128_sintdigest(PyObject *self, PyObject *args)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
@@ -570,7 +570,7 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_x64_128_uintdigest(PyObject *self, PyObject *args, PyObject *keywds)
+mmh3_mmh3_x64_128_uintdigest(PyObject *self, PyObject *args)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
@@ -620,8 +620,7 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_x64_128_stupledigest(PyObject *self, PyObject *args,
-                               PyObject *keywds)
+mmh3_mmh3_x64_128_stupledigest(PyObject *self, PyObject *args)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
@@ -659,8 +658,7 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_x64_128_utupledigest(PyObject *self, PyObject *args,
-                               PyObject *keywds)
+mmh3_mmh3_x64_128_utupledigest(PyObject *self, PyObject *args)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
@@ -697,7 +695,7 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
[…]
