### 2024-09-16 56be9ce
Fix casting errors
