### 2024-09-16 5a169b9
Add tox testenv for benchmark and plot
