+++ tox.ini
@@ -12,6 +12,7 @@ commands =
 
 [testenv:lint]
 description = run linters with formatting
+skip_install = true
 allowlist_externals =
     find
     npx
@@ -40,6 +41,7 @@ allowlist_externals =
 commands_pre =
     pip install ".[docs]"
 commands =
+    make -C docs clean
     make -C docs html
 
 [testenv:build_cfiles]
@@ -52,3 +54,17 @@ commands =
     git submodule update --init
     python util/refresh.py
     find ./src/mmh3 -name '*.[ch]' -exec clang-format -i {} +
+
+[testenv:benchmark]
+description = run benchmarks
+commands_pre =
+    pip install ".[benchmark]"
+commands =
+    python benchmark/benchmark.py {posargs}
+
+[testenv:plot]
+description = plot benchmark results
+commands_pre =
+    pip install ".[benchmark,plot]"
+commands =
+    python benchmark/plot_graph.py {posargs}
