### 2024-09-16 c341ca0
Make some functions METH_FASTCALL
