### 2024-09-16 551c477
Make all *digest funcs METH_FASTCALL
