+++ src/mmh3/mmh3module.c
@@ -435,15 +435,16 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_32_sintdigest(PyObject *self, PyObject *args)
+mmh3_mmh3_32_sintdigest(PyObject *self, PyObject *const *args,
+                        Py_ssize_t nargs)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
     int32_t result[1];
 
-    if (!PyArg_ParseTuple(args, "y*|I", &target_buf, &seed)) {
-        return NULL;
-    }
+    MMH3_VALIDATE_ARGS_AND_SET_SEED(nargs, args, seed);
+
+    GET_BUFFER_VIEW_OR_ERROUT(args[0], &target_buf);
 
     murmurhash3_x86_32(target_buf.buf, target_buf.len, seed, result);
     PyBuffer_Release(&target_buf);
@@ -470,15 +471,16 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_32_uintdigest(PyObject *self, PyObject *args)
+mmh3_mmh3_32_uintdigest(PyObject *self, PyObject *const *args,
+                        Py_ssize_t nargs)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
     uint32_t result[1];
 
-    if (!PyArg_ParseTuple(args, "y*|I", &target_buf, &seed)) {
-        return NULL;
-    }
+    MMH3_VALIDATE_ARGS_AND_SET_SEED(nargs, args, seed);
+
+    GET_BUFFER_VIEW_OR_ERROUT(args[0], &target_buf);
 
     murmurhash3_x86_32(target_buf.buf, target_buf.len, seed, result);
     PyBuffer_Release(&target_buf);
@@ -547,15 +549,16 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_x64_128_sintdigest(PyObject *self, PyObject *args)
+mmh3_mmh3_x64_128_sintdigest(PyObject *self, PyObject *const *args,
+                             Py_ssize_t nargs)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
     uint64_t result[2];
 
-    if (!PyArg_ParseTuple(args, "y*|I", &target_buf, &seed)) {
-        return NULL;
-    }
+    MMH3_VALIDATE_ARGS_AND_SET_SEED(nargs, args, seed);
+
+    GET_BUFFER_VIEW_OR_ERROUT(args[0], &target_buf);
 
     murmurhash3_x64_128(target_buf.buf, target_buf.len, seed, result);
     PyBuffer_Release(&target_buf);
@@ -596,15 +599,16 @@ PyDoc_STRVAR(
     ".. versionadded:: 5.0.0\n");
 
 static PyObject *
-mmh3_mmh3_x64_128_uintdigest(PyObject *self, PyObject *args)
+mmh3_mmh3_x64_128_uintdigest(PyObject *self, PyObject *const *args,
+                             Py_ssize_t nargs)
 {
     Py_buffer target_buf;
     uint32_t seed = 0;
     uint64_t result[2];
 
-    if (!PyArg_ParseTuple(args, "y*|I", &target_buf, &seed)) {
-        return NULL;
-    }
+    MMH3_VALIDATE_ARGS_AND_SET_SEED(nargs, args, seed);
+
[…]
