+++ README.md
@@ -135,11 +135,14 @@ complete changelog.
 
 #### Added
 
-- Add `digest` functions that accept a non-immutable buffer as input
-  and process it without internal copying
-  ([#75](https://github.com/hajimes/mmh3/issues/75)).
-- Slightly improve the performance of the `hash_bytes` function.
 - Add support for Python 3.13.
+- Add `digest` functions that support the new buffer protocol
+  ([PEP 688](https://peps.python.org/pep-0688/)) as input
+  ([#75](https://github.com/hajimes/mmh3/pull/75)).
+  These functions are implemented with
+  [METH_FASTCALL](https://docs.python.org/3/c-api/structures.html#c.METH_FASTCALL),
+  offering improved performance over legacy functions.
+- Slightly improve the performance of the `hash_bytes()` function.
 - Add Read the Docs documentation
   ([#54](https://github.com/hajimes/mmh3/issues/54)).
 - (planned: Document benchmark results
@@ -149,16 +152,22 @@ complete changelog.
 
 - Change the format of CHANGELOG.md to conform to the
   [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) standard
-  ([#63](https://github.com/hajimes/mmh3/issues/63)).
+  ([#63](https://github.com/hajimes/mmh3/pull/63)).
 - **Backward-incompatible**: Change the constructors of hasher classes to
-  accept a buffer as the first argument.
+  accept a buffer as the first argument
+  ([#83](https://github.com/hajimes/mmh3/pull/83)).
+
+#### Deprecated
+
+- Deprecate the `hash_from_buffer()` function.
+  Use `mmh3_32_sintdigest()` or `mmh3_32_uintdigest()` as alternatives.
 
 #### Fixed
 
 - Fix a reference leak in the `hash_from_buffer()` function
-  ([#75](https://github.com/hajimes/mmh3/issues/75)).
-- Fix type hints ([#76](https://github.com/hajimes/mmh3/issues/76),
-  [#77](https://github.com/hajimes/mmh3/issues/77)).
+  ([#75](https://github.com/hajimes/mmh3/pull/75)).
+- Fix type hints ([#76](https://github.com/hajimes/mmh3/pull/76),
+  [#77](https://github.com/hajimes/mmh3/pull/77)).
 
 ### [4.1.0] - 2024-01-09
 
@@ -172,7 +181,7 @@ complete changelog.
   ([#50](https://github.com/hajimes/mmh3/issues/50)).
 - Fix incorrect type hints ([#51](https://github.com/hajimes/mmh3/issues/51)).
 - Fix invalid results on s390x when the arg `x64arch` of `hash64` or
-  `hash_bytes` is set to `False`
+  `hash_bytes()` is set to `False`
   ([#52](https://github.com/hajimes/mmh3/issues/52)).
 
 ## License
