### 2024-09-17 2ad4331
Add value range check for seed
