+++ tox.ini
@@ -8,7 +8,7 @@ description = run unit tests
 commands_pre =
     pip install ".[test]"
 commands =
-    pytest
+    pytest {posargs}
 
 [testenv:lint]
 description = run linters with formatting
