### 2024-09-17 5a38ca4
Fix mypy errors
