+++ tests/test_invalid_inputs.py
@@ -1,9 +1,12 @@
 # pylint: disable=missing-module-docstring, missing-function-docstring
 # pylint: disable=no-value-for-parameter, too-many-function-args
+from typing import no_type_check
+
 import mmh3
 import pytest
 
 
+@no_type_check
 def test_mmh3_32_digest_raises_typeerror() -> None:
     with pytest.raises(TypeError):
         mmh3.mmh3_32_digest()
@@ -17,6 +20,7 @@ def test_mmh3_32_digest_raises_typeerror() -> None:
         mmh3.mmh3_32_digest([1, 2, 3], 42)
 
 
+@no_type_check
 def test_mmh3_32_digest_raises_valueerror() -> None:
     with pytest.raises(ValueError):
         mmh3.mmh3_32_digest(b"hello, world", -1)
