+++ tox.ini
@@ -30,7 +30,7 @@ commands =
 [testenv:type]
 description = run type checks
 commands_pre =
-    pip install ".[type]"
+    pip install ".[test,type]"
 commands =
     mypy --strict tests
 
