### 2024-09-17 7106202
Add seed range check for all funcs
