### 2024-09-17 fc3ed7d
Fix return type warning
