### 2024-09-17 cd68416
Fix hasher init; add more unit tests
