+++ tests/test_invalid_inputs.py
@@ -29,6 +29,94 @@ def test_hash_raises_valueerror() -> None:
 
 
 @no_type_check
+def test_hash128_raises_typeerror() -> None:
+    with pytest.raises(TypeError):
+        mmh3.hash128()
+    with pytest.raises(TypeError):
+        mmh3.hash128(b"hello, world", "42")
+    with pytest.raises(TypeError):
+        mmh3.hash128(b"hello, world", 42, True, False, 1234)
+    with pytest.raises(TypeError):
+        mmh3.hash128(b"hello, world", seed="42")
+    with pytest.raises(TypeError):
+        mmh3.hash128([1, 2, 3], 42)
+
+
+@no_type_check
+def test_hash128_raises_valueerror() -> None:
+    with pytest.raises(ValueError):
+        mmh3.hash128(b"hello, world", -1)
+    with pytest.raises(ValueError):
+        mmh3.hash128(b"hello, world", 2**32)
+
+
+@no_type_check
+def test_hash64_raises_typeerror() -> None:
+    with pytest.raises(TypeError):
+        mmh3.hash64()
+    with pytest.raises(TypeError):
+        mmh3.hash64(b"hello, world", "42")
+    with pytest.raises(TypeError):
+        mmh3.hash64(b"hello, world", 42, True, False, 1234)
+    with pytest.raises(TypeError):
+        mmh3.hash64(b"hello, world", seed="42")
+    with pytest.raises(TypeError):
+        mmh3.hash64([1, 2, 3], 42)
+
+
+@no_type_check
+def test_hash64_raises_valueerror() -> None:
+    with pytest.raises(ValueError):
+        mmh3.hash64(b"hello, world", -1)
+    with pytest.raises(ValueError):
+        mmh3.hash64(b"hello, world", 2**32)
+
+
+@no_type_check
+def test_hash_bytes_raises_typeerror() -> None:
+    with pytest.raises(TypeError):
+        mmh3.hash_bytes()
+    with pytest.raises(TypeError):
+        mmh3.hash_bytes(b"hello, world", "42")
+    with pytest.raises(TypeError):
+        mmh3.hash_bytes(b"hello, world", 42, True, 1234)
+    with pytest.raises(TypeError):
+        mmh3.hash_bytes(b"hello, world", seed="42")
+    with pytest.raises(TypeError):
+        mmh3.hash_bytes([1, 2, 3], 42)
+
+
+@no_type_check
+def test_hash_bytes_raises_valueerror() -> None:
+    with pytest.raises(ValueError):
+        mmh3.hash_bytes(b"hello, world", -1)
+    with pytest.raises(ValueError):
+        mmh3.hash_bytes(b"hello, world", 2**32)
+
+
+@no_type_check
+def test_hash_from_buffer_raises_typeerror() -> None:
+    with pytest.raises(TypeError):
+        mmh3.hash_from_buffer()
+    with pytest.raises(TypeError):
+        mmh3.hash_from_buffer(b"hello, world", "42")
+    with pytest.raises(TypeError):
+        mmh3.hash_from_buffer(b"hello, world", 42, True, 1234)
+    with pytest.raises(TypeError):
+        mmh3.hash_from_buffer(b"hello, world", seed="42")
+    with pytest.raises(TypeError):
+        mmh3.hash_from_buffer([1, 2, 3], 42)
+
+
+@no_type_check
+def test_hash_from_buffer_raises_valueerror() -> None:
+    with pytest.raises(ValueError):
+        mmh3.hash_from_buffer(b"hello, world", -1)
[…]
