### 2024-09-17 b8dbc59
Improve docs and docstrings
