+++ src/mmh3/mmh3module.c
@@ -91,10 +91,10 @@ PyDoc_STRVAR(
     "\n"
     "Args:\n"
     "    key (bytes | str): The input data to hash.\n"
-    "    seed (int): The seed value. Must be an integer in the range [0, "
-    "0xFFFFFFFF].\n"
-    "    signed (Any): If True, return a signed integer. Otherwise, return "
-    "an unsigned integer.\n"
+    "    seed (int): The seed value. Must be an integer in the range\n"
+    "        [0, 0xFFFFFFFF].\n"
+    "    signed (Any): If True, return a signed integer. Otherwise, return\n"
+    "        an unsigned integer.\n"
     "\n"
     "Returns:\n"
     "    int: The hash value as a 32-bit integer.\n"
@@ -171,10 +171,10 @@ PyDoc_STRVAR(
     "    key (Buffer | str): The bufer to hash. String inputs are also\n"
     "        supported and are automatically converted to `bytes` using\n"
     "        UTF-8 encoding before hashing.\n"
-    "    seed (int): The seed value. Must be an integer in the range [0, "
-    "0xFFFFFFFF].\n"
-    "    signed (Any): If True, return a signed integer. Otherwise, return "
-    "an unsigned integer.\n"
+    "    seed (int): The seed value. Must be an integer in the range\n"
+    "        [0, 0xFFFFFFFF].\n"
+    "    signed (Any): If True, return a signed integer. Otherwise, return\n"
+    "        an unsigned integer.\n"
     "\n"
     "Returns:\n"
     "    int: The hash value as a 32-bit integer.\n"
@@ -251,12 +251,13 @@ PyDoc_STRVAR(
     "\n"
     "Args:\n"
     "    key (bytes | str): The input data to hash.\n"
-    "    seed (int): The seed value. Must be an integer in the range [0, "
-    "0xFFFFFFFF].\n"
-    "    x64arch (Any): If True, use an algorithm optimized for 64-bit "
-    "architecture. Otherwise, use one optimized for 32-bit architecture.\n"
-    "    signed (Any): If True, return a signed integer. Otherwise, return "
-    "an unsigned integer.\n"
+    "    seed (int): The seed value. Must be an integer in the range\n"
+    "        [0, 0xFFFFFFFF].\n"
+    "    x64arch (Any): If True, use an algorithm optimized for 64-bit\n"
+    "        architecture. Otherwise, use one optimized for 32-bit\n"
+    "        architecture.\n"
+    "    signed (Any): If True, return a signed integer. Otherwise, return\n"
+    "        an unsigned integer.\n"
     "\n"
     "Returns:\n"
     "    tuple[int, int]: The hash value as a tuple of two 64-bit "
@@ -311,12 +312,13 @@ PyDoc_STRVAR(
     "\n"
     "Args:\n"
     "    key (bytes | str): The input data to hash.\n"
-    "    seed (int): The seed value. Must be an integer in the range [0, "
-    "0xFFFFFFFF].\n"
-    "    x64arch (Any): If True, use an algorithm optimized for 64-bit "
-    "architecture. Otherwise, use one optimized for 32-bit architecture.\n"
-    "    signed (Any): If True, return a signed integer. Otherwise, return "
-    "an unsigned integer.\n"
+    "    seed (int): The seed value. Must be an integer in the range\n"
+    "        [0, 0xFFFFFFFF].\n"
[…]
