### 2024-09-17 30da46e
Improve buffer-accepting hashes and more (#84)
