+++ tests/test_mmh3.py
@@ -13,11 +13,9 @@ def test_hash() -> None:
     # https://stackoverflow.com/a/31929528
     assert mmh3.hash(b"", seed=0) == 0
     assert mmh3.hash(b"", seed=1) == 0x514E28B7
-    assert mmh3.hash(b"", seed=u32_to_s32(0xFFFFFFFF)) == u32_to_s32(0x81F16F39)
+    assert mmh3.hash(b"", seed=0xFFFFFFFF) == u32_to_s32(0x81F16F39)
     assert mmh3.hash(b"\x21\x43\x65\x87", 0) == u32_to_s32(0xF55B516B)
-    assert mmh3.hash(b"\x21\x43\x65\x87", u32_to_s32(0x5082EDEE)) == u32_to_s32(
-        0x2362F9DE
-    )
+    assert mmh3.hash(b"\x21\x43\x65\x87", 0x5082EDEE) == u32_to_s32(0x2362F9DE)
     assert mmh3.hash(b"\x21\x43\x65", 0) == u32_to_s32(0x7E4A8634)
     assert mmh3.hash(b"\x21\x43", 0) == u32_to_s32(0xA0F7B07A)
     assert mmh3.hash(b"\x21", 0) == u32_to_s32(0x72661CF4)
@@ -27,23 +25,21 @@ def test_hash() -> None:
     assert mmh3.hash(b"\x00\x00", 0) == u32_to_s32(0x30F4C306)
     assert mmh3.hash(b"\x00", 0) == u32_to_s32(0x514E28B7)
 
-    assert mmh3.hash("aaaa", u32_to_s32(0x9747B28C)) == u32_to_s32(0x5A97808A)
-    assert mmh3.hash("aaa", u32_to_s32(0x9747B28C)) == u32_to_s32(0x283E0130)
-    assert mmh3.hash("aa", u32_to_s32(0x9747B28C)) == u32_to_s32(0x5D211726)
-    assert mmh3.hash("a", u32_to_s32(0x9747B28C)) == u32_to_s32(0x7FA09EA6)
+    assert mmh3.hash("aaaa", 0x9747B28C) == u32_to_s32(0x5A97808A)
+    assert mmh3.hash("aaa", 0x9747B28C) == u32_to_s32(0x283E0130)
+    assert mmh3.hash("aa", 0x9747B28C) == u32_to_s32(0x5D211726)
+    assert mmh3.hash("a", 0x9747B28C) == u32_to_s32(0x7FA09EA6)
 
-    assert mmh3.hash("abcd", u32_to_s32(0x9747B28C)) == u32_to_s32(0xF0478627)
-    assert mmh3.hash("abc", u32_to_s32(0x9747B28C)) == u32_to_s32(0xC84A62DD)
-    assert mmh3.hash("ab", u32_to_s32(0x9747B28C)) == u32_to_s32(0x74875592)
-    assert mmh3.hash("a", u32_to_s32(0x9747B28C)) == u32_to_s32(0x7FA09EA6)
+    assert mmh3.hash("abcd", 0x9747B28C) == u32_to_s32(0xF0478627)
+    assert mmh3.hash("abc", 0x9747B28C) == u32_to_s32(0xC84A62DD)
+    assert mmh3.hash("ab", 0x9747B28C) == u32_to_s32(0x74875592)
+    assert mmh3.hash("a", 0x9747B28C) == u32_to_s32(0x7FA09EA6)
 
-    assert mmh3.hash("Hello, world!", u32_to_s32(0x9747B28C)) == u32_to_s32(0x24884CBA)
+    assert mmh3.hash("Hello, world!", 0x9747B28C) == u32_to_s32(0x24884CBA)
 
[…]
