+++ src/mmh3/mmh3module.c
@@ -32,12 +32,58 @@ typedef unsigned __int64 uint64_t;
 #define MMH3_32_BLOCKSIZE 12
 #define MMH3_128_BLOCKSIZE 32
 
+#define MMH3_VALIDATE_SEED_RETURN_NULL(seed)                       \
+    if (seed < 0 || seed > 0xFFFFFFFF) {                           \
+        PyErr_SetString(PyExc_ValueError, "seed is out of range"); \
+        return NULL;                                               \
+    }
+
+#define MMH3_VALIDATE_SEED_RETURN_INT(seed, buf)                   \
+    if (seed < 0 || seed > 0xFFFFFFFF) {                           \
+        PyBuffer_Release(&buf);                                    \
+        PyErr_SetString(PyExc_ValueError, "seed is out of range"); \
+        return -1;                                                 \
+    }
+
+#define MMH3_VALIDATE_ARGS_AND_SET_SEED(nargs, args, seed)                  \
+    if (nargs < 1) {                                                        \
+        PyErr_SetString(PyExc_TypeError,                                    \
+                        "function takes at least 1 argument (0 given)");    \
+        return NULL;                                                        \
+    }                                                                       \
+    if (nargs > 2) {                                                        \
+        PyErr_Format(PyExc_TypeError,                                       \
+                     "function takes at most 2 arguments (%d given)",       \
+                     (int)nargs);                                           \
+        return NULL;                                                        \
+    }                                                                       \
+    if (nargs == 2) {                                                       \
+        if (!PyLong_Check(args[1])) {                                       \
+            PyErr_Format(PyExc_TypeError,                                   \
+                         "'%s' object cannot be interpreted as an integer", \
+                         Py_TYPE(args[1])->tp_name);                        \
+            return NULL;                                                    \
+        }                                                                   \
+        const unsigned long seed_tmp = PyLong_AsUnsignedLong(args[1]);      \
+        if (seed_tmp == (unsigned long)-1 && PyErr_Occurred()) {            \
+            if (PyErr_ExceptionMatches(PyExc_OverflowError)) {              \
+                PyErr_SetString(PyExc_ValueError, "seed is out of range");  \
+                return NULL;                                                \
+            }                                                               \
+        }                                                                   \
+        if (seed_tmp > 0xFFFFFFFF) {                                        \
+            PyErr_SetString(PyExc_ValueError, "seed is out of range");      \
+            return NULL;                                                    \
+        }                                                                   \
+        seed = (uint32_t)seed_tmp;                                          \
+    }
[…]
