### 2024-09-17 d99b6ed
Add tag check to main upload workflow
