+++ docs/CONTRIBUTORS.md
@@ -1,0 +1,65 @@
+# Contributors
+
+This page acknowledges contributors to the project. For details on the
+project's history and changes, please refer to the
+[Changelog](./changelog.md)
+page. If you're interested in contributing, be sure to review the
+[Contributing](./CONTRIBUTING.md) guide.
+
+## Code Contributors
+
+We gratefully acknowledge the contributions of the following individuals:
+
+- [Alexander Maznev](https://github.com/pik),
+  [#6](https://github.com/hajimes/mmh3/pull/6).
+- [@arieleizenberg](https://github.com/arieleizenberg),
+  [#34](https://github.com/hajimes/mmh3/pull/34).
+- [Micha Gorelick](https://github.com/mynameisfiber),
+  [#1](https://github.com/hajimes/mmh3/pull/1).
+- [Danil Shein](https://github.com/dshein-alt),
+  [#40](https://github.com/hajimes/mmh3/pull/40).
+- [Derek Wilson](https://github.com/underrun),
+  [#2](https://github.com/hajimes/mmh3/pull/2),
+  [#3](https://github.com/hajimes/mmh3/pull/3).
+- [Dimitri Vorona](https://github.com/alendit),
+  [#13](https://github.com/hajimes/mmh3/pull/13).
+- [@doozr](https://github.com/doozr),
+  [#15](https://github.com/hajimes/mmh3/pull/15).
+- [Dušan Nikolić](https://github.com/n-dusan),
+  [#37](https://github.com/hajimes/mmh3/pull/37).
+- [Matthew Honnibal](https://github.com/honnibal),
+  [#22](https://github.com/hajimes/mmh3/pull/22).
+- [wouter bolsterlee](https://github.com/wbolster),
+  [#35](https://github.com/hajimes/mmh3/pull/35).
+
+## Community Contributors
+
+We would also like to thank the following contributors for their valuable
+bug reports, feature suggestions, and other contributions:
+
+- [Antoine Pitrou](https://github.com/pitrou),
+  [#10](https://github.com/hajimes/mmh3/issues/10).
+- [Benjamin Bengfort](https://github.com/bbengfort),
+  [#46](https://github.com/hajimes/mmh3/issues/46).
+- [Christian von Schultz](https://github.com/vonschultz),
+  [#50](https://github.com/hajimes/mmh3/issues/50).
+- [Dan Blanchard](https://github.com/dan-blanchard),
+  [#8](https://github.com/hajimes/mmh3/issues/8).
+- [Heather Lapointe](https://github.com/Alphadelta14),
+  [#25](https://github.com/hajimes/mmh3/issues/25).
+- [Jacques Dark](https://github.com/jqdark),
+  [#12](https://github.com/hajimes/mmh3/issues/12).
+- [Niklas Semmler](https://github.com/niklassemmler),
+  [#7](https://github.com/hajimes/mmh3/issues/7).
+- [Ryan](https://github.com/ryanfwy),
+  [#25](https://github.com/hajimes/mmh3/issues/25).
+- [Sebastian Kreft](https://github.com/sk-),
+  [#17](https://github.com/hajimes/mmh3/issues/17).
+- [Tom Mitchell](https://github.com/tcmitchell),
+  [#51](https://github.com/hajimes/mmh3/issues/51).
+- [Varunkumar Nagarajan](https://github.com/varunkumar),
+  [#39](https://github.com/hajimes/mmh3/issues/39).
+- [@xqdd](https://github.com/xqdd),
+  [#9](https://github.com/hajimes/mmh3/issues/9).
+- [@yzssbo](https://github.com/yzssbo),
+  [#25](https://github.com/hajimes/mmh3/issues/25).
