### 2024-09-17 fb065c4
Add experimental tuning to mmh3_32
