### 2024-09-17 41b2e39
Enable autovectorize
