### 2024-09-17 6a5c513
Improve mmh3.hash
