+++ src/mmh3/mmh3module.c
@@ -45,6 +45,59 @@ typedef unsigned __int64 uint64_t;
         return -1;                                                 \
     }
 
+// obj: PyObject*
+// target_str: const char *
+// len: Py_ssize_t
+#define MMH3_HASH_VALIDATE_AND_SET_BYTES(obj, target_str, len)          \
+    if (PyBytes_Check(obj)) {                                           \
+        target_str_len = PyBytes_Size(obj);                             \
+        target_str = PyBytes_AS_STRING(obj);                            \
+    }                                                                   \
+    else if (PyUnicode_Check(obj)) {                                    \
+        target_str_len = PyUnicode_GET_LENGTH(obj);                     \
+        target_str = PyUnicode_AsUTF8AndSize(obj, &target_str_len);     \
+    }                                                                   \
+    else {                                                              \
+        PyErr_Format(PyExc_TypeError,                                   \
+                     "argument 1 must be read-only bytes-like object, " \
+                     "not '%s'",                                        \
+                     Py_TYPE(obj)->tp_name);                            \
+        return NULL;                                                    \
+    }
+
+// obj: PyObject*
+// seed: unsigned long
+#define MMH3_HASH_VALIDATE_AND_SET_SEED(obj, seed)                      \
+    if (!PyLong_Check(obj)) {                                           \
+        PyErr_Format(PyExc_TypeError,                                   \
+                     "'%s' object cannot be interpreted as an integer", \
+                     Py_TYPE(obj)->tp_name);                            \
+        return NULL;                                                    \
+    }                                                                   \
+    seed = PyLong_AsUnsignedLong(obj);                                  \
+    if (seed == (unsigned long)-1 && PyErr_Occurred()) {                \
+        if (PyErr_ExceptionMatches(PyExc_OverflowError)) {              \
+            PyErr_SetString(PyExc_ValueError, "seed is out of range");  \
+            return NULL;                                                \
+        }                                                               \
+    }                                                                   \
+    if (seed > 0xFFFFFFFF) {                                            \
+        PyErr_SetString(PyExc_ValueError, "seed is out of range");      \
+        return NULL;                                                    \
+    }
+
+// nargs: Py_ssize_t
+// name: const char *
+// pos: int
+#define MMH3_HASH_VALIDATE_ARG_DUPLICATION(nargs, name, pos) \
+    if (nargs >= pos) {                                      \
+        PyErr_Format(PyExc_TypeError,                        \
+                     "argument for function given by name "  \
+                     "('%s') and position (%d)",             \
+                     name, pos);                             \
+        return NULL;                                         \
+    }
+
 #define MMH3_VALIDATE_ARGS_AND_SET_SEED(nargs, args, seed)                  \
     if (nargs < 1) {                                                        \
         PyErr_SetString(PyExc_TypeError,                                    \
@@ -105,16 +158,15 @@ PyDoc_STRVAR(
     "    ``bool`` to ``Any``.\n");
 
 static PyObject *
-mmh3_hash(PyObject *self, PyObject *args, PyObject *keywds)
[…]
