### 2024-09-17 4381e78
Add benchmark workflow for base hash
