+++ benchmark/benchmark.py
@@ -223,9 +223,20 @@ def add_cmdline_args(cmd: list, args) -> None:
     cmd.extend(("--test-buffer-size-max", str(args.test_buffer_size_max)))
 
 
+# "if hasattr" is used to check for the existence of the function in the
+# module, to compare the performance of the current implementation with the
+# old one (version 4.1.0), which does not implement the new functions.
+# These conditions should be removed in the future.
 HASHES = {
-    "mmh3_32": mmh3.mmh3_32_digest,
-    "mmh3_128": mmh3.mmh3_x64_128_digest,
+    "mmh3_base_hash": mmh3.hash,
+    "mmh3_32": (
+        mmh3.mmh3_32_digest if hasattr(mmh3, "mmh3_32_digest") else mmh3.hash_bytes
+    ),
+    "mmh3_128": (
+        mmh3.mmh3_x64_128_digest
+        if hasattr(mmh3, "mmh3_x64_128_digest")
+        else mmh3.hash128
+    ),
     "xxh_32": xxhash.xxh32_digest,
     "xxh_64": xxhash.xxh64_digest,
     "xxh3_64": xxhash.xxh3_64_digest,
@@ -257,7 +268,7 @@ if __name__ == "__main__":
     runner.argparser.add_argument(
         "--test-type",
         type=str,
-        help="Type of benchmarking to perform",
+        help="Type of benchmarking to perform (experimental)",
         choices=BENCHMARKING_TYPES.keys(),
         default="random",
     )
