+++ .github/workflows/benchmark-base-hash.yml
@@ -2,7 +2,10 @@
 name: Benchmark Base Hash
 
 on:
-  workflow_dispatch:
+  push:
+    branches:
+      - master
+      - feature/**
 
 permissions: {}
 
