### 2024-09-17 6bb9987
Improve basic hash by METH_FASTCALL (#87)
