+++ docs/api.md
@@ -1,15 +1,40 @@
+<!-- markdownlint-disable MD051 -->
+
 # API Reference
 
-```{caution}
-  This reference contains functions and APIs that are not yet released in the
-  stable version (4.1.0).
-```
+The MurmurHash3 algorithm has three variants:
 
-## Functions for immutables
+- MurmurHash3_x86_32: Generates 32-bit hashes using 32-bit arithmetic.
+- MurmurHash3_x64_128: Generates 128-bit hashes using 64-bit arithmetic.
+- MurmurHash3_x86_128: Generates 128-bit hashes using 32-bit arithmetic.
+
+The `mmh3` library provides functions and classes for each variant.
+
+Although this API reference is comprehensive, you may find the following
+functions particularly useful:
+
+- [mmh3.hash()](#mmh3.hash): Uses the 32-bit variant as its backend and accepts
+  `bytes` or `str` as input (strings are UTF-8 encoded). This
+  function is slower than the x64_128 variant in 64-bit environments but is
+  portable across different architectures. It can also be used to calculate
+  favicon hash footprints for platforms like
+  [Shodan](https://www.shodan.io) and [ZoomEye](https://www.zoomeye.hk).
+- [mmh3.mmh3_x64_128_digest()](#mmh3.mmh3_x64_128_digest): Uses the x64_128
+  variant as its backend. This function accepts a buffer (e.g., `bytes`,
+  `bytearray`, `memoryview`, and `numpy` arrays) and returns a 128-bit hash as
+  a `bytes` object, similar to the `hashlib` module in the Python Standard
+  Library. It performs faster than the 32-bit variant on 64-bit machines.
+
+## Basic Hash Functions
 
 The following functions are used to hash immutable types, specifically
 `bytes` and `str`. String inputs are automatically converted to `bytes` using
 UTF-8 encoding before hashing.
+
+Although `hash128()`, `hash64()`, and `mmh3.hash_bytes()` are provided for
+compatibility with previous versions and are not marked for deprecation,
+the [buffer-accepting hashe functions](#buffer-accepting-hash-functions)
+introduced in version 5.0.0 are recommended for new code.
 
 ```{eval-rst}
 .. autofunction:: mmh3.hash
@@ -18,7 +43,7 @@ UTF-8 encoding before hashing.
 .. autofunction:: mmh3.hash_bytes
 ```
 
-## Functions for Buffer
+## Buffer-Accepting Hash Functions
 
 The following functions are used to hash types that implement the buffer
 protocol such as `bytes`, `bytearray`, `memoryview`, and `numpy` arrays.
@@ -56,6 +81,36 @@ the backported type hint for `collections.abc.Buffer`.
 ```
 
 ## Hasher Classes
+
+`mmh3` implements hashers with interfaces similar to those in `hashlib` from
+the standard library: `mmh3_32()` for 32-bit hashing, `mmh3_x64_128()` for
+128-bit hashing optimized for x64 architectures, and `mmh3_x86_128()` for
+128-bit hashing optimized for x86 architectures.
+
+In addition to the standard `digest()` method, each hasher provides
+`sintdigest()`, which returns a signed integer, and `uintdigest()`, which
+returns an unsigned integer. The 128-bit hashers also include `stupledigest()`
+and `utupledigest()`, which return two 64 bit integers.
+
+Please note that as of version 5.0.0, the implementation is still experimental,
+and performance may be unsatisfactory (particularly `mmh3_x86_128()`).
[…]
