+++ docs/benchmark.md
@@ -2,46 +2,186 @@
 
 ## Settings
 
-- Python environment
-  - CPython 3.12.5 (64-bit)
+### Machine
+
 - Ubuntu 22.04 instance on GitHub Actions
-- Hash libraries
+  - The bencharmking suits are implemented as GitHub Actions workflows.
+  - [4 Processors, 16 GB RAM, 14 GB storage (SSD)](https://docs.github.com/en/actions/using-github-hosted-runners/using-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories)
+  - According to profiling with `pyperf`, each processor operates at a
+    frequency between 2.4 and 3.3 GHz.
+- Tuning by the folloiwing settings:
+  - All tests in a benchmarking suite are executed within the same GitHub
+    Actions job. For more details, refer to
+    [Rodríguez-Guerra (2021)](https://labs.quansight.org/blog/2021/08/github-actions-benchmarks).
+  - [CPU pinning](https://manuel.bernhardt.io/posts/2023-11-16-core-pinning/)
+    to isolate the benchmarking process.
+  - See the
+    [documentation of pyperf](https://pyperf.readthedocs.io/en/latest/system.html)
+    for more details on the following settings:
+    - Stop `irqbalance`.
+    - Set `/proc/irq/default_smp_affinity` to `3` (CPU 0 and 1), where
+      the benrchmarking processes are pinned to CPU 2 and 3.
+    - Set `/proc/sys/kernel/perf_event_max_sample_rate` to `1`.
+    - `/proc/sys/kernel/randomize_va_space` = 2 (default)
+
+### Software
+
+- Python environment:
+  - CPython 3.12.5 (64-bit)
+- Hash libraries:
   - mmh3 5.0.0-dev
-  - xxhash 3.5.0
-- pyperf 2.7.0
+  - [python-xxhash](https://github.com/ifduyue/python-xxhash) 3.5.0
+  - [hashlib](https://docs.python.org/3/library/hashlib.html) (Standard library)
+    - `md5` is tested for `lambda x: hashlib.md5(x).digest()`, and so is `sha1`.
+      therefore, the results for these functions include
+      the overhead of creating the hash object and a function call.
+- Benchmarking library:
+  - [pyperf](https://github.com/psf/pyperf) 2.7.0
+    - Used the
+      [bench_time_func](https://pyperf.readthedocs.io/en/latest/api.html#Runner.bench_time_func)
+      interface to eliminate the overhead of the function call.
+    - Processed time are measured by
+      [time.perf_counter()](https://docs.python.org/3/library/time.html#time.perf_counter)
+      in nanoseconds.
+
+## Method
+
+- A benchmarking test is performed for each specified byte size, which is
+  derived from the Fibonacci sequence.
+- For each input size, the test generates a set of 10 `bytes` instances, where
+  each instance's size is pseudo-randomly selected from the range
+  [ceil(input * 0.9), floor(input * 1.1)].
+  - This randomization is crucial as it increases the difficulty of branch
+    predictions, creating a more realistic scenario. For further details, see
+    [xxHash: Performance comparison](https://github.com/Cyan4973/xxHash/wiki/Performance-comparison#throughput-on-small-data-of-random-length-1-n).
+- This inner loop of 10 iterations is repeated for a certain number of cycles,
+  referred to as the outer loop, which is auto-calibrated by `pyperf`.
+- To avoid the overhead during the loop, iterators are pre-generated
+  using `itertools.repeat()` outside the loop. See
+  [Peters (2002)](https://www.oreilly.com/library/view/python-cookbook/0596001673/ch17.html)
[…]
