### 2024-09-18 ac005e5
Update documentation to prepare next release (#89)
