+++ paper/paper.bib
@@ -16,6 +16,12 @@
 	year         = 2011,
 	url          = {https://github.com/aappleby/smhasher}
 }
+@misc{Bernhardt2023,
+	title        = {On pinning and isolating CPU cores},
+	author       = {Bernhardt, Manuel},
+	year         = 2023,
+	url          = {https://manuel.bernhardt.io/posts/2023-11-16-core-pinning/}
+}
 @article{Bloom1970,
 	title        = {Space/Time Trade-Offs in Hash Coding with Allowable Errors},
 	author       = {Bloom, Burton H.},
@@ -130,6 +136,25 @@
 	publisher    = {Manning},
 	isbn         = {978-1-61729-803-5}
 }
+@incollection{Peters2002,
+  author        = {Peters, Tim},
+  title         = {Algorithms: Introduction},
+  chapter       = {17},
+  year          = 2002,
+  month		    = {jul},
+  booktitle     = {Python Cookbook},
+  publisher     = {O'Reilly Media},
+  editor        = {Martelli, Alex and Ascher, David},
+  edition       = {1st edition}
+}
+@techreport{RodriguezGuerra2021,
+	title        = {Is GitHub Actions suitable for running benchmarks?},
+	author       = {Rodríguez-Guerra, Jaime},
+	year         = 2021,
+	month        = {aug},
+	url          = {https://labs.quansight.org/blog/2021/08/github-actions-benchmarks},
+	institution  = {Quansight Labs}
+}
 @inproceedings{Senuma2011,
 	title        = {K-means {Clustering} with {Feature} {Hashing}},
 	author       = {Senuma, Hajime},
@@ -168,6 +193,21 @@
 	month        = oct,
 	journal      = {Twitter},
 	url          = {https://twitter.com/shodanhq/status/1314023874128998400}
+}
+@misc{Stinner2016,
+	title        = {My journey to stable benchmark},
+	author       = {Stinner, Victor},
+	year         = 2016,
+	url          = {https://vstinner.github.io/journey-to-stable-benchmark-system.html}
+}
+@inproceedings{Tang2024,
+	title        = {Data Splitting based Double Layer Encryption for
+Secure Ciphertext Deduplication in Cloud Storage},
+	author       = {Tang, Xin and Jin, Luchao},
+	year         = 2024,
+	journal      = {2024 IEEE 17th International Conference on Cloud Computing (CLOUD)},
+	pages        = {153--163}
+	doi          = {10.1109/CLOUD62652.2024.00027},
 }
 @inproceedings{Weinberger2009,
 	title        = {Feature {Hashing} for {Large} {Scale} {Multitask} {Learning}},
