### 2024-09-20 e7ac2db
Fix bibtex
