+++ paper/paper.bib
@@ -206,8 +206,8 @@ Secure Ciphertext Deduplication in Cloud Storage},
 	author       = {Tang, Xin and Jin, Luchao},
 	year         = 2024,
 	journal      = {2024 IEEE 17th International Conference on Cloud Computing (CLOUD)},
-	pages        = {153--163}
-	doi          = {10.1109/CLOUD62652.2024.00027},
+	pages        = {153--163},
+	doi          = {10.1109/CLOUD62652.2024.00027}
 }
 @inproceedings{Weinberger2009,
 	title        = {Feature {Hashing} for {Large} {Scale} {Multitask} {Learning}},
