+++ paper/paper.bib
@@ -53,10 +53,10 @@
 	note         = {ISSN: 0818681322},
 	keywords     = {MinHash}
 }
-@misc{collet_xxhash_2012,
+@misc{collet_xxhash_2014,
 	title        = {{xxHash}},
 	author       = {Collet, Yan},
-	year         = 2012,
+	year         = 2014,
 	url          = {https://github.com/Cyan4973/xxHash}
 }
 @misc{du_xxhash_2014,
