### 2024-09-20 fa40da5
Revise IoT section
