+++ paper/paper.bib
@@ -59,6 +59,12 @@
 	year         = 2014,
 	url          = {https://github.com/Cyan4973/xxHash}
 }
+@misc{collet_xxhash_comparison_2024,
+	title        = {{xxHash: Performance comparison (2024)}},
+	author       = {Collet, Yan},
+	year         = 2024,
+	url          = {https://github.com/Cyan4973/xxHash/wiki/Performance-comparison}
+}
 @misc{du_xxhash_2014,
 	title        = {{xxhash}},
 	author       = {Du, Yue},
@@ -136,6 +142,21 @@
 	publisher    = {Manning},
 	isbn         = {978-1-61729-803-5}
 }
+@techreport{Matherly2017,
+	title        = {Complete Guide to Shodan: Collect. Analyze. Visualize. Make Internet Intelligence Work for You.},
+	author       = {Matherly, John},
+	year         = 2017,
+	edition	     = {Version 2017-08-23},
+	institution  = {Shodan}
+}
+@misc{Matherly2024,
+	title        = {Deep Dive: http.favicon},
+	author       = {Matherly, John},
+	year         = 2024,
+	month        = {jan},
+	url          = {https://blog.shodan.io/deep-dive-http-favicon/},
+	institution  = {Shodan}
+}
 @incollection{Peters2002,
   author        = {Peters, Tim},
   title         = {Algorithms: Introduction},
@@ -185,14 +206,6 @@
 	month        = may,
 	journal      = {Twitter},
 	url          = {https://twitter.com/shodanhq/status/1395501365456261122}
-}
-@misc{shodan_we_2020,
-	title        = {"we created this technique and decided on mmh3 due to space efficiency and performance at scale."},
-	author       = {Shodan},
-	year         = 2020,
-	month        = oct,
-	journal      = {Twitter},
-	url          = {https://twitter.com/shodanhq/status/1314023874128998400}
 }
 @misc{Stinner2016,
 	title        = {My journey to stable benchmark},
