### 2024-09-20 #6124 @hajimes
Dear @cassiersg,

I sincerely apologize for the delay in my response and greatly appreciate your patience and thoughtful comments. Below, I outline the improvements made based on your feedback and hope that they address your concerns.

> **Statement of need** & **State of the field**: The documentation & paper do not discuss much murmurhash3 itself and how it compares to other hash functions. Since performance is claimed as an important characteristic, I think that a discussion of performance should be included or linked somewhere. This could provide answer such as "How does murmurhash3 compare to other hash functions (xxhash, cityhash, spookyhash, fnv...), as a function of the input size?", "How does mmh3 compare to pure-python implementations of murmurhash3?".

I am thankful to your insightful comments and suggestions. To address the issue, I have carefully formulated benchmarking settings and conducted a series of experiments. The benchmarking results are now available in [the documetation for mmh3](https://mmh3.readthedocs.io/en/latest/benchmark.html). Furthermore, in order to ensure reproducibility, I have published the results as JSON files in a [public GitHub repository](https://github.com/hajimes/mmh3-benchmarks).

For meaningful comparisons, I benchmarked various Python-C hash implementations across the following four categories:

- Classical cryptographic hash functions: `md5` and `sha1` from the Python Standard Library. These algorithms were both developed in 1990s. They are secure but significantly slower than non-cryptographic families, making them less suitable for scientific computing.
- **`mmh3` (our library)**: Backed by the MurmurHash3 family. This family of algorithms, developed by Austin Appleby in 2011 , can be viewed as representative of the "second generation" of non-cryptographic hash functions, with [Fowler-Noll-Vo](https://en.wikipedia.org/wiki/Fowler–Noll–Vo_hash_function) (FNV), developed in 1991, being representative of the first. The MurmurHash3 family is significant both historically and practically, as it was the first non-cryptographic hash function that gained widespread recognition and usage beyond existing communities. The simplicity of MurmurHash3, particularly the 32-bit version, also makes it useful for educational purposes.
- `python-xxhash` (by Yue Du): Backed by the first version of xxHash known as `xxh` or `xxh1`, developed by Yann Collet in 2014 . In my opinion, this family can be regarded as "2.5th generation", with several claimed improvements over earlier algorithms.
- `python-xxhash.xxh3` (by Yue Du): This is a Python extension for `xxh3`, a set of two hash functions developed by Yann Collet in 2019, as the successor to `xxh1`. This family can be thought to represent the "third generation" non-cryptographic hash functions, leveraging vectorized arithmetic and SIMD computation for even faster performance.

Regarding the comparison with pure Python implementations, I evaluated my library against [pymmh3](https://github.com/wc-duck/pymmh3) by Fredrik Kihlander. The results indicate that pymmh3 is more than 100 times slower than the native implementation, that is, our library on CPython (the official Python interpreter), although the performance gap may be smaller on PyPy (the JIT implementation of Python). I have decided to omit a detailed discussion of pure Python implementations due to their significantly lower performance from the analysis below.

Then, for high-quality microbenchmarking, I focused on accuracy, reproducibility, and reliability. To achieve accuracy and reliability, I used [pyperf](https://github.com/psf/pyperf) by Victor Stinner as the primary tool. `pyperf` serves as the backend for [pyperformance](https://github.com/python/pyperformance), the authoritative benchmarking suit for various Python implementations, which in turn serves as the backend of the [Python Speed Center](https://speed.python.org), the official continuous benchmarking web-site of CPython to monitor its performance. To maximize reproducibility, I implemented my benchmarking suite as a GitHub Actions workflow.

I will skip the technical details (which are documented in the `mmh3` documentation), but the following three graphs illustrate how my microbenchmarking methodology evolved through experiments and how it contributed to the performance improvements in the new version, `mmh3 5.0.0`.

The first figure represents my initial experiment on `mmh3 4.1.0` , the previous version of my library against other hash libraries. While this setup ensured maximum reproducibility—thanks to the GitHub Actions workflow, which anyone can execute—the reliability was limited, as the tests were run on virtual machines shared by many developers.

![throughput_small](https://github.com/user-attachments/assets/23799f5d-bd89-4033-b28b-ea033d99fb63)

The second graph presents more stable results, reflecting improved reliability. These enhancements were achieved by applying advanced microbenchmarking techniques, including Linux tuning techniques for microbenchmarking, developed by Victor Stinner, and the randomized input method utilized by Yann Collet, the creator of the xxHash algorithm.

![throughput_small](https://github.com/user-attachments/assets/f60acb38-1ff0-4d36-a565-f0bf1e82bc51)

Finally, the third graph shows the performance improvements in `mmh3 5.0.0`, the latest version of my library. This enhancement was realized by the implementation of the Python-C interface `METH_FASTCALL`, a new feature introduced in Python 3.7 that reduces the overhead of function calls. Initially, I hesitated to adopt this complex interface due to uncertainties about its impact on performance and the time required for implementation.

However, with the stability provided by reliable benchmarks, the improvements from integrating `METH_FASTCALL` became clear. As a result, our library is now particularly well-suited for use cases involving repeated hashing of small keys, a common scenario for non-cryptographic hash functions.  For instance, for data sizes between 1–2 KB (e.g., 48x48 favicons), performance is 10%–20% faster. For smaller data (~500 bytes, like 16x16 favicons), performance increases by approximately 30%.

![throughput_small](https://github.com/user-attachments/assets/a87df050-8415-45fc-a670-766162479576)

> **Functionality documentation**: I did not see any exhaustive online API documentation. It seems that docstrings are of good quality, though. I would recommend setting up a API documentation with sphinx (e.g. on readthedocs).

As per your recommendation, I have adopted `Sphinx` for building the documentation, [which is now hosted on Read the Docs](https://mmh3.readthedocs.io/en/latest/). In addition, I have greatly enhanced the docstrings, which can be seen in  the [API Reference](https://mmh3.readthedocs.io/en/latest/api.html). All functions are now documented with detailed inputs and outputs with the Google-Style format.

Beyond improving the docstrings, I have also thoroughly revised the overall documentation. For instance,  the [previous version of CHANGELOG](https://github.com/hajimes/mmh3/blob/3e62a197d85b483d8b51ad93943bb45ce2b17966/CHANGELOG.md)  was disorganized, with uncategorized update items and numerous comments for contributor acknowledgments. To improve clarity, I have moved  acknowledgements to a dedicated [CONTRIBUTORS](https://github.com/hajimes/mmh3/blob/master/docs/CONTRIBUTORS.md)  file and [revised the CHANGELOG](https://github.com/hajimes/mmh3/blob/master/CHANGELOG.md) according to the [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) convention. This reorganization has made the project’s history more structured and easier to navigate.

In conclusion, I have made significant improvements to my library in both performance and documentation, thanks to your concrete suggestions and insightful review comments.
