### 2024-09-22 c5a061f
Add sdist check to gha (should fail)
