### 2024-09-22 e59f33f
Include *.h for sdist
