+++ pyproject.toml
@@ -70,6 +70,12 @@ ext-modules = [
   {name = "mmh3", sources=["./src/mmh3/mmh3module.c", "./src/mmh3/murmurhash3.c"]}
 ]
 
+[tool.setuptools.package-data]
+mmh3 = [
+  "./src/mmh3/murmurhash3.h",
+  "./src/mmh3/hashlib.h"
+]
+
 [tool.isort]
 profile = "black"
 src_paths = ["src/mmh3/__init__.pyi", "util", "tests", "benchmark", "docs"]
