### 2024-09-22 ca84d08
Try include-dirs for setuptools config
