### 2024-09-22 c76785b
Try include-package-data = true
