+++ .github/workflows/build.yml
@@ -46,5 +46,9 @@ jobs:
         run: |
           pip uninstall -y mmh3
           python -m build --sdist
-          python -m pip install dist/*.tar.gz
+          if [[ "${{ matrix.os }}" == "windows-2022" ]]; then
+            for %i in (dist\*.tar.gz) do python -m pip install %i
+          else
+            python -m pip install dist/*.tar.gz
+          fi
           python -m pytest
