### 2024-09-22 e6d6ad7
Revise os detection
