+++ .github/workflows/build.yml
@@ -42,13 +42,21 @@ jobs:
           pip install ".[test,type]"
       - name: Test with pytest
         run: python -m pytest
-      - name: Check source distribution
+      - name: Test type hints with mypy
+        run: mypy --strict tests
+      - name: Test building from the source distribution (Linux/macOS)
+        if: runner.os != "Windows"
         run: |
           pip uninstall -y mmh3
           python -m build --sdist
-          if [[ "${{ matrix.os }}" == "windows-2022" ]]; then
-            for %i in (dist\*.tar.gz) do python -m pip install %i
-          else
-            python -m pip install dist/*.tar.gz
-          fi
+          python -m pip install dist/*.tar.gz
           python -m pytest
+          mypy --strict tests
+      - name: Test building from the source distribution (Windows)
+        if: runner.os == "Windows"
+        run: |
+          pip uninstall -y mmh3
+          python -m build --sdist
+          for %i in (dist\*.tar.gz) do python -m pip install %i
+          python -m pytest
+          mypy --strict tests
