### 2024-09-22 70ef57d
Remove stub file
