### 2024-09-22 01f4499
Add sdist building test to publish workflow
