### 2024-09-22 14edaed
Delineate endian issue
