### 2024-09-22 124e267
Fix issue with not building from sdist (#91)
