+++ .github/workflows/wheels.yml
@@ -78,6 +78,14 @@ jobs:
           python -m pip install --upgrade pip
           pip install setuptools build
           python -m build --sdist
+      - name: Test building from the source distribution
+        shell: bash
+        run: |
+          pip install ".[test,type]"
+          pip uninstall -y mmh3
+          python -m pip install dist/*.tar.gz
+          python -m pytest
+          mypy --strict tests
       - uses: actions/upload-artifact@v4
         with:
           path: dist/*.tar.gz
