+++ pyproject.toml
@@ -56,7 +56,7 @@ benchmark = [
 ]
 plot = [
   "matplotlib == 3.9.2",
-  "pandas == 2.2.2"
+  "pandas == 2.2.3"
 ]
 
 [project.urls]
