### 2024-09-24 bf1091d
Bump the dependencies group with 3 updates (#95)
Bumps the dependencies group with 3 updates: [clang-format](https://github.com/ssciwr/clang-format-wheel), [pylint](https://github.com/pylint-dev/pylint) and [pandas](https://github.com/pandas-dev/pandas).


Updates `clang-format` from 18.1.8 to 19.1.0
- [Release notes](https://github.com/ssciwr/clang-format-wheel/releases)
- [Commits](https://github.com/ssciwr/clang-format-wheel/compare/v18.1.8...v19.1.0)

Updates `pylint` from 3.2.7 to 3.3.1
- [Release notes](https://github.com/pylint-dev/pylint/releases)
- [Commits](https://github.com/pylint-
[…]
