+++ pyproject.toml
@@ -36,9 +36,9 @@ test = [
 ]
 lint = [
   "black == 24.8.0",
-  "clang-format == 18.1.8",
+  "clang-format == 19.1.0",
   "isort == 5.13.2",
-  "pylint == 3.2.7"
+  "pylint == 3.3.1"
 ]
 type = [
   "mypy == 1.11.2"
@@ -56,7 +56,7 @@ benchmark = [
 ]
 plot = [
   "matplotlib == 3.9.2",
-  "pandas == 2.2.2"
+  "pandas == 2.2.3"
 ]
 
 [project.urls]
