### 2024-09-24 #6124 @cassiersg
Hello @hajimes

Congratulations for the great work!

I am surprised by your benchmark results:

1. The general downwards trend of the performance w.r.t. the input size: generally, very low input sizes have lower throughput due to function call overheads, worse branch prediction, etc. (The maximum throughput is usually achieve for sizes of a few kiB, since for larger sizes the limited cache capacity may lead to memory throughput bottlenecks.)
2. The absolute value of the reported throughput, which is a few *orders of magnitude* lower than the state of the art. See for example https://github.com/Cyan4973/xxHash/wiki/Performance-comparison (I do not know how scientific these benchmarks are, but they roughly match my expectations. Other benchmarks can be found at https://www.blake2.net/  and https://github.com/BLAKE3-team/BLAKE3-specs/blob/master/blake3.pdf , showing that even fast cryptographic hash functions can achieve multiple GiB/s).
