### 2024-09-25 #6124 @hajimes
Thank you very much for your review!

In my case, throughput is defined as "output throughput", or the performance of the output size per unit of processing time. As you correctly pointed out, output throughput typically improves with larger inputs, because, in many functions, the output size tends to increase in relation to the input size.

However, in the case of hash functions, the output size is fixed (e.g., MD5 consistently produces 128-bit outputs). Consequently, as the input size grows, the output throughput of hash functions decreases.

Unlike output throughput, which considers the quantity of output, the table of xxHash's benchmarking measures input bandwidth, or the number of input bytes processed per unit of time.

The table below compares the result of input bandwidth between benchmarking. In my case, the input bandwidth is measured at around 192 KiB. Given that the latency graph shows a linear relationship for sufficiently large inputs (e.g., 5 KiB), I believe this bandwidth remains stable for larger inputs. Also, in my benchmarks, "GB/s" refers to 1 IEC GB (10^9 bytes) per second, though I am not certain whether xxHash’s benchmark table uses IEC GB or GiB (2^30 bytes).

| HASH_NAME | Our bench (CPU: 2.4-3.3 GHz) | xxHash's bench (CPU: 3.6-4.9 GHz) |
| --------- | ---------------------------- | --------------------------------- |
| xxh3_128  | 24.1 GB/s (22.4 GiB/s)       | 29.6 GB/s                         |
| xxh3_64   | 24.0 GB/s (22.4 GiB/s)       | 31.5 GB/s                         |
| xxh_64    | 9.53 GB/s (8.88 GiB/s)       | 19.4 GB/s                         |
| mmh3_128  | 7.41 GB/s (6.90 GiB/s)       | N/A                               |
| xxh_32    | 6.61 GB/s (6.16 GiB/s)       | 9.7 GB/s                          |
| mmh3_32   | 3.07 GB/s (2.86 GiB/s)       | 3.9 GB/s                          |
| sha1      | 1.75 GB/s (1.63 GiB/s)       | 0.8 GB/s                          |
| md5       | 0.699 GB/s (0.651 GiB/s)     | 0.6 GB/s                          |

The clock speed of GitHub Actions Ubuntu VM's core is only (2.4 + 3.3) / (3.6 + 4.9) ≒ 67.1 (%) of Intel i7-9700K used in the official xxHash benchmark. Therefore, I believe my benchmarking results are consistent with those published by xxHash's author.

One exceptions is the faster performance of `md5` and `sha1` in my benchmarks, likely due to the efficiency of the underlying hash libraries. While xxHash does mention the exact implementation used, it is plausible that they employed OpenSSL, as it is widely recognized as a standard implementation.  On the other hand, [Python's `hashlib` is backed by `HACL*`](https://github.com/python/cpython/tree/8447c933da308939b06e33544ca9abc9fc46aa8b/Modules/_hacl), a highly optimized library for various cryptographic hash functions, written by F*, a proof-oriented language. The roughly 2x performance difference between the original C implementation of `xxh_64`  between and its Python extension (`python-xxhash`) is also notable, although exploring its cause may be beyond the scope of my current work.
