### 2024-09-25 #6124 @cassiersg
Thanks for the detailed explanations. I have never seen the "output throughput" of a hash function considered before, input throughput is the most common metric. Could you consider changing your plots to use that metric?

(FYI, python's hashlib uses openssl if available, the F* implementation is only a fallback.)
