### 2024-09-26 #6124 @hajimes
## About Performance Visualization

Thank you for your thoughtful feedback! While you suggested altering the graph itself, I wonder if we could instead adopt a new, intuitive term for a metric that better reflects the digest (output) of hash functions. A potential candidate could be "**digest-weighted performance**".

Regarding the input throughput, it is not particularly useful in my case because it is inherently equivalent to latency, which I have already visualized with graphs. Therefore, adding a separate graph for input throughput would be redundant. The key point here is that the digest size plays an important role in the stochastic quality of probabilistic data structures. For instance, a hash function with a 128-bit output can be treated as having four 32-bit hash functions computed in parallel. By leveraging these "parallelly"-computed "multiple" hash functions, we can enhance the stochastic performance of probabilistic data structures, such as reducing the false positive rate in Bloom filters.

The digest size of cryptographic hash functions contributes to security strength, which is orthogonal to performance (measured in terms of speed), and, thus, doesn't need to be visualized alongside performance metrics. However, for non-cryptographic hash functions, the digest size is directly related to performance; 4 times larger is equivalent to 4 times faster in certain use cases. For this reason, it is essential to have a metric that visualizes the quality of hash functions based on digest size.

More specifically, I would like to have two sets of figures that illustrate the relationship between `xxh_64` (the flagship of the first generation of xxHash) and `mmh3_128` (MurmurHash3_x64_128, the flagship of MurmurHash3). In one figure (which I called latency), `xxh_64` outperforms `mmh3_128` by 28.7% due to its faster clock-time speed. However, in another figure (currently labeled "throughput", though I acknowledge the potential confusion), `mmh3_128` outperforms `xxh_64` by 55.3% due to its twice larger digest size, which in some contexts translates to a virtual speed gain. These two sets complement each other, and while the term "throughput" was used in my first plots, I am open to adopting a clearer term that emphasizes the relationship to digest size.

Note: Regarding terminology, [xxHash's benchmarking](https://github.com/Cyan4973/xxHash/wiki/Performance-comparison#benchmarks-concentrating-on-small-data-)  defines throughput as the number of hash functions executed per second given a specific input size. Here the author also uses output throughput, and thus xxHash's throughput graph shows a similar decreasing trend as mine. The main difference in my graph is that my graph's y-axis values are scaled by a constant specific to each hash function (i.e., digest size). Therefore, this usage of terminology is not unique in the non-cryptographic hash function community. However, as we have seen, to avoid confusion, adopting a new term that clearly indicates its relationship to the digest (output) would be more suitable.

## About MD5/SHA1 Performance

Thank you for pointing this out. You are correct in highlighting this, and I appreciate your attention. The documentation for `hashlib` clearly states that, starting from Python 3.12 (the version I used for benchmarking), `[f]or any of the MD5, SHA1, SHA2, or SHA3 algorithms that the linked OpenSSL does not provide we fall back to a verified implementation from the HACL* project`. I had overlooked this important detail.

Given this, a plausible explanation for the performance discrepancy between MD5 and SHA1 could be the presence of Intel SHA extensions. The CPU used in xxHash's benchmarking, a Core i7-9700k, belongs to Intel's 9th-generation desktop processors (Coffee Lake Refresh), which do not support these extensions.

For further reference, [this site benchmarked the performance of OpenSSL hash functions](https://pablotron.org/2022/06/10/hash-speeds/). On a Thinkpad X1 Carbon, which uses an Intel i7-1185G7, OpenSSL SHA1 is around twice faster than OpenSSL MD5 for large inputs. Intel i7-1185G7 belongs to Tiger Lake, the 11th generation Intel mobile processors, which support Intel SHA extensions. Conversely, on a Raspberry Pi, where such extensions are not available, the performance of MD5 and SHA1 is nearly identical (with SHA1 being slightly slower).

These findings are consistent with both xxHash's results and my own, although they still do not fully explain why MD5 in my benchmarks outperforms the one in xxHash's tests. In any case, I will later add this observation to my documentation.
