### 2024-09-26 #6124 @cassiersg
I think that it would be good to split benchmark results:

1. provide detailed benchmarks and discussion in the repo
2. provide only a quick performance overview in the paper (keeping the paper length within JOSS's guidelines)

While there are many ways to report performance of hash functions, it seems to me that the most commonly-used metrics are (1) throughput (w.r.t. the input, either as a function of the input size, or as a single number, for "long" inputs) and (2) number of hashes/s as a function of the input size. These metrics (either, or both) therefore seem the most natural to include in the paper.

That being said, feel free to include whichever metric you find most relevant in the paper, along with a clear explanation of what is being measured. Please keep in mind JOSS's guidelines on paper length.
