+++ paper/paper.md
@@ -19,34 +19,31 @@ date: 28 Sep 2024
 bibliography: paper.bib
 ---
 
-<!-- markdownlint-disable MD025 -->
+<!-- markdownlint-disable single-h1 -->
 
 # Summary
 
-This decade has witnessed the rapid evolution of artificial intelligence (AI),
-notably in the field of natural language processing (NLP), as represented by
-the popularity of OpenAI’s ChatGPT. Another important advancement in computer
-science and engineering is found in the field of the Internet of Things (IoT),
-a crucial component of ubiquitous computing, as represented by the development
-of Shodan, the world's first IoT search engine.
+In recent years, artificial intelligence (AI) has rapidly evolved, particularly
+in natural language processing (NLP) with services like OpenAI's ChatGPT.
+Likewise, the Internet of Things (IoT) continues to grow as a key area of
+ubiquitous computing, exemplified by Shodan, the first IoT search engine.
 
 Underlying these advancements are high-performance algorithms and data
-structures that use non-cryptographic hash functions. This type of hash
-functions is generally characterized by four properties; they are fast, their
-resulting bits are statistically well-distributed, they have an avalanche
-effect, meaning a one-bit difference in a key changes at least half of the
-resulting bits, and they are collision resistant. Because cryptographic
-strength is not required in these use cases, they can leverage the efficiency
-of non-cryptographic hash functions.
+structures relying on non-cryptographic hash functions, which are
+characteristically fast, produce statistically well-distributed bits, exhibit
+an avalanche effect (where a one-bit change in the input alters at least half
+of the output), and are collision resistant. Because cryptographic strength is
+unnecessary in these cases, they benefit from the efficiency of
+non-cryptographic hashes.
 
 MurmurHash3 and its test suite, SMHasher, was developed
 by @appleby_murmurhash3_2011 and is one of the earliest and most continuously
 popular hash functions specifically designed to implement the characteristics
 mentioned above.
 
-`mmh3` was launched in 2011 as a Python wrapper for MurmurHash3 and has been
-maintained ever since. Its API is very simple to use for Python programmers,
-as it offers both simple one-shot hash functions and hasher classes that allow
+`mmh3` was launched in 2011 as a Python extension for MurmurHash3 and has been
+maintained ever since. Its API is simple to use for Python programmers,
+as it offers both one-shot hash functions and hasher classes that allow
 incremental updating, whose methods are compliant to `hashlib`, a part of the
 Python Standard Library. The library provides Python wheels (i.e., pre-built
 binary packages) for immediate use on various platforms, including Linux
@@ -62,10 +59,9 @@ in the PyPI ecosystem are more popular [@hugo_van_kemenade_2024_13624792].
 According to PePy, as of September 1, 2024, the total downloads of
 this library exceeded 130 millions.
 
-Libraries and organizations that directly use `mmh3` include
-Shodan (the world’s first IoT search engine),
-Azure SDK for Python (the official open-source Azure libraries for Python by
-Microsoft), Apache Iceberg (open table format for analytic datasets),
+Libraries and organizations that use `mmh3` include
+Shodan, Microsoft Azure SDK for Python,
+Apache Iceberg (open table format for analytic datasets),
 Feast (feature store for machine learning),
 PyMilvus (Python SDK for Milvus, an open-source vector database),
 and pocsuite3 (open-source remote vulnerability testing framework).
[…]
