+++ paper/paper.bib
@@ -59,10 +59,10 @@
 	year         = 2014,
 	url          = {https://github.com/Cyan4973/xxHash}
 }
-@misc{collet_xxhash_comparison_2024,
-	title        = {{xxHash: Performance comparison (2024)}},
+@misc{collet_xxhash_comparison_2020,
+	title        = {{xxHash: Performance comparison (2020)}},
 	author       = {Collet, Yan},
-	year         = 2024,
+	year         = 2020,
 	url          = {https://github.com/Cyan4973/xxHash/wiki/Performance-comparison}
 }
 @misc{du_xxhash_2014,
