### 2024-09-28 #6124 @hajimes
I fully agree with your suggestions and have revised the draft accordingly. Could you please run the ``@editorialbot generate pdf`` command again and review the Benchmarks section?

As recommended, I have provided a brief overview of the benchmarks in the paper, reserving more detailed explanations for my repository, which I plan to update later.

In the Benchmarks section, I have included a new table that describes the following metrics, ranging from the most commonly used hash function metrics to less famous ones relevant to non-cryptographic literature:

- **Bandwidth**: The amount of input data processed per second (following the terminology used by xxHash).
- **Small Data Velocity**: In my paper, I define this as the "inverse of the mean latency (in microseconds) for inputs in the range of 1–256 bytes". While my definition may differ from the one used in xxHash benchmarks (as the author does provide an exact definition), it is intended to capture the same property and trend.
- **Weighted Small Data Velocity** (shown as "x Width" in the paper): This metric adjusts small data velocity by multiplying it by the hash output's width (digest size) to account for differences in output sizes.
- **Reference to xxHash's benchmarks**: Lastly, I refer to xxHash's benchmarks (Collet, 2020) as a comparison point. Note: When reviewing xxHash's code, I found the line where the macro constant `GB` is defined as `(1 << 30)`, suggesting that the author may have meant GiB instead of GB in the benchmarks too.

For figures, I have included only a graph showing the latency for small to medium-sized inputs, because latency is an unambiguous term that readers will likely find straightforward to interpret.

Finally, I have condensed the paper to adhere to the <1000-word limit of of the JOSS guideline. The word count extension in VSCode shows "983 words", excluding metadata (that is, instructions before `# Summary`), markdown-lint comments, and the table data (but including the table caption).

I hope these revisions align with your expectations and reflect the feedback. Thank you again for your thorough review.
