### 2024-09-28 #6124 @hajimes
For clarity,

1. Bandwidth (term chosen for comparison with Collet 2020 = xxHash) corresponds to "(1) throughput (w.r.t. the input, either as a function of the input size, or as a single number, for "long" inputs)"
2. The latency graph essentially represents "(2) number of hashes/s as a function of the input size" (because latency is its inverse)

Thus, I believe both of the common metrics are effectively covered.
