+++ .github/workflows/wheels.yml
@@ -48,7 +48,7 @@ jobs:
         if: runner.os == 'Linux'
         uses: docker/setup-qemu-action@v3
       - name: Build wheels
-        uses: pypa/cibuildwheel@v2.21.1
+        uses: pypa/cibuildwheel@v2.21.2
         with:
           output-dir: wheelhouse
         env:
