### 2024-10-14 33649d9
Bump the dependencies group across 1 directory with 5 updates
Bumps the dependencies group with 5 updates in the / directory:

| Package | From | To |
| --- | --- | --- |
| [black](https://github.com/psf/black) | `24.8.0` | `24.10.0` |
| [clang-format](https://github.com/ssciwr/clang-format-wheel) | `19.1.0` | `19.1.1` |
| [mypy](https://github.com/python/mypy) | `1.11.2` | `1.12.0` |
| [sphinx](https://github.com/sphinx-doc/sphinx) | `8.0.2` | `8.1.3` |
| [pyperf](https://github.com/psf/pyperf) | `2.7.0` | `2.8.0` |



Updates `black` from 24.8.0 to 24.10.0
- [Release notes](https://github.co
[…]
