+++ pyproject.toml
@@ -35,23 +35,23 @@ test = [
   "pytest-sugar == 1.0.0"
 ]
 lint = [
-  "black == 24.8.0",
-  "clang-format == 19.1.0",
+  "black == 24.10.0",
+  "clang-format == 19.1.1",
   "isort == 5.13.2",
   "pylint == 3.3.1"
 ]
 type = [
-  "mypy == 1.11.2"
+  "mypy == 1.12.0"
 ]
 docs = [
   "myst-parser == 4.0.0",
   "shibuya == 2024.8.30",
-  "sphinx == 8.0.2",
+  "sphinx == 8.1.3",
   "sphinx-copybutton == 0.5.2"
 ]
 benchmark = [
   "pymmh3 == 0.0.5",
-  "pyperf == 2.7.0",
+  "pyperf == 2.8.0",
   "xxhash == 3.5.0"
 ]
 plot = [
