### 2024-10-21 28e3f7b
Bump the dependencies group with 3 updates
Bumps the dependencies group with 3 updates: [clang-format](https://github.com/ssciwr/clang-format-wheel), [mypy](https://github.com/python/mypy) and [shibuya](https://github.com/lepture/shibuya).


Updates `clang-format` from 19.1.1 to 19.1.2
- [Release notes](https://github.com/ssciwr/clang-format-wheel/releases)
- [Commits](https://github.com/ssciwr/clang-format-wheel/compare/v19.1.1...v19.1.2)

Updates `mypy` from 1.12.0 to 1.12.1
- [Changelog](https://github.com/python/mypy/blob/master/CHANGELOG.md)
- [Commits](https://github.com/python/mypy/comp
[…]
