+++ pyproject.toml
@@ -36,16 +36,16 @@ test = [
 ]
 lint = [
   "black == 24.10.0",
-  "clang-format == 19.1.1",
+  "clang-format == 19.1.2",
   "isort == 5.13.2",
   "pylint == 3.3.1"
 ]
 type = [
-  "mypy == 1.12.0"
+  "mypy == 1.12.1"
 ]
 docs = [
   "myst-parser == 4.0.0",
-  "shibuya == 2024.8.30",
+  "shibuya == 2024.10.15",
   "sphinx == 8.1.3",
   "sphinx-copybutton == 0.5.2"
 ]
