### 2024-11-02 8fd7f77
Bump mypy from 1.12.1 to 1.13.0 in the dependencies group (#101)
Bumps the dependencies group with 1 update: [mypy](https://github.com/python/mypy).


Updates `mypy` from 1.12.1 to 1.13.0
- [Changelog](https://github.com/python/mypy/blob/master/CHANGELOG.md)
- [Commits](https://github.com/python/mypy/compare/v1.12.1...v1.13.0)

---
updated-dependencies:
- dependency-name: mypy
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: dependencies
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependab
[…]
