### 2024-11-04 d3595f1
Bump clang-format from 19.1.2 to 19.1.3 in the dependencies group
Bumps the dependencies group with 1 update: [clang-format](https://github.com/ssciwr/clang-format-wheel).


Updates `clang-format` from 19.1.2 to 19.1.3
- [Release notes](https://github.com/ssciwr/clang-format-wheel/releases)
- [Commits](https://github.com/ssciwr/clang-format-wheel/compare/v19.1.2...v19.1.3)

---
updated-dependencies:
- dependency-name: clang-format
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: dependencies
...

Signed-off-by: dependabot[bot] <support@github.
[…]
