+++ pyproject.toml
@@ -36,7 +36,7 @@ test = [
 ]
 lint = [
   "black == 24.10.0",
-  "clang-format == 19.1.2",
+  "clang-format == 19.1.3",
   "isort == 5.13.2",
   "pylint == 3.3.1"
 ]
