### 2024-11-18 cb60be0
Bump the dependencies group across 1 directory with 2 updates
Bumps the dependencies group with 2 updates in the / directory: [clang-format](https://github.com/ssciwr/clang-format-wheel) and [pyperf](https://github.com/psf/pyperf).


Updates `clang-format` from 19.1.2 to 19.1.3
- [Release notes](https://github.com/ssciwr/clang-format-wheel/releases)
- [Commits](https://github.com/ssciwr/clang-format-wheel/compare/v19.1.2...v19.1.3)

Updates `pyperf` from 2.8.0 to 2.8.1
- [Release notes](https://github.com/psf/pyperf/releases)
- [Changelog](https://github.com/psf/pyperf/blob/main/doc/changelog.
[…]
