+++ pyproject.toml
@@ -36,7 +36,7 @@ test = [
 ]
 lint = [
   "black == 24.10.0",
-  "clang-format == 19.1.2",
+  "clang-format == 19.1.4",
   "isort == 5.13.2",
   "pylint == 3.3.1"
 ]
@@ -51,7 +51,7 @@ docs = [
 ]
 benchmark = [
   "pymmh3 == 0.0.5",
-  "pyperf == 2.8.0",
+  "pyperf == 2.8.1",
   "xxhash == 3.5.0"
 ]
 plot = [
