### 2024-11-28 #6124 @cassiersg
Hi, sorry for the late reply.

My concerns have been addressed.

Figure 1 is easy to understand and well-explained in the paper.

Minor (non-blocking) comments for Table 1:
- The "width" column would benefit from having the unit stated, e.g. "Width (bit)"
- I still don't see the interpretation of the "x width" column (i.e., I don't know any usecase where this metric would be the relevant one for choosing a hash function).
