### 2024-12-02 cbbfcd6
Bump the dependencies group across 1 directory with 5 updates
Bumps the dependencies group with 5 updates in the / directory:

| Package | From | To |
| --- | --- | --- |
| [pytest](https://github.com/pytest-dev/pytest) | `8.3.3` | `8.3.4` |
| [clang-format](https://github.com/ssciwr/clang-format-wheel) | `19.1.2` | `19.1.4` |
| [pylint](https://github.com/pylint-dev/pylint) | `3.3.1` | `3.3.2` |
| [pyperf](https://github.com/psf/pyperf) | `2.8.0` | `2.8.1` |
| [matplotlib](https://github.com/matplotlib/matplotlib) | `3.9.2` | `3.9.3` |



Updates `pytest` from 8.3.3 to 8.3.4
- [Release notes]
[…]
