### 2024-12-14 7a0aa7c
Bump the dependencies group across 1 directory with 3 updates
Bumps the dependencies group with 3 updates in the / directory: [clang-format](https://github.com/ssciwr/clang-format-wheel), [shibuya](https://github.com/lepture/shibuya) and [matplotlib](https://github.com/matplotlib/matplotlib).


Updates `clang-format` from 19.1.4 to 19.1.5
- [Release notes](https://github.com/ssciwr/clang-format-wheel/releases)
- [Commits](https://github.com/ssciwr/clang-format-wheel/commits)

Updates `shibuya` from 2024.10.15 to 2024.12.13
- [Release notes](https://github.com/lepture/shibuya/releases)
- [Chang
[…]
