+++ pyproject.toml
@@ -36,7 +36,7 @@ test = [
 ]
 lint = [
   "black == 24.10.0",
-  "clang-format == 19.1.4",
+  "clang-format == 19.1.5",
   "isort == 5.13.2",
   "pylint == 3.3.2"
 ]
@@ -45,7 +45,7 @@ type = [
 ]
 docs = [
   "myst-parser == 4.0.0",
-  "shibuya == 2024.10.15",
+  "shibuya == 2024.12.13",
   "sphinx == 8.1.3",
   "sphinx-copybutton == 0.5.2"
 ]
@@ -55,7 +55,7 @@ benchmark = [
   "xxhash == 3.5.0"
 ]
 plot = [
-  "matplotlib == 3.9.3",
+  "matplotlib == 3.10.0",
   "pandas == 2.2.3"
 ]
 
