### 2024-12-14 4a606d7
Revise table in the paper
