### 2024-12-14 99e6489
Fix table in the paper
