### 2024-12-15 #6124 @hajimes
@cassiersg

I appreciate your careful review and am delighted to hear that my revisions addressed your suggestions.

I agree with your comments and have made the following additional minor revisions to the paper:

- Added the ``bits`` unit to the ``Width`` column.
- Removed the ``x Width`` column from the table.
	- As you questioned, the usefulness of this field depends on its context, and I acknowledge that I failed to provide an adequate explanation. This column pertains to performance in some stochastic techniques such as Bloom filters, but is less relevant to footprint usage (e.g., as in Shodan). While I could amend this oversight by adding an explanation, considering that JOSS papers should be as concise as possible, I figure that it is best to remove this column entirely. These details should instead be moved to the advanced documentation of the library.

I hope these revisions ensure that my paper now meets the journal's standards.
Once again, I truly value your thoughtful feedback.
