### 2024-12-18 1456b66
Bump shibuya from 2024.12.13 to 2024.12.15 in the dependencies group (#109)
Bumps the dependencies group with 1 update: [shibuya](https://github.com/lepture/shibuya).


Updates `shibuya` from 2024.12.13 to 2024.12.15
- [Release notes](https://github.com/lepture/shibuya/releases)
- [Changelog](https://github.com/lepture/shibuya/blob/master/docs/changelog.rst)
- [Commits](https://github.com/lepture/shibuya/commits)

---
updated-dependencies:
- dependency-name: shibuya
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: dependencies
...

Signed-off-
[…]
