+++ pyproject.toml
@@ -45,7 +45,7 @@ type = [
 ]
 docs = [
   "myst-parser == 4.0.0",
-  "shibuya == 2024.12.13",
+  "shibuya == 2024.12.15",
   "sphinx == 8.1.3",
   "sphinx-copybutton == 0.5.2"
 ]
