### 2024-12-23 b32492e
Bump the dependencies group with 2 updates
Bumps the dependencies group with 2 updates: [mypy](https://github.com/python/mypy) and [shibuya](https://github.com/lepture/shibuya).


Updates `mypy` from 1.13.0 to 1.14.0
- [Changelog](https://github.com/python/mypy/blob/master/CHANGELOG.md)
- [Commits](https://github.com/python/mypy/compare/v1.13.0...v1.14.0)

Updates `shibuya` from 2024.12.15 to 2024.12.21
- [Release notes](https://github.com/lepture/shibuya/releases)
- [Changelog](https://github.com/lepture/shibuya/blob/master/docs/changelog.rst)
- [Commits](https://github.com/lepture/shibuya/co
[…]
