+++ pyproject.toml
@@ -41,11 +41,11 @@ lint = [
   "pylint == 3.3.2"
 ]
 type = [
-  "mypy == 1.13.0"
+  "mypy == 1.14.0"
 ]
 docs = [
   "myst-parser == 4.0.0",
-  "shibuya == 2024.12.15",
+  "shibuya == 2024.12.21",
   "sphinx == 8.1.3",
   "sphinx-copybutton == 0.5.2"
 ]
