### 2024-12-30 02fd7f9
Bump the dependencies group across 1 directory with 3 updates
Bumps the dependencies group with 3 updates in the / directory: [pylint](https://github.com/pylint-dev/pylint), [mypy](https://github.com/python/mypy) and [shibuya](https://github.com/lepture/shibuya).


Updates `pylint` from 3.3.2 to 3.3.3
- [Release notes](https://github.com/pylint-dev/pylint/releases)
- [Commits](https://github.com/pylint-dev/pylint/compare/v3.3.2...v3.3.3)

Updates `mypy` from 1.13.0 to 1.14.0
- [Changelog](https://github.com/python/mypy/blob/master/CHANGELOG.md)
- [Commits](https://github.com/python/mypy/compar
[…]
