### 2025-01-06 055da99
Bump the dependencies group across 1 directory with 4 updates
Bumps the dependencies group with 4 updates in the / directory: [clang-format](https://github.com/ssciwr/clang-format-wheel), [pylint](https://github.com/pylint-dev/pylint), [mypy](https://github.com/python/mypy) and [shibuya](https://github.com/lepture/shibuya).


Updates `clang-format` from 19.1.5 to 19.1.6
- [Release notes](https://github.com/ssciwr/clang-format-wheel/releases)
- [Commits](https://github.com/ssciwr/clang-format-wheel/commits/v19.1.6)

Updates `pylint` from 3.3.2 to 3.3.3
- [Release notes](https://github.com/pylin
[…]
