+++ pyproject.toml
@@ -36,16 +36,16 @@ test = [
 ]
 lint = [
   "black == 24.10.0",
-  "clang-format == 19.1.5",
+  "clang-format == 19.1.6",
   "isort == 5.13.2",
-  "pylint == 3.3.2"
+  "pylint == 3.3.3"
 ]
 type = [
-  "mypy == 1.13.0"
+  "mypy == 1.14.1"
 ]
 docs = [
   "myst-parser == 4.0.0",
-  "shibuya == 2024.12.15",
+  "shibuya == 2024.12.21",
   "sphinx == 8.1.3",
   "sphinx-copybutton == 0.5.2"
 ]
