### 2025-01-06 #6124 @drvinceknight
@hajimes I have made some minor suggestions to the text here: https://github.com/hajimes/mmh3/pull/113

After you have decided to either include my edits or not could you make a tagged release and archive, and report the version number and archive DOI here please. Note that it's important that the title and author list of the paper match the title and author list of the archive.
