### 2025-01-17 #6124 @danielskatz
👋 @hajimes - I've requested some changes to casing in the bib in https://github.com/hajimes/mmh3/pull/114  Please merge this, or let me know what you disagree with, then we can continue to acceptance and publication.
