### 2025-01-18 59e12d3
minor changes in paper.bib (#114)
