+++ paper/paper.bib
@@ -44,7 +44,7 @@
 	title        = {On the resemblance and containment of documents},
 	author       = {Broder, Andrei Z.},
 	year         = 1997,
-	booktitle    = {Proceedings. {Compression} and {Complexity} of {SEQUENCES} 1997},
+	booktitle    = {Proceedings. Compression and Complexity of {SEQUENCES} 1997},
 	publisher    = {IEEE Comput. Soc},
 	pages        = {21--29},
 	doi          = {10.1109/SEQUEN.1997.666900},
@@ -60,7 +60,7 @@
 	url          = {https://github.com/Cyan4973/xxHash}
 }
 @misc{collet_xxhash_comparison_2020,
-	title        = {{xxHash: Performance comparison (2020)}},
+	title        = {{xxHash}: Performance comparison (2020)},
 	author       = {Collet, Yan},
 	year         = 2020,
 	url          = {https://github.com/Cyan4973/xxHash/wiki/Performance-comparison}
@@ -79,7 +79,7 @@
 	url          = {https://faradaysec.com/understanding-spring4shell/}
 }
 @book{gorelick_high_2020,
-	title        = {High {Performance} {Python}: {Practical} {Performant} {Programming} for {Humans}},
+	title        = {High Performance {P}ython: Practical Performant Programming for Humans},
 	author       = {Gorelick, Micha and Ozsvald, Ian},
 	year         = 2020,
 	month        = jun,
@@ -100,11 +100,11 @@
   url          = {https://doi.org/10.5281/zenodo.13624792}
 }
 @inproceedings{kakwani_indicnlpsuite_2020,
-	title        = {{IndicNLPSuite}: {Monolingual} {Corpora}, {Evaluation} {Benchmarks} and {Pre}-trained {Multilingual} {Language} {Models} for {Indian} {Languages}},
+	title        = {{IndicNLPSuite}: Monolingual Corpora, Evaluation Benchmarks and Pre-trained Multilingual Language Models for {I}ndian Languages},
 	author       = {Kakwani, Divyanshu and Kunchukuttan, Anoop and Golla, Satish and N.C., Gokul and Bhattacharyya, Avik and Khapra, Mitesh M. and Kumar, Pratyush},
 	year         = 2020,
 	month        = nov,
-	booktitle    = {Findings of the {Association} for {Computational} {Linguistics}: {EMNLP} 2020},
+	booktitle    = {Findings of the {A}ssociation for {C}omputational {L}inguistics: {EMNLP} 2020},
 	publisher    = {Association for Computational Linguistics},
 	address      = {Online},
 	pages        = {4948--4961},
@@ -126,7 +126,7 @@
 	institution  = {SANS Internet Storm Center}
 }
 @book{kumar_probabilistic_2021,
-	title        = {Probabilistic {Data} {Structures} for {Blockchain}-{Based} {Internet} of {Things} {Applications}},
+	title        = {Probabilistic Data Structures for Blockchain-Based {I}nternet of {T}hings Applications},
 	author       = {Kumar, Neeraj and Miglani, Arzoo},
 	year         = 2021,
 	month        = jan,
@@ -135,7 +135,7 @@
 	isbn         = {978-0-367-52990-1}
 }
 @book{medjedovic_algorithms_2022,
-	title        = {Algorithms and {Data} {Structures} for {Massive} {Datasets}},
+	title        = {Algorithms and Data Structures for Massive Datasets},
 	author       = {Medjedovic, Dzejla and Tahirovic, Emin and Dedovic, Ines},
 	year         = 2022,
 	month        = jul,
@@ -177,22 +177,22 @@
 	institution  = {Quansight Labs}
 }
 @inproceedings{Senuma2011,
[…]
