### 2025-01-18 #6124 @hajimes
@danielskatz
I have merged the patch. Thank you very much for ensuring consistent capitalization in the bibliography titles and for having overseen the review process since the pre-review stage.
