### 2025-01-18 #6124 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
