### 2025-01-18 #6124 @danielskatz
Congratulations to @hajimes (Hajime Senuma) on your publication!!
